### 2023-12-04 #6121 @editorialbot
[PRE REVIEW]: SpeedyWeather.jl: Reinventing atmospheric general circulation models towards interactivity and extensibility
**Submitting author:** @milankl (<a href="http://orcid.org/0000-0002-3920-4356">Milan Klöwer</a>)
**Repository:** https://github.com/SpeedyWeather/SpeedyWeather.jl
**Branch with paper.md** (empty if default branch): mk/josspaper
**Version:** v0.7
