### 2023-12-04 3bacbde
allow run!(simulation,period=1.23)
