+++ test/dates.jl
@@ -34,13 +34,13 @@
     SpeedyWeather.initialize!(c1, L6)
     @test c1.n_timesteps == 10 
 
-    SpeedyWeather.set_period!(c1, 10)
+    SpeedyWeather.set_period!(c1, 10)   # assumed to be in days
     SpeedyWeather.initialize!(c1, L6)
-    @test c1.n_timesteps == 10
+    @test c1.n_timesteps == 24*10
 
-    SpeedyWeather.set_period!(c1, 10.0)
+    SpeedyWeather.set_period!(c1, 10.0) # also assumed to be in days
     SpeedyWeather.initialize!(c1, L6)
-    @test c1.n_timesteps == 10 
+    @test c1.n_timesteps == 24*10 
 
     SpeedyWeather.set_period!(c1, Day(2))
     SpeedyWeather.initialize!(c1, L6)
