+++ docs/joss/paper.bib
@@ -14,7 +14,6 @@
   abstract = {Abstract In a recent study, Williams introduced a simple modification to the widely used Robert\textendash Asselin (RA) filter for numerical integration. The main purpose of the Robert\textendash Asselin\textendash Williams (RAW) filter is to avoid the undesired numerical damping of the RA filter and to increase the accuracy. In the present paper, the effects of the modification are comprehensively evaluated in the Simplified Parameterizations, Primitive Equation Dynamics (SPEEDY) atmospheric general circulation model. First, the authors search for significant changes in the monthly climatology due to the introduction of the new filter. After testing both at the local level and at the field level, no significant changes are found, which is advantageous in the sense that the new scheme does not require a retuning of the parameterized model physics. Second, the authors examine whether the new filter improves the skill of short- and medium-term forecasts. January 1982 data from the NCEP\textendash NCAR reanalysis are used to evaluate the forecast skill. Improvements are found in all the model variables (except the relative humidity, which is hardly changed). The improvements increase with lead time and are especially evident in medium-range forecasts (96\textendash 144 h). For example, in tropical surface pressure predictions, 5-day forecasts made using the RAW filter have approximately the same skill as 4-day forecasts made using the RA filter. The results of this work are encouraging for the implementation of the RAW filter in other models currently using the RA filter.},
   chapter = {Monthly Weather Review},
   langid = {english},
-  file = {/Users/milan/Zotero/storage/KWWGE38V/Amezcua et al. - 2011 - The Effects of the RAW Filter on the Climatology a.pdf}
 }
 
 @article{Bezanson2017,
@@ -32,7 +31,6 @@
   doi = {10.1137/141000671},
   urldate = {2020-02-25},
   abstract = {Bridging cultures that have often been distant, Julia combines expertise from the diverse fields of computer science and computational science to create a new approach to numerical  computing. Julia is  designed to be easy and fast and questions notions generally held to be ``laws of nature"  by practitioners of numerical computing: \textbackslash beginlist \textbackslash item  High-level dynamic programs have to be slow. \textbackslash item  One must prototype in one language and then rewrite in another language for speed or deployment. \textbackslash item There are parts of a system appropriate for the programmer, and other parts that are best left untouched as they have been built by the experts. \textbackslash endlist We introduce the  Julia programming language and its design---a  dance between specialization and abstraction. Specialization allows for custom treatment. Multiple dispatch,  a  technique from computer science, picks  the right algorithm for the right circumstance. Abstraction, which is what good computation is really about, recognizes what remains the same after differences are stripped away. Abstractions in mathematics are captured as code through another technique from computer science, generic programming. Julia shows that  one can achieve machine performance without sacrificing human convenience.},
-  file = {/Users/milan/Zotero/storage/SZM3ENMR/Bezanson et al. - 2017 - Julia A Fresh Approach to Numerical Computing.pdf;/Users/milan/Zotero/storage/UG5NFZ28/141000671.html}
 }
 
 @article{Bourke1972,
@@ -48,7 +46,6 @@
   doi = {10.1175/1520-0493(1972)100<0683:AEOPSM>2.3.CO;2},
   urldate = {2022-01-07},
   langid = {english},
-  file = {/Users/milan/Zotero/storage/29QHT8QP/Bourke - 1972 - An Efficient, One-Level, Primitive-Equation Spectr.pdf}
 }
 
 @article{Burns2020,
@@ -65,7 +62,6 @@
   doi = {10.1103/PhysRevResearch.2.023068},
   urldate = {2023-09-14},
[…]
