+++ src/physics/surface_fluxes.jl
@@ -7,8 +7,8 @@ function surface_fluxes!(column::ColumnVariables,model::PrimitiveEquation)
     surface_wind_stress!(column,model.surface_wind)
 
     # now call other heat and humidity fluxes
-    sensible_heat_flux!(column,model.surface_heat_flux,model.constants)
-    evaporation!(column,model)
+    # sensible_heat_flux!(column,model.surface_heat_flux,model.constants)
+    # evaporation!(column,model)
 end
 
 struct SurfaceThermodynamicsConstant{NF<:AbstractFloat} <: AbstractSurfaceThermodynamics{NF} end
@@ -23,13 +23,13 @@ function surface_thermodynamics!(   column::ColumnVariables,
                                     model::PrimitiveWet)
 
     # surface value is same as lowest model level
-    column.temp[end] = column.temp[end-1]       # todo use constant POTENTIAL temperature
-    column.humid[end] = column.humid[end-1]     # humidity at surface is the same as 
+    column.surface_temp = column.temp[end]      # todo use constant POTENTIAL temperature
+    column.surface_humid = column.humid[end]    # humidity at surface is the same as 
 
     # surface air density via virtual temperature
     (;R_dry,μ_virt_temp) = C
-    T = column.temp[end]        # surface air temperature
-    q = column.humid[end]       # surface humidity
+    T = column.surface_temp     # surface air temperature
+    q = column.surface_humid    # surface humidity
     Tᵥ = T*(1 + μ_virt_temp*q)  # virtual temperature
     column.surface_air_density = column.pres[end]/R_dry/Tᵥ
 end
@@ -40,8 +40,8 @@ function surface_thermodynamics!(   column::ColumnVariables,
                                     model::PrimitiveDry)
 
     # surface value is same as lowest model level
-    column.temp[end] = column.temp[end-1]       # todo use constant POTENTIAL temperature
-    column.surface_air_density = column.pres[end]/C.R_dry/column.temp[end]
+    column.surface_temp = column.temp   # todo use constant POTENTIAL temperature
+    column.surface_air_density = column.pres[end]/C.R_dry/column.surface_temp
 end
 
 Base.@kwdef struct SurfaceWind{NF<:AbstractFloat} <: AbstractSurfaceWind{NF}
@@ -58,8 +58,8 @@ Base.@kwdef struct SurfaceWind{NF<:AbstractFloat} <: AbstractSurfaceWind{NF}
     "Drag coefficient over sea [1]"
     drag_sea::NF = 1.8e-3
 
-    "Flux limiter [kg/m/s²]"
-    max_flux::NF = 1.5
+    # "Flux limiter [kg/m/s²]"
+    # max_flux::NF = 1.5
 end
 
 SurfaceWind(SG::SpectralGrid;kwargs...) = SurfaceWind{SG.NF}(;kwargs...)
@@ -67,15 +67,17 @@ SurfaceWind(SG::SpectralGrid;kwargs...) = SurfaceWind{SG.NF}(;kwargs...)
 function surface_wind_stress!(  column::ColumnVariables{NF},
                                 surface_wind::SurfaceWind) where NF
 
-    (;u,v,land_fraction) = column
-    (;f_wind, V_gust, drag_land, drag_sea, max_flux) = surface_wind     
+    (;land_fraction) = column
+    (;f_wind, V_gust, drag_land, drag_sea) = surface_wind
+    # (;max_flux) = surface_wind
 
     # SPEEDY documentation eq. 49
-    u[end] = f_wind*u[end-1] 
-    v[end] = f_wind*v[end-1]
[…]
