### 2023-12-11 3c0f4fb
AbstractSimulation
