+++ src/abstract_types.jl
@@ -1,4 +1,5 @@
 # MODELS
+abstract type AbstractSimulation{Model} end
 abstract type ModelSetup end
 abstract type Barotropic <: ModelSetup end
 abstract type ShallowWater <: ModelSetup end
