### 2023-12-11 af8de41
get_column function
