+++ src/physics/boundary_layer.jl
@@ -24,12 +24,15 @@ end
 """Linear boundary layer drag Following Held and Suarez, 1996 BAMS
 $(TYPEDFIELDS)"""
 Base.@kwdef struct LinearDrag{NF<:AbstractFloat} <: BoundaryLayerDrag{NF}
+    
+    # DIMENSIONS
+    nlev::Int
+    
     # PARAMETERS
-    σb::Float64 = 0.7           # sigma coordinate below which linear drag is applied
-    time_scale::Float64 = 24    # [hours] time scale for linear drag coefficient at σ=1 (=1/kf in HS96)
+    σb::NF = 0.7                    # sigma coordinate below which linear drag is applied
+    time_scale::Second = Hour(24)   # time scale for linear drag coefficient at σ=1 (=1/kf in HS96)
 
     # PRECOMPUTED CONSTANTS
-    nlev::Int = 0
     drag_coefs::Vector{NF} = zeros(NF,nlev)
 end
 
@@ -45,9 +48,9 @@ function initialize!(   scheme::LinearDrag,
                         model::PrimitiveEquation)
 
     (;σ_levels_full) = model.geometry
-    (;σb,time_scale,drag_coefs) = scheme
+    (;σb, time_scale, drag_coefs) = scheme
 
-    kf = 1/(time_scale*3600)                        # scale with radius as ∂ₜu is; hrs -> sec
+    kf = 1/time_scale.value
 
     for (k,σ) in enumerate(σ_levels_full)
         drag_coefs[k] = -kf*max(0,(σ-σb)/(1-σb))    # drag only below σb, lin increasing to kf at σ=1
@@ -60,7 +63,7 @@ Compute tendency for boundary layer drag of a `column` and add to its tendencies
 function boundary_layer_drag!(  column::ColumnVariables,
                                 scheme::LinearDrag)
 
-    (;u,v,u_tend,v_tend) = column
+    (;u, v, u_tend, v_tend) = column
     (;drag_coefs) = scheme
 
     @inbounds for k in eachlayer(column)
