### 2023-12-11 12c7689
humidity diffusion
