+++ src/physics/tendencies.jl
@@ -32,6 +32,7 @@ function parameterization_tendencies!(
 
         # VERTICAL DIFFUSION
         static_energy_diffusion!(column,model)
+        humidity_diffusion!(column,model)
 
         # HELD-SUAREZ
         temperature_relaxation!(column,model)
@@ -62,17 +63,17 @@ function fluxes_to_tendencies!(
     constants::DynamicsConstants,
 )
     
-    (;nlev,u_tend,flux_u_upward,flux_u_downward) = column
-    (;v_tend,flux_v_upward,flux_v_downward) = column
-    (;humid_tend,flux_humid_upward,flux_humid_downward) = column
-    (;temp_tend,flux_temp_upward,flux_temp_downward) = column
+    (;nlev, u_tend, flux_u_upward, flux_u_downward) = column
+    (;      v_tend, flux_v_upward, flux_v_downward) = column
+    (;humid_tend, flux_humid_upward, flux_humid_downward) = column
+    (;temp_tend,  flux_temp_upward,  flux_temp_downward) = column
 
     Δσ = geometry.σ_levels_thick
     pₛ = column.pres[end]               # surface pressure
     (;radius) = constants               # used for scaling
-    
+
     # TODO ONLY FOR TESTING TO MAKE PARAMETERIZATIONS WEAKER
-    radius /= 1.5
+    radius /= 2
 
     # for g/Δp and g/(Δp*cₚ), see Fortran SPEEDY documentation eq. (3,5)
     g_pₛ = constants.gravity/pₛ
