### 2023-12-14 cbe761e
Make prognostics available for forcing, drag in Barotropic/SWM
