### 2023-12-14 3ffc815
Move lf outside of gridded! Move spectral_truncation! into time_integration
