+++ src/dynamics/tendencies_dynamics.jl
@@ -131,7 +131,6 @@ function surface_pressure_tendency!(
     @. pres_tend = -pres_tend - D̄   # the -D̄ term in spectral and swap sign
     
     pres_tend[1] = 0                # for mass conservation
-    spectral_truncation!(pres_tend) # remove lmax+1 row, only vectors use it
     return nothing
 end
 
@@ -292,8 +291,6 @@ function temperature_tendency!(
     # now add the -∇⋅((u,v)*T') term
     flux_divergence!(temp_tend,temp_grid,diagn,G,S,add=true,flipsign=true)
     
-    # only vectors make use of the lmax+1 row, set to zero for scalars
-    spectral_truncation!(temp_tend)
     return nothing
 end
 
@@ -505,7 +502,6 @@ function bernoulli_potential!(  diagn::DiagnosticVariablesLayer{NF},
     spectral!(bernoulli,bernoulli_grid,S)                   # to spectral space
     bernoulli .+= geopot                                    # add geopotential Φ
     ∇²!(div_tend,bernoulli,S,add=true,flipsign=true)        # add -∇²(½(u² + v²) + ϕ)
-    spectral_truncation!(div_tend)                          # set lmax+1 row to zero
 end
 
 """
@@ -574,7 +570,8 @@ function SpeedyTransforms.gridded!(
 
     # all variables on layers
     for (progn_layer,diagn_layer) in zip(progn.layers,diagn.layers)
-        gridded!(diagn_layer,progn_layer,lf,model)
+        progn_layer_lf = progn_layer.timesteps[lf]
+        gridded!(diagn_layer,progn_layer_lf,model)
     end
 
     # surface only for ShallowWaterModel or PrimitiveEquation
@@ -592,22 +589,21 @@ Propagate the spectral state of the prognostic variables `progn` to the
 diagnostic variables in `diagn` for the barotropic vorticity model.
 Updates grid vorticity, spectral stream function and spectral and grid velocities u,v."""
 function SpeedyTransforms.gridded!( diagn::DiagnosticVariablesLayer,   
-                                    progn::PrognosticLayerTimesteps,
-                                    lf::Int,                            # leapfrog index
+                                    progn::PrognosticVariablesLayer,
                                     model::Barotropic)
     
     (; vor_grid, u_grid, v_grid ) = diagn.grid_variables
+    (; vor ) = progn                    # relative vorticity
     U = diagn.dynamics_variables.a      # reuse work arrays for velocities in spectral
     V = diagn.dynamics_variables.b      # U = u*coslat, V=v*coslat
     S = model.spectral_transform
 
-    vor_lf = progn.timesteps[lf].vor    # relative vorticity at leapfrog step lf
-    gridded!(vor_grid,vor_lf,S)         # get vorticity on grid from spectral vor
+    gridded!(vor_grid,vor,S)            # get vorticity on grid from spectral vor
     
     # get spectral U,V from spectral vorticity via stream function Ψ
     # U = u*coslat = -coslat*∂Ψ/∂lat
     # V = v*coslat = ∂Ψ/∂lon, radius omitted in both cases
-    UV_from_vor!(U,V,vor_lf,S)
+    UV_from_vor!(U,V,vor,S)
 
     # transform from U,V in spectral to u,v on grid (U,V = u,v*coslat)
     gridded!(u_grid,U,S,unscale_coslat=true)
@@ -623,26 +619,23 @@ diagnostic variables in `diagn` for the shallow water model. Updates grid vortic
 grid divergence, grid interface displacement (`pres_grid`) and the velocities
 u,v."""
 function SpeedyTransforms.gridded!( diagn::DiagnosticVariablesLayer,
[…]
