+++ src/dynamics/time_integration.jl
@@ -132,7 +132,6 @@ function leapfrog!( A_old::LowerTriangularMatrix{Complex{NF}},      # prognostic
     
     A_lf = lf == 1 ? A_old : A_new                      # view on either t or t+dt to dis/enable Williams filter        
     (;robert_filter, williams_filter) = L               # coefficients for the Robert and Williams filter
-    two = convert(NF,2)                                 # 2 in number format NF
     dt_NF = convert(NF,dt)                              # time step dt in number format NF
 
     # LEAP FROG time step with or without Robert+Williams filter
@@ -140,13 +139,13 @@ function leapfrog!( A_old::LowerTriangularMatrix{Complex{NF}},      # prognostic
     # see Williams (2009), Eq. 7-9
     # for lf == 1 (initial time step) no filter applied (w1=w2=0)
     # for lf == 2 (later steps) Robert+Williams filter is applied
-    w1 = lf == 1 ? zero(NF) : robert_filter*williams_filter/two         # = ν*α/2 in Williams (2009, Eq. 8)
-    w2 = lf == 1 ? zero(NF) : robert_filter*(1-williams_filter)/two     # = ν(1-α)/2 in Williams (2009, Eq. 9)
+    w1 = lf == 1 ? zero(NF) : robert_filter*williams_filter/2       # = ν*α/2 in Williams (2009, Eq. 8)
+    w2 = lf == 1 ? zero(NF) : robert_filter*(1-williams_filter)/2   # = ν(1-α)/2 in Williams (2009, Eq. 9)
 
     @inbounds for lm in eachharmonic(A_old,A_new,A_lf,tendency)
         a_old = A_old[lm]                       # double filtered value from previous time step (t-Δt)
         a_new = a_old + dt_NF*tendency[lm]      # Leapfrog/Euler step depending on dt=Δt,2Δt (unfiltered at t+Δt)
-        a_update = a_old - two*A_lf[lm] + a_new # Eq. 8&9 in Williams (2009), calculate only once
+        a_update = a_old - 2A_lf[lm] + a_new # Eq. 8&9 in Williams (2009), calculate only once
         A_old[lm] = A_lf[lm] + w1*a_update      # Robert's filter: A_old[lm] becomes 2xfiltered value at t
         A_new[lm] = a_new - w2*a_update         # Williams filter: A_new[lm] becomes 1xfiltered value at t+Δt
     end
@@ -168,8 +167,22 @@ function leapfrog!( progn::PrognosticLayerTimesteps,
         var_old = getproperty(progn.timesteps[1],var)
         var_new = getproperty(progn.timesteps[2],var)
         var_tend = getproperty(diagn.tendencies,Symbol(var,:_tend))
+        spectral_truncation!(var_tend)      # set lmax+1 mode to zero
         leapfrog!(var_old,var_new,var_tend,dt,lf,model.time_stepping)
     end
+end
+
+function leapfrog!( progn::PrognosticSurfaceTimesteps,
+                    diagn::SurfaceVariables,
+                    dt::Real,               # time step (mostly =2Δt, but for init steps =Δt,Δt/2)
+                    lf::Int,                # leapfrog index to dis/enable Williams filter
+                    model::ModelSetup)
+               
+    (;pres_tend) = diagn
+    pres_old = progn.timesteps[1].pres
+    pres_new = progn.timesteps[2].pres
+    spectral_truncation!(pres_tend)         # set lmax+1 mode to zero
+    leapfrog!(pres_old,pres_new,pres_tend,dt,lf,model.time_stepping)
 end
 
 """
@@ -184,7 +197,7 @@ function first_timesteps!(
 )
     
     (;clock) = progn
[…]
