### 2023-12-16 3c8259d
Add forcing and drag extensions to tests
