+++ test/extending.jl
@@ -1,0 +1,196 @@
+@testset "Extending forcing and drag" begin
+    Base.@kwdef struct JetDrag{NF} <: SpeedyWeather.AbstractDrag{NF}
+
+        # DIMENSIONS from SpectralGrid
+        "Spectral resolution as max degree of spherical harmonics"
+        trunc::Int
+    
+        # OPTIONS
+        "Relaxation time scale τ"
+        time_scale::Second = Day(6)
+    
+        "Jet strength [m/s]"
+        u₀::Float64 = 20
+    
+        "latitude of Gaussian jet [˚N]"
+        latitude::Float64 = 30
+    
+        "Width of Gaussian jet [˚]"
+        width::Float64 = 6
+    
+        # TO BE INITIALISED
+        "Relaxation back to reference vorticity"
+        ζ₀::LowerTriangularMatrix{Complex{NF}} = zeros(LowerTriangularMatrix{Complex{NF}},trunc+2,trunc+1)
+    end
+    
+    function JetDrag(SG::SpectralGrid;kwargs...)
+        return JetDrag{SG.NF}(;SG.trunc,kwargs...)
+    end
+    
+    function SpeedyWeather.initialize!( drag::JetDrag,
+                                        model::SpeedyWeather.ModelSetup)
+    
+        (;spectral_grid, geometry) = model
+        (;Grid,NF,nlat_half) = spectral_grid
+        u = zeros(Grid{NF},nlat_half)
+    
+        lat = geometry.latds
+    
+        for ij in eachindex(u)
+            u[ij] = drag.u₀ * exp(-(lat[ij]-drag.latitude)^2/(2*drag.width^2))
+        end
+    
+        û = SpeedyTransforms.spectral(u,one_more_degree=true)
+        v̂ = zero(û)
+        SpeedyTransforms.curl!(drag.ζ₀,û,v̂,model.spectral_transform)
+        return nothing
+    end
+    
+    function SpeedyWeather.drag!(   diagn::SpeedyWeather.DiagnosticVariablesLayer,
+                                    progn::SpeedyWeather.PrognosticVariablesLayer,
+                                    drag::JetDrag,
+                                    time::DateTime,
+                                    model::SpeedyWeather.ModelSetup)
+    
+        (;vor) = progn
+        (;vor_tend) = diagn.tendencies
+        (;ζ₀) = drag
+        
+        (;radius) = model.spectral_grid
+        r = radius/drag.time_scale.value
+        for lm in eachindex(vor,vor_tend,ζ₀)
+            vor_tend[lm] -= r*(vor[lm] - ζ₀[lm])
+        end
+        
+        SpeedyTransforms.spectral_truncation!(vor_tend)    # set lmax+1 to zero
+        
+        return nothing
+    end
+    
+    Base.@kwdef struct StochasticStirring{NF} <: SpeedyWeather.AbstractForcing{NF}
+        
+        # DIMENSIONS from SpectralGrid
+        "Spectral resolution as max degree of spherical harmonics"
+        trunc::Int
+        
+        "Number of latitude rings, used for latitudinal mask"
+        nlat::Int
+    
+        
+        # OPTIONS
+        "Decorrelation time scale τ [days]"
+        decorrelation_time::Second = Day(2)
+    
+        "Stirring strength A [1/s²]"
+        strength::Float64 = 1e-11
+    
+        "Stirring latitude [˚N]"
+        latitude::Float64 = 45
+    
+        "Stirring width [˚]"
+        width::Float64 = 24
+    
+        
+        # TO BE INITIALISED
+        "Stochastic stirring term S"
+        S::LowerTriangularMatrix{Complex{NF}} = zeros(LowerTriangularMatrix{Complex{NF}},trunc+2,trunc+1)
+        
[…]
