### 2023-12-16 f90fcbb
Stochastic stirring video
