### 2023-12-28 716cf72
stochastic stirring extension test with lmin/max & mmin/max
