+++ Project.toml
@@ -47,11 +47,11 @@ LinearAlgebra = "1.8"
 NCDatasets = "0.12, 0.13"
 Primes = "0.5"
 Printf = "1.8"
-ProgressMeter = "^1.7"
+ProgressMeter = "1.7"
 Random = "1.8"
 Statistics = "1.8"
 TOML = "1"
-UnicodePlots = "^3.3"
+UnicodePlots = "3.3"
 julia = "1.8"
 
 [extras]
