+++ .github/workflows/CI.yml
@@ -15,6 +15,7 @@ jobs:
         version:
           - '1.8'
           - '1.9'
+          - '1.10'
         os:
           - ubuntu-latest
         arch:
