### 2024-01-17 d80609d
introduce DEFAULT_RADIUS, _TRUNC, _NLEV
