### 2024-01-17 83f1fe4
adjust time step with radius
