### 2024-01-17 486da34
adapt initialize!(::Leapfrog) accordingly
