### 2024-01-18 90a09fe
dynamics tendencies merged into one file
