+++ src/dynamics/tendencies_dynamics.jl
@@ -1,704 +1,0 @@
-function pressure_gradient!(diagn::DiagnosticVariables,
-                            progn::PrognosticVariables,
-                            lf::Integer,                # leapfrog index
-                            S::SpectralTransform)
-    
-    (;pres) = progn.surface.timesteps[lf]               # log of surface pressure
-    ∇lnp_x_spec = diagn.layers[1].dynamics_variables.a  # reuse work arrays for gradients
-    ∇lnp_y_spec = diagn.layers[1].dynamics_variables.b  # in spectral space
-    (;∇lnp_x, ∇lnp_y) = diagn.surface                   # but store in grid space
-
-    ∇!(∇lnp_x_spec,∇lnp_y_spec,pres,S)                  # CALCULATE ∇ln(pₛ)
-    gridded!(∇lnp_x,∇lnp_x_spec,S,unscale_coslat=true)  # transform to grid: zonal gradient
-    gridded!(∇lnp_y,∇lnp_y_spec,S,unscale_coslat=true)  # meridional gradient
-end
-
-function pressure_flux!(diagn::DiagnosticVariablesLayer,
-                        surf::SurfaceVariables)
-
-    (; ∇lnp_x, ∇lnp_y ) = surf   # zonal, meridional gradient of log surface pressure
-    (; u_grid, v_grid ) = diagn.grid_variables
-    (; uv∇lnp ) = diagn.dynamics_variables
-    
-    @. uv∇lnp = u_grid*∇lnp_x + v_grid*∇lnp_y           # the (u,v)⋅∇lnpₛ term
-end
-
-"""Convert absolute and virtual temperature to anomalies wrt to the reference profile"""
-function temperature_anomaly!(
-    diagn::DiagnosticVariablesLayer,
-    I::ImplicitPrimitiveEq,
-    )
-                    
-    Tₖ = I.temp_profile[diagn.k]    # reference temperature on this layer
-    (; temp_grid, temp_virt_grid ) = diagn.grid_variables
-
-    @. temp_grid -= Tₖ              # absolute temperature -> anomaly
-    @. temp_virt_grid -= Tₖ         # virtual temperature -> anomaly
-end
-
-"""
-    vertical_integration!(Diag::DiagnosticVariables,G::Geometry)
-
-Calculates the vertically averaged (weighted by the thickness of the σ level)
-velocities (*coslat) and divergence. E.g.
-
-    u_mean = ∑_k=1^nlev Δσ_k * u_k
-
-u,v are averaged in grid-point space, divergence in spectral space.
-"""
-function vertical_integration!( diagn::DiagnosticVariables{NF},
-                                progn::PrognosticVariables{NF},
-                                lf::Int,    # leapfrog index for D̄_spec
-                                G::Geometry{NF}) where NF
-    
-    (; σ_levels_thick, nlev ) = G
-    (; ∇lnp_x, ∇lnp_y ) = diagn.surface  # zonal, meridional grad of log surface pressure
-    
-    ū = diagn.surface.u_mean_grid           # rename for convenience
-    v̄ = diagn.surface.v_mean_grid
-    D̄ = diagn.surface.div_mean_grid
-    D̄_spec = diagn.surface.div_mean
-
-    @boundscheck nlev == diagn.nlev || throw(BoundsError)
-
-    fill!(ū,0)     # reset accumulators from previous vertical average
-    fill!(v̄,0)
-    fill!(D̄,0)
-    fill!(D̄_spec,0)
-
-    @inbounds for k in 1:nlev
-
[…]
