### 2024-01-22 3f18779
get_column with parameterization_tendencies! call
