### 2024-01-22 bc38773
Tetens equation and relative humidity centrally calculated
