+++ src/physics/define_column.jl
@@ -62,6 +62,7 @@ Base.@kwdef mutable struct ColumnVariables{NF<:AbstractFloat} <: AbstractColumnV
     # THERMODYNAMICS
     surface_air_density::NF = 0
     const sat_humid::Vector{NF} = zeros(NF,nlev)                # Saturation specific humidity [kg/kg]
+    const rel_humid::Vector{NF} = zeros(NF,nlev)                # Relative humidity [1]
     const sat_vap_pres::Vector{NF} = zeros(NF,nlev)             # Saturation vapour pressure [Pa]
     const dry_static_energy::Vector{NF} = zeros(NF,nlev)        # Dry static energy
     const moist_static_energy::Vector{NF} = zeros(NF,nlev)      # Moist static energy
@@ -125,7 +126,7 @@ Base.@kwdef mutable struct ColumnVariables{NF<:AbstractFloat} <: AbstractColumnV
     qcloud::NF = NF(NaN)       # Equivalent specific humidity of clouds
     fmask::NF = NF(NaN)        # Fraction of land
     # Shortwave radiation: shortwave_radiation
-    rel_hum::Vector{NF} = fill(NF(NaN), nlev) # Relative humidity
+    # rel_hum::Vector{NF} = fill(NF(NaN), nlev) # Relative humidity
     grad_dry_static_energy::NF = NF(NaN)      # gradient of dry static energy
 end
 
