+++ src/physics/thermodynamics.jl
@@ -1,25 +1,33 @@
 """
-Parameters for computing saturation vapour pressure using the August-Roche-Magnus formula,
+Parameters for computing saturation vapour pressure of water using the Tetens' equation,
 
-    eᵢ(T) = e₀ * exp(Cᵢ * (T - T₀) / (T - Tᵢ)),
+    eᵢ(T) = e₀ * exp(Cᵢ * (T - T₀) / (T + Tᵢ)),
 
-where T is in Kelvin and i = 1,2 for saturation with respect to water and ice,
-respectively.
+where T is in Kelvin and i = 1,2 for saturation above and below freezing,
+respectively. From Tetens (1930), and Murray (1967) for below freezing.
 $(TYPEDFIELDS)"""
-Base.@kwdef struct MagnusCoefs{NF<:AbstractFloat}
-    "Saturation vapour pressure at 0°C [Pa]"
-    e₀::NF = 610.8
+Base.@kwdef struct TetensCoefs{NF<:AbstractFloat}
+    "Saturation water vapour pressure at 0°C [Pa]"
+    e₀::NF = 610.78
 
     "0°C in Kelvin"
-    T₀::NF = 273.16
-    T₁::NF = 35.86
-    T₂::NF = 7.66
-    C₁::NF = 17.269
+    T₀::NF = 273.15
+
+    "Tetens denominator (water) [˚C]"
+    T₁::NF = 237.3
+
+    "Tetens denominator following Murray (1967, ice) [˚C]"
+    T₂::NF = 265.5
+
+    "Tetens numerator scaling [1], above freezing"
+    C₁::NF = 17.27
+
+    "Tetens numerator scaling [1], below freezing"
     C₂::NF = 21.875
 end
 
 Base.@kwdef struct Thermodynamics{NF} <: AbstractThermodynamics{NF}
-    magnus_coefs::MagnusCoefs{NF} = MagnusCoefs{NF}()
+    tetens_coefs::TetensCoefs{NF} = TetensCoefs{NF}()
     mol_ratio::NF 
     latent_heat_condensation::NF
     latent_heat_sublimation::NF
@@ -93,12 +101,12 @@ function dry_static_energy!(column::ColumnVariables,constants::DynamicsConstants
 end
 
 """$(TYPEDSIGNATURES)
-Compute (1) the saturation vapour pressure as a function of temperature using the
-August-Roche-Magnus formula,
+Compute (1) the saturation water vapour pressure as a function of temperature using the
+Tetens equation,
 
     eᵢ(T) = e₀ * exp(Cᵢ * (T - T₀) / (T - Tᵢ)),
 
-where T is in Kelvin and i = 1,2 for saturation with respect to water and ice,
+where T is in Kelvin and i = 1,2 for saturation above and below freezing,
 respectively. And (2) the saturation specific humidity according to the formula,
 
     0.622 * e / (p - (1 - 0.622) * e),
@@ -109,20 +117,38 @@ function saturation_humidity!(
     column::ColumnVariables,
     thermodynamics::Thermodynamics,
 )
-    (;sat_humid, sat_vap_pres, pres, temp) = column
+    (;sat_humid, rel_humid, sat_vap_pres, pres, temp, humid) = column
     (;mol_ratio) = thermodynamics      # = mol_mass_vapour/mol_mass_dry_air = 0.622
 
     for k in eachlayer(column)
-        sat_vap_pres[k] = saturation_vapour_pressure(temp[k],thermodynamics.magnus_coefs)
+        sat_vap_pres[k] = saturation_vapour_pressure(temp[k],thermodynamics.tetens_coefs)
[…]
