+++ src/physics/vertical_diffusion.jl
@@ -23,10 +23,10 @@ gradient of static energy wrt to geopotential, see Fortran SPEEDY documentation.
 $(TYPEDFIELDS)"""
 Base.@kwdef struct StaticEnergyDiffusion{NF<:AbstractFloat} <: VerticalDiffusion{NF}
     "time scale for strength"
-    time_scale::Second = Hour(24)
+    time_scale::Second = Hour(6)
 
     "[1] ∂SE/∂Φ, vertical gradient of static energy SE with geopotential Φ"
-    static_energy_lapse_rate::NF = 0.1
+    static_energy_lapse_rate::NF = 0.1          
     
     # precomputations
     Fstar::Base.RefValue{NF} = Ref(zero(NF))    # excluding the surface pressure pₛ
@@ -56,14 +56,21 @@ function static_energy_diffusion!(  column::ColumnVariables,
     pₛ = column.pres[end]               # surface pressure
     Fstar = scheme.Fstar[]*pₛ
     Γˢᵉ = scheme.static_energy_lapse_rate 
-    
+
     # relax static energy profile back to a reference gradient Γˢᵉ
     @inbounds for k in 1:nlev-1
         # Fortran SPEEDY doc eq (74)
         SEstar = dry_static_energy[k+1] + Γˢᵉ*(geopot[k] - geopot[k+1])
         SE = dry_static_energy[k]
         if SE < SEstar
-            flux_temp_upward[k+1] += Fstar*(SEstar - SE)
+            # SPEEDY code applies the flux to all layers below
+            # effectively fluxing temperature from below the surface to
+            # given height. It doesn't really make sense to me in
+            # a meteorology sense to me since it can flux through stable layers,
+            # but it's more stable and that's good enough for now.
+            for kk = k+1:nlev
+                flux_temp_upward[kk] += Fstar*(SEstar - SE)
+            end
         end
     end
 end
@@ -79,9 +86,6 @@ Base.@kwdef struct HumidityDiffusion{NF<:AbstractFloat} <: VerticalDiffusion{NF}
     "[1] ∂RH/∂σ, vertical gradient of relative humidity RH wrt sigma coordinate σ"
     humidity_gradient::NF = 0.5
 
-    "[1] disable above this σ level"
-    σ_min::NF = 0.5
-    
     # precomputations
     Fstar::Base.RefValue{NF} = Ref(zero(NF))    # excluding the surface pressure pₛ
 end
@@ -99,6 +103,9 @@ function initialize!(   scheme::HumidityDiffusion,
     
     # Fortran SPEEDY documentation equation (70), excluding the surface pressure pₛ
     scheme.Fstar[] = C₀/gravity/scheme.time_scale.value
+
+    # SPEEDY code version
+    # scheme.Fstar[] = C₀/scheme.time_scale.value
 end
 
 # function barrier
@@ -124,114 +131,33 @@ function humidity_diffusion!(   column::ColumnVariables,
                                 scheme::HumidityDiffusion,
                                 geometry::Geometry)
     
-    (;nlev, humid, sat_humid, flux_humid_upward) = column
+    (;nlev, sat_humid, rel_humid, flux_humid_upward) = column
     pₛ = column.pres[end]               # surface pressure
-    Fstar = scheme.Fstar[]*pₛ
+    Fstar = scheme.Fstar[]*pₛ           
+    # Fstar = scheme.Fstar[]              # SPEEDY code version
     Γ = scheme.humidity_gradient        # relative humidity wrt σ
     σ = geometry.σ_levels_full
-
[…]
