### 2024-01-23 b551600
relative_humidity convenience func
