### 2024-01-23 0659d4c
large_scale_condensation with max heating
