### 2024-01-24 9ce15b3
provide default SpectralGrid for models
