### 2024-01-26 b3179b8
Solar zenith angle calculation
