+++ src/abstract_types.jl
@@ -33,6 +33,11 @@ abstract type AbstractDrag{NF} end
 # VERTICAL ADVECTION (PrimitiveEquation)
 abstract type VerticalAdvection{NF,B} end
 
+# SOLAR RADIATION
+abstract type AbstractSolarDeclination{NF} end
+abstract type AbstractSolarTimeCorrection{NF} end
+abstract type AbstractZenith{NF,Grid} end
+
 # PARAMETERIZATIONS
 abstract type AbstractParameterization{NF} end
 abstract type BoundaryLayerDrag{NF} <: AbstractParameterization{NF} end
