### 2024-01-26 dffcba4
daily cycle, seasonal cycle options for SolarZenithAngle
