+++ src/physics/zenith.jl
@@ -1,3 +1,34 @@
+"""Coefficients to calculate the solar declination angle δ [radians] based on a simple
+sine function, with Earth's axial tilt as amplitude, equinox as phase shift.
+$(TYPEDFIELDS)"""
+Base.@kwdef struct SinSolarDeclination{NF} <: AbstractSolarDeclination{NF}
+    axial_tilt::NF = 23.44
+    equinox::DateTime = DateTime(2000,3,20)
+    length_of_year::Second = Day(365.25)
+    length_of_day::Second = Hour(24)
+end
+
+"""Generator function pulling the number format NF from a SpectralGrid."""
+SinSolarDeclination(SG::SpectralGrid;kwargs...) = SinSolarDeclination{SG.NF}(;kwargs...)
+
+"""Generator function using the planet's orbital parameters to adapt the
+solar declination calculation."""
+function SinSolarDeclination(SG::SpectralGrid,P::AbstractPlanet)
+    (;axial_tilt, equinox, length_of_year, length_of_day) = P
+    SinSolarDeclination{SG.NF}(;axial_tilt, equinox, length_of_year, length_of_day)
+end
+
+"""
+$(TYPEDSIGNATURES)
+SinSolarDeclination functor, computing the solar declination angle of
+angular fraction of year g [radians] using the coefficients of the
+SinSolarDeclination struct."""
+function (S::SinSolarDeclination)(g::NF) where NF
+    axial_tilt = deg2rad(S.axial_tilt)
+    equinox = S.length_of_day.value*Dates.dayofyear(S.equinox)/S.length_of_year.value
+    return axial_tilt*sin(g-2*(π*convert(NF,equinox)))
+end
+
 """Coefficients to calculate the solar declination angle δ from
 
     δ = 0.006918    - 0.399912*cos(g)  + 0.070257*sin(g)
@@ -39,6 +70,9 @@ function Base.show(io::IO,L::AbstractSolarDeclination)
     print_fields(io,L,keys)
 end
 
+"""Coefficients for the solar time correction (also called
+Equation of time) which adjusts the solar hour to an oscillation
+of sunrise/set by about +-16min throughout the year."""
 Base.@kwdef struct SolarTimeCorrection{NF} <: AbstractSolarTimeCorrection{NF}
     a::NF =   0.004297      # the offset +
     s1::NF = -1.837877      # s1*sin(g) +
@@ -47,6 +81,10 @@ Base.@kwdef struct SolarTimeCorrection{NF} <: AbstractSolarTimeCorrection{NF}
     c2::NF = -0.837378      # c2*cos(2g)
 end
 
+"""
+$(TYPEDSIGNATURES)
+Functor that returns the time correction for a angular
+fraction of the year g [radians], so that g=0 for Jan-01 and g=2π for Dec-31."""
 function (STC::SolarTimeCorrection)(g)
     (;a,s1,s2,c1,c2) = STC
     sin1g,cos1g = sincos(g)
@@ -60,6 +98,8 @@ function Base.show(io::IO,L::AbstractSolarTimeCorrection)
     print_fields(io,L,keys)
 end
 
+"""Struct containing all parameters for the calculation of the
+cos of the solar zenith angle, cos_zenith."""
 Base.@kwdef struct SolarZenithAngle{NF<:AbstractFloat,Grid<:AbstractGrid{NF}} <: AbstractZenith{NF,Grid}
     # DIMENSIONS
     nlat_half::Int
@@ -71,19 +111,21 @@ Base.@kwdef struct SolarZenithAngle{NF<:AbstractFloat,Grid<:AbstractGrid{NF}} <:
     length_of_year::Second = Day(365.25)
 
     # COEFFICIENTS
[…]
