+++ src/dynamics/spectral_grid.jl
@@ -121,6 +121,9 @@ Base.@kwdef struct Geometry{NF<:AbstractFloat} <: AbstractGeometry{NF}     # NF:
     "array of colatitudes in radians (0...π)"
     colat::Vector{Float64} = get_colat(Grid,nlat_half)
 
+    "array of latitudes in radians (π...-π)"
+    lat::Vector{NF} = get_lat(Grid,nlat_half)
+
     "array of latitudes in degrees (90˚...-90˚)"
     latd::Vector{Float64} = get_latd(Grid,nlat_half)
 
