### 2024-01-26 0bb8706
initialize!  call with model
