### 2024-01-30 df87a5b
SolarZenith, SolarZenithSeason
