+++ src/physics/zenith.jl
@@ -98,17 +98,43 @@ function Base.show(io::IO,L::AbstractSolarTimeCorrection)
     print_fields(io,L,keys)
 end
 
-"""Struct containing all parameters for the calculation of the
-cos of the solar zenith angle, cos_zenith."""
-Base.@kwdef struct SolarZenithAngle{NF<:AbstractFloat,Grid<:AbstractGrid{NF}} <: AbstractZenith{NF,Grid}
+"""
+$(TYPEDSIGNATURES)
+Chooses from SolarZenith (daily and seasonal cycle), SolarZenithDay (daily cycle only),
+SolarZenithSeason (seasonal cycle only) and SolarZenithConstant (no daily or seasonal cycle)
+given the parameters in model.planet."""
+function WhichZenith(SG::SpectralGrid,P::AbstractPlanet;kwargs...)
+    (;NF, Grid, nlat_half) = SG
+    (;daily_cycle, seasonal_cycle, length_of_day, length_of_year) = P
+    solar_declination = SinSolarDeclination(SG,P)
+
+    if daily_cycle
+        return SolarZenith{NF,Grid{NF}}(;
+            nlat_half, length_of_day, length_of_year, solar_declination, seasonal_cycle, kwargs...)
+
+    else
+        return SolarZenithSeason{NF,Grid{NF}}(;
+            nlat_half, length_of_day, length_of_year, solar_declination, seasonal_cycle, kwargs...)
+    end
+end
+
+function Base.show(io::IO,L::AbstractZenith)
+    println(io,"$(typeof(L)) <: AbstractZenith")
+    keys = propertynames(L)
+    print_fields(io,L,keys)
+end
+
+"""Solar zenith angle varying with daily and seasonal cycle.
+$(TYPEDFIELDS)"""
+Base.@kwdef struct SolarZenith{NF<:AbstractFloat,Grid<:AbstractGrid{NF}} <: AbstractZenith{NF,Grid}
     # DIMENSIONS
     nlat_half::Int
 
     # OPTIONS
-    daily_cycle::Bool = true
     length_of_day::Second = Hour(24)
-    seasonal_cycle::Bool = true
     length_of_year::Second = Day(365.25)
+    equation_of_time::Bool = true
+    seasonal_cycle::Bool = true
 
     # COEFFICIENTS
     solar_declination::SinSolarDeclination{NF} = SinSolarDeclination{NF}()
@@ -119,56 +145,62 @@ Base.@kwdef struct SolarZenithAngle{NF<:AbstractFloat,Grid<:AbstractGrid{NF}} <:
     cos_zenith::Grid = zeros(Grid,nlat_half)
 end
 
-# generator function
-function SolarZenithAngle(SG::SpectralGrid,P::AbstractPlanet;kwargs...)
-    (;NF, Grid, nlat_half) = SG
-    (;daily_cycle, seasonal_cycle, length_of_day, length_of_year) = P
-    solar_declination = SinSolarDeclination(SG,P)
-    SolarZenithAngle{NF,Grid{NF}}(;nlat_half,daily_cycle,seasonal_cycle,
-        length_of_day,length_of_year,solar_declination,kwargs...)
-end
-
-function Base.show(io::IO,L::AbstractZenith)
-    println(io,"$(typeof(L)) <: AbstractZenith")
-    keys = propertynames(L)
-    print_fields(io,L,keys)
-end
-
 function initialize!(
-    S::SolarZenithAngle,
+    S::AbstractZenith,
     initial_time::DateTime,
     model::ModelSetup
 )
-    S.initial_time[] = initial_time
+    S.initial_time[] = initial_time     # to fix the season if no seasonal cycle
     cos_zenith!(S,initial_time,model.geometry)
 end
 
+"""
+$(TYPEDSIGNATURES)
[…]
