### 2024-01-30 e295f8c
add cos_zenith! call to physics tendencies
