### 2024-01-30 62cc41d
latitude -pi/2 to pi/2 typo
Co-authored-by: Simone Silvestri <silvestri.simone0@gmail.com>
