### 2024-01-30 #6121 @trontrytel
I work with some of the co-authors of this paper on a different project. Is that considered conflict of interests. I know that Julia Earth sciences modeling community is not that big, but maybe its too close?
