### 2024-01-30 3b3affb
add boundserror check
