### 2024-02-01 #6121 @kthyng
Here is our conflict of interest policy: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy
Please take a look and see if it applies to you.

Indeed it's hard to find people who aren't conflicted! I already skipped past all the MIT Julia people and hoped by going to a different group I'd get some success.

@trontrytel and @charleskawczynski if you are both conflicted could you suggest someone you think might not be?
