+++ src/dynamics/models.jl
@@ -255,12 +255,12 @@ Base.@kwdef mutable struct PrimitiveWetModel{NF<:AbstractFloat, D<:AbstractDevic
 
     # PHYSICS/PARAMETERIZATIONS
     physics::Bool = true
-    thermodynamics::Thermodynamics{NF} = Thermodynamics(spectral_grid,atmosphere)
+    clausis_clapeyron::AbstractClausiusClapeyron{NF} = ClausiusClapeyron(spectral_grid,atmosphere)
     boundary_layer_drag::BoundaryLayerDrag{NF} = NoBoundaryLayerDrag(spectral_grid)
     temperature_relaxation::TemperatureRelaxation{NF} = HeldSuarez(spectral_grid)
     static_energy_diffusion::VerticalDiffusion{NF} = StaticEnergyDiffusion(spectral_grid)
     humidity_diffusion::VerticalDiffusion{NF} = HumidityDiffusion(spectral_grid)
-    large_scale_condensation::AbstractCondensation{NF} = SpeedyCondensation(spectral_grid)
+    large_scale_condensation::AbstractCondensation{NF} = ImmediateCondensation(spectral_grid)
     surface_thermodynamics::AbstractSurfaceThermodynamics{NF} = SurfaceThermodynamicsConstant(spectral_grid)
     surface_wind::AbstractSurfaceWind{NF} = SurfaceWind(spectral_grid)
     surface_heat_flux::AbstractSurfaceHeat{NF} = SurfaceSensibleHeat(spectral_grid)
