### 2024-02-01 68e07b8
Immediate condensation from Frierson 2006
