### 2024-02-01 89e73c3
disable vertical diffusion by default
