### 2024-02-02 #6121 @trontrytel
> Here is our conflict of interest policy: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy Please take a look and see if it applies to you.

I think both me and @charleskawczynski should not be the reviewers then.

> Indeed it's hard to find people who aren't conflicted! I already skipped past all the MIT Julia people and hoped by going to a different group I'd get some success.

:) It's even worse for cloud parameterizations. We are a small community.

> @trontrytel and @charleskawczynski if you are both conflicted could you suggest someone you think might not be?

@slayoo will be a great fit for reviewing this. I think another great reviewer would be @claresinger
