### 2024-02-02 #6121 @Datseris
I can review but in one month from now (March) because at the moment I am committed to another JOSS review.
