+++ src/dynamics/vertical_coordinates.jl
@@ -32,7 +32,14 @@ Vertical sigma coordinates defined by their nlev+1 half levels `σ_levels_half`.
 fraction of surface pressure (p/p0) and are sorted from top (stratosphere) to bottom (surface).
 The first half level is at 0 the last at 1. Default levels are equally spaced from 0 to 1 (including)."""
 function default_sigma_coordinates(nlev::Integer)
-    σ_half = collect(range(0,1,nlev+1))     # equi-distant in σ
+
+    # equi-distant in σ
+    # σ_half = collect(range(0,1,nlev+1))
+
+    # Frierson 2006 spacing, higher resolution in surface boundary layer and in stratosphere
+    z = collect(range(1,0,nlev+1))
+    σ_half = @. exp(-5*(0.05*z + 0.95*z^3))
+    σ_half[1] = 0
     return σ_half
 end
 
