+++ src/physics/surface_fluxes.jl
@@ -54,12 +54,12 @@ Base.@kwdef struct SurfaceWind{NF<:AbstractFloat} <: AbstractSurfaceWind{NF}
     # TODO make this orography dependent
     "Drag coefficient over land (orography = 0) [1]"
     drag_land::NF = 2.4e-3
-
+    
     "Drag coefficient over sea [1]"
     drag_sea::NF = 1.8e-3
 
-    # "Flux limiter [kg/m/s²]"
-    # max_flux::NF = 1.5
+    "Flux limiter [kg/m/s²]"
+    max_flux::NF = 0.1
 end
 
 SurfaceWind(SG::SpectralGrid;kwargs...) = SurfaceWind{SG.NF}(;kwargs...)
@@ -69,7 +69,7 @@ function surface_wind_stress!(  column::ColumnVariables{NF},
 
     (;land_fraction) = column
     (;f_wind, V_gust, drag_land, drag_sea) = surface_wind
-    # (;max_flux) = surface_wind
+    (;max_flux) = surface_wind
 
     # SPEEDY documentation eq. 49
     column.surface_u = f_wind*column.u[end] 
@@ -84,22 +84,23 @@ function surface_wind_stress!(  column::ColumnVariables{NF},
     V₀ = column.surface_wind_speed
     drag = land_fraction*drag_land + (1-land_fraction)*drag_sea
 
-    # add flux limiter to avoid heavy drag in initial shock
-    # u_flux = sign(u[end])*min(abs(ρdragV₀*u[end]),max_flux)
-    # v_flux = sign(v[end])*min(abs(ρdragV₀*v[end]),max_flux)
-    # column.flux_u_upward[end] -= u_flux
-    # column.flux_v_upward[end] -= v_flux
-    
+    # add flux limiter to avoid heavy drag in (initial) shock
+    u_flux = ρ*drag*V₀*surface_u
+    v_flux = ρ*drag*V₀*surface_v
+    column.flux_u_upward[end] -= clamp(u_flux,-max_flux,max_flux)
+    column.flux_v_upward[end] -= clamp(v_flux,-max_flux,max_flux)
+
     # SPEEDY documentation eq. 52, 53, accumulate fluxes with +=
-    column.flux_u_upward[end] -= ρ*drag*V₀*surface_u
-    column.flux_v_upward[end] -= ρ*drag*V₀*surface_v
+    # column.flux_u_upward[end] -= ρ*drag*V₀*surface_u
+    # column.flux_v_upward[end] -= ρ*drag*V₀*surface_v
     
     return nothing
 end
 
 Base.@kwdef struct SurfaceSensibleHeat{NF<:AbstractFloat} <: AbstractSurfaceHeat{NF}
-    heat_exchange_land::Float64 = 1.2e-3    # for neutral stability
-    heat_exchange_sea::Float64 = 0.9e-3
+    heat_exchange_land::NF = 1.2e-3    # for neutral stability
+    heat_exchange_sea::NF = 0.9e-3
+    max_flux::NF = 100
 end
 
 SurfaceSensibleHeat(SG::SpectralGrid;kwargs...) = SurfaceSensibleHeat{SG.NF}(;kwargs...)
@@ -108,8 +109,7 @@ function sensible_heat_flux!(   column::ColumnVariables{NF},
                                 heat_flux::SurfaceSensibleHeat,
                                 C::DynamicsConstants) where NF
     (;cₚ) = C
-    heat_exchange_land = convert(NF,heat_flux.heat_exchange_land)
-    heat_exchange_sea  = convert(NF,heat_flux.heat_exchange_sea)
+    (;heat_exchange_land, heat_exchange_sea, max_flux) = heat_flux
 
     ρ = column.surface_air_density
[…]
