### 2024-02-02 5c471a5
Make equispaced sigma default again
