### 2024-02-03 a35cb86
Calculate bulk Richardson number
