### 2024-02-05 4f7ac6a
bulk richardson typo
