+++ src/physics/boundary_layer.jl
@@ -1,3 +1,5 @@
+abstract type BoundaryLayerDrag{NF} <: AbstractParameterization{NF} end
+
 """Concrete type that disables the boundary layer drag scheme."""
 struct NoBoundaryLayerDrag{NF} <: BoundaryLayerDrag{NF} end
 
@@ -9,15 +11,16 @@ function initialize!(   scheme::NoBoundaryLayerDrag,
     return nothing
 end 
 
-# function barrier
+# function barrier for all boundary layer drags
 function boundary_layer_drag!(  column::ColumnVariables,
                                 model::PrimitiveEquation)
-    boundary_layer_drag!(column,model.boundary_layer_drag)
+    boundary_layer_drag!(column,model.boundary_layer_drag,model)
 end
 
 """NoBoundaryLayer scheme just passes."""
 function boundary_layer_drag!(  column::ColumnVariables,
-                                scheme::NoBoundaryLayerDrag)
+                                scheme::NoBoundaryLayerDrag,
+                                model::PrimitiveEquation)
     return nothing
 end
 
@@ -57,6 +60,13 @@ function initialize!(   scheme::LinearDrag,
     end
 end 
 
+# function barrier
+function boundary_layer_drag!(  column::ColumnVariables,
+                                scheme::LinearDrag,
+                                model::PrimitiveEquation)
+    boundary_layer_drag!(column,scheme)
+end
+
 """
 $(TYPEDSIGNATURES)
 Compute tendency for boundary layer drag of a `column` and add to its tendencies fields"""
@@ -74,3 +84,80 @@ function boundary_layer_drag!(  column::ColumnVariables,
         end
     end
 end
+
+Base.@kwdef struct ConstantDrag{NF} <: BoundaryLayerDrag{NF}
+    drag::NF = 1e-3
+end
+
+ConstantDrag(SG::SpectralGrid;kwargs...) = ConstantDrag{SG.NF}(;kwargs...)
+
+initialize!(::ConstantDrag,::PrimitiveEquation) = nothing
+
+function boundary_layer_drag!(  column::ColumnVariables,
+                                scheme::ConstantDrag,
+                                model::PrimitiveEquation)
+    column.boundary_layer_drag = scheme.drag
+end
+
+"""Boundary layer drag coefficient from the bulk Richardson number,
+following Frierson, 2006, Journal of the Atmospheric Sciences.
+$(TYPEDFIELDS)"""
+Base.@kwdef struct BulkRichardsonDrag{NF} <: BoundaryLayerDrag{NF}
+    "von Kármán constant [1]"
+    κ::NF = 0.4
+
+    "roughness length [m]"
+    z₀::NF = 3.21e-5
+
+    "Critical Richardson number for stable mixing cutoff [1]"
+    Ri_c::NF = 1
+
+    "Maximum drag coefficient, κ²/log(zₐ/z₀) for zₐ from reference temperature"
+    drag_max::Base.RefValue{NF} = Ref(zero(NF))
+end
+
+BulkRichardsonDrag(SG::SpectralGrid,kwargs...) = BulkRichardsonDrag{SG.NF}(;kwargs...)
+
+function initialize!(scheme::BulkRichardsonDrag, model::PrimitiveEquation)
+
+    # Typical height Z of lowermost layer from geopotential of reference surface temperature
+    # minus surface geopotential (orography * gravity)
+    (;temp_ref) = model.atmosphere
+    (;Δp_geopot_full, gravity) = model.constants
+    Z = temp_ref*Δp_geopot_full[end] / gravity
+    
[…]
