### 2024-02-07 890eef3
column with orography and geopotential
