+++ src/abstract_types.jl
@@ -32,7 +32,6 @@ abstract type AbstractDrag{NF} end
 
 # PARAMETERIZATIONS
 abstract type AbstractParameterization{NF} end
-abstract type BoundaryLayerDrag{NF} <: AbstractParameterization{NF} end
 abstract type TemperatureRelaxation{NF} <: AbstractParameterization{NF} end
 abstract type VerticalDiffusion{NF} <: AbstractParameterization{NF} end
 abstract type AbstractCondensation{NF} <: AbstractParameterization{NF} end
