### 2024-02-07 fafb262
column with orography and surface geopotential
