### 2024-02-07 3c388de
get_column arguments with model
