### 2024-02-07 33d1a21
calculate column geopotential from virtual temperature
