+++ src/physics/thermodynamics.jl
@@ -113,6 +113,7 @@ end
 $(TYPEDSIGNATURES)
 Calculate the dry static energy for the primitive dry model."""
 function get_thermodynamics!(column::ColumnVariables,model::PrimitiveDry)
+    # use temp here as temp = temp_virt
     geopotential!(column.geopot,column.temp,model.constants)
     dry_static_energy!(column, model.constants)
 end
@@ -123,7 +124,7 @@ Calculate thermodynamic quantities like saturation vapour pressure,
 saturation specific humidity, dry static energy, moist static energy
 and saturation moist static energy from the prognostic column variables."""
 function get_thermodynamics!(column::ColumnVariables,model::PrimitiveWet)
-    geopotential!(column.geopot, column.temp, model.constants, column.surface_geopotential)
+    geopotential!(column.geopot, column.temp_virt, model.constants, column.surface_geopotential)
     dry_static_energy!(column, model.constants)
     saturation_humidity!(column, model.clausis_clapeyron)
     moist_static_energy!(column, model.clausis_clapeyron)
