### 2024-02-13 #6323 @kthyng
Hi @vavrines, @natgeo-wong, @slayoo! I see some progress in these checklists, thanks for that! Feel free to open issues in the software repo as detailed questions come up and link them back to this review issue. Thanks.
