### 2024-02-13 3e96248
make Simplified BettsMiller convection default
