+++ src/physics/convection.jl
@@ -81,12 +81,20 @@ function initialize!(convection::SpeedyConvection,model::PrimitiveWet)
     entrainment_profile .*= convection.max_entrainment/entrainment_profile[nlev-1]
 end
 
-# dry model doesn't have convection
+# dry model doesn't have convection TODO maybe it should?
 convection!(column::ColumnVariables,model::PrimitiveDry) = nothing
 
-# function barrier
+# function barrier for all AbstractConvection
 function convection!(
     column::ColumnVariables,
+    model::PrimitiveEquation,
+)
+    convection!(column,model.convection,model)
+end
+
+function convection!(
+    column::ColumnVariables,
+    scheme::SpeedyConvection,
     model::PrimitiveEquation,
 )
     # always diagnose convection
@@ -311,4 +319,196 @@ function convection!(
     return nothing
 end
 
+Base.@kwdef struct SimplifiedBettsMiller{NF} <: AbstractConvection{NF}
+    "number of vertical layers/levels"
+    nlev::Int
 
+    "Relaxation time for profile adjustment"
+    time_scale::Second = Hour(4)
+
+    "Relative humidity for reference profile"
+    relative_humidity::NF = 0.7
+
+    "temperature [K] reference profile to adjust to"
+    temp_ref_profile::Vector{NF} = zeros(NF,nlev)
+
+    "specific humidity [kg/kg] profile to adjust to"
+    humid_ref_profile::Vector{NF} = zeros(NF,nlev)
+end
+
+SimplifiedBettsMiller(SG::SpectralGrid;kwargs...) = SimplifiedBettsMiller{SG.NF}(nlev=SG.nlev;kwargs...)
+
+initialize!(::SimplifiedBettsMiller,::PrimitiveWet) = nothing
+
+# function barrier to unpack model
+function convection!(
+    column::ColumnVariables,
+    scheme::SimplifiedBettsMiller,
+    model::PrimitiveEquation,
+)
+    convection!(column,scheme,model.clausis_clapeyron,model.geometry)
+end
+
+function convection!(
+    column::ColumnVariables{NF},
+    SBM::SimplifiedBettsMiller,
+    clausius_clapeyron::AbstractClausiusClapeyron,
+    geometry::Geometry
+) where NF
+
+    σ = geometry.σ_levels_full
+    σ_half = geometry.σ_levels_half
+    Δσ = geometry.σ_levels_thick
+    (;temp_ref_profile, humid_ref_profile) = SBM
+    (;geopot, nlev, temp, humid, temp_tend, humid_tend) = column
+    pₛ = column.pres[end]
+    (;Lᵥ, cₚ) = clausius_clapeyron
+
+    # CONVECTIVE CRITERIA AND FIRST GUESS RELAXATION
+    temp_parcel = temp[nlev]
+    humid_parcel = humid[nlev]
+    level_zero_buoyancy = moist_adiabat!(temp_ref_profile,
+                                            temp_parcel, humid_parcel,
+                                            temp, geopot, pₛ, σ,
+                                            clausius_clapeyron)
+            
+    for k in level_zero_buoyancy:nlev
+        qsat = saturation_humidity(temp_ref_profile[k],pₛ*σ[k],clausius_clapeyron)
+        humid_ref_profile[k] = qsat*SBM.relative_humidity
+    end
+
+    local Pq::NF = 0        # precipitation due to drying
+    local PT::NF = 0        # precipitation due to cooling
+    local ΔT::NF = 0        # vertically uniform temperature profile adjustment
+    local Qref::NF = 0      # = ∫_pₛ^p_LZB -humid_ref_profile dp
+
+    # skip constants compared to Frierson 2007, i.e. no /τ, /gravity, *cₚ/Lᵥ
+    for k in level_zero_buoyancy:nlev
[…]
