### 2024-02-13 e38f13b
add heat capacity to ClausiusClapeyron
