### 2024-02-14 7ead7e9
pull virtual temperature from diagnostic variables into column
