### 2024-02-14 SpeedyWeather/SpeedyWeather.jl#443 @natgeo-wong
[JOSS review] Nat's Review of SpeedyWeather.jl
Review of SpeedyWeather.jl by Nat (https://github.com/openjournals/joss-reviews/issues/6323#issuecomment-1927459691)

See the pdf for detailed comments.
[reviewv1.pdf](https://github.com/SpeedyWeather/SpeedyWeather.jl/files/14288668/reviewv1.pdf)

Current outstanding issues on the checklist:
     - [x] We need a more detailed explanation of the different modules / `AbstractTypes` available for the different models
     - [x] which of these`AbstractType` modules are relevant for each of the different types of models (e.g. `ShallowWaterModel`, `PrimitiveDryModel`) available in SpeedyWeather.jl
    - [x] A reorganization of the documentation into a more hierarchical form
