### 2024-02-14 SpeedyWeather/SpeedyWeather.jl#443 @natgeo-wong
https://github.com/openjournals/joss-reviews/issues/6323
