### 2024-02-14 SpeedyWeather/SpeedyWeather.jl#443 @milankl
Nat also suggests looking into [Paulius and Garner, 2006](https://doi.org/10.1175/JAS3705.1) for a simple radiation scheme
