### 2024-02-14 20693bb
0.8 as higher relative humidity for SimplifiedBettsMiller
