### 2024-02-14 be29d70
calculate saturation humidity only in large_scale_condensation!
