### 2024-02-14 3f5d49f
get_thermodynamics! now same for wet and dry models
