### 2024-02-15 112b346
remove SpeedyConvection scheme
