+++ src/physics/convection.jl
@@ -1,328 +1,5 @@
 abstract type AbstractConvection{NF} <: AbstractParameterization{NF} end
 
-Base.@kwdef struct SpeedyConvection{NF} <: AbstractConvection{NF}
-
-    "Number of vertical levels"
-    nlev::Int
-
-    "Minimum (normalised) surface pressure for the occurrence of convection [1]"
-    pres_threshold::NF = 0.8
-
-    "Relative humidity threshold for convection in PBL [1]"
-    humid_threshold_boundary::NF = 0.9
-
-    "Relative humidity threshold for convection in the troposphere [1]"
-    humid_threshold_troposphere::NF = 0.7
-
-    "Relaxation time for PBL humidity"
-    time_scale::Second = Hour(6)
-
-    "Maximum entrainment as a fraction of cloud-base mass flux"
-    max_entrainment::NF = 0.5
-
-    "Ratio between secondary and primary mass flux at cloud-base"
-    ratio_secondary_mass_flux::NF = 0.8
-
-    "Super saturation of relative humidity in the planteray boundar layer [1]"
-    super_saturation::NF = 1.01
-
-    "Mass flux limiter for humidity [kg/kg]"
-    mass_flux_limiter_humid::NF = 5e-3
-
-    # precomputed in initialize!
-    "Reference surface pressure [Pa]"
-    pres_ref::Base.RefValue{NF} = Base.Ref(zero(NF))
-
-    "latent heat of condensation [J/kg] for consistency with specific humidity [kg/kg], also called alhc"
-    latent_heat_condensation::Base.RefValue{NF} = Base.Ref(zero(NF))
-
-    "specific heat [J/kg/K] of air"
-    specific_heat::Base.RefValue{NF} = Base.Ref(zero(NF))
-
-    "Number of vertical levels for stratosphere"
-    n_stratosphere_levels::Base.RefValue{Int} = Base.Ref(0)
-
-    "Number of vertical levels for planetary boundary layer"
-    n_boundary_levels::Base.RefValue{Int} = Base.Ref(0)
-
-    "Mass flux entrainment profile in the vertical [1]"
-    entrainment_profile::Vector{NF} = zeros(NF,nlev)
-end
-
-SpeedyConvection(SG::SpectralGrid;kwargs...) = SpeedyConvection{SG.NF}(nlev=SG.nlev;kwargs...)
-
-function initialize!(convection::SpeedyConvection,model::PrimitiveWet)
-
-    (;σ_levels_full) = model.geometry
-    (;σ_tropopause, σ_boundary_layer) = model.atmosphere
-    (;entrainment_profile, nlev) = convection
-
-    # number of stratospheric levels, for nlev very small n can be nothing, then use 0
-    n = findlast(σ->σ<=σ_tropopause,σ_levels_full)
-    convection.n_stratosphere_levels[] = isnothing(n) ? 0 : n
-
-    # number of levels for the planetary boundary layer, same as above
-    n = findlast(σ->σ<=σ_boundary_layer,σ_levels_full)
-    convection.n_boundary_levels[] = isnothing(n) ? 0 : nlev - n
-    
-    # reference pressure
-    convection.pres_ref[] = model.atmosphere.pres_ref*100     # [hPa] -> [Pa]
-    convection.latent_heat_condensation[] = model.atmosphere.latent_heat_condensation
-    convection.specific_heat[] = model.atmosphere.cₚ
-
-    # Mass entrainment profile
-    entrainment_profile[1] = 0      # no entrainment in top layer
-    entrainment_profile[nlev] = 0   # no entrainment in bottom layer
-    for k = 2:nlev-1                # intermediate layers with minimum at σ=0
[…]
