### 2024-02-15 281b55c
remove unused tests for old schemes
