### 2024-02-15 06a5edd
UniformCooling and TransparentShortwave defined, old removed
