+++ src/physics/define_column.jl
@@ -64,84 +64,13 @@ Base.@kwdef mutable struct ColumnVariables{NF<:AbstractFloat} <: AbstractColumnV
     # THERMODYNAMICS
     surface_air_density::NF = 0
     const sat_humid::Vector{NF} = zeros(NF,nlev)                # Saturation specific humidity [kg/kg]
-    const rel_humid::Vector{NF} = zeros(NF,nlev)                # Relative humidity [1]
-    const dry_static_energy::Vector{NF} = zeros(NF,nlev)        # Dry static energy
-    const moist_static_energy::Vector{NF} = zeros(NF,nlev)      # Moist static energy
-    const sat_moist_static_energy::Vector{NF} = zeros(NF,nlev)  # Saturation moist static energy
+    const dry_static_energy::Vector{NF} = zeros(NF,nlev)        # Dry static energy [J/kg]
     const temp_virt::Vector{NF} = zeros(NF,nlev)                # virtual temperature [K]
     const bulk_richardson::Vector{NF} = zeros(NF,nlev)          # bulk richardson number [1]
     const geopot::Vector{NF} = zeros(NF,nlev)                   # gepotential height [m]
 
-    # and interpolated to half levels
-    const humid_half::Vector{NF} = zeros(NF,nlev)                    # Specific humidity interpolated to half-levels
-    const sat_humid_half::Vector{NF} = zeros(NF,nlev)                # Saturation specific humidity interpolated to half-levels
-    const dry_static_energy_half::Vector{NF} = zeros(NF,nlev)        # Dry static energy interpolated to half-levels
-    const sat_moist_static_energy_half::Vector{NF} = zeros(NF,nlev)  # Saturation moist static energy interpolated to half-levels
-
-    # Convection
-    conditional_instability::Bool = false                       # in this column? (condition 1)
-    activate_convection::Bool = false                           # in this column? (condition 2)
-    cloud_top::Int = nlev+1                                     # Top-of-convection layer
-    excess_humid::NF = 0                                        # due to convection
-    cloud_base_mass_flux::NF = 0                                # Mass flux at the top of the PBL
-    precip_convection::NF = 0                                   # Precipitation due to convection [m]
-
-    # Large-scale condensation
-    precip_large_scale::NF = 0                                  # precipitation [m]
-
-    # Longwave radiation
-    ## New vars in radlw_down!
-    wvi::Matrix{NF} = fill(NF(NaN), nlev, 2)  # Weights for vertical interpolation
-    tau2::Matrix{NF} = fill(NF(NaN), nlev, nband) # Transmissivity of atmospheric layers
-    dfabs::Vector{NF} = fill(NF(NaN), nlev)   # Flux of sw rad. absorbed by each atm. layer
-    fsfcd::NF = NaN                       # Downward-only flux of sw rad. at the surface
-    st4a::Matrix{NF} = fill(NF(NaN), nlev, 2) # Blackbody emission from full and half atmospheric levels
-    flux::Vector{NF} = fill(NF(NaN), nband)       # Radiative flux in different spectral bands
-
-    ## New vars in compute_bbe!
-    fsfcu::NF = NaN # surface blackbody emission (upward)
-    ts::NF = NaN    # surface temperature
-
-    ## New vars in radlw_up!
-    fsfc::NF = NaN # Net (downw.) flux of sw rad. at the surface
-    ftop::NF = NaN # Net (downw.) flux of sw rad. at the atm. top
[…]
