+++ src/physics/longwave_radiation.jl
@@ -1,241 +1,38 @@
-"""
-    function initialize_longwave_radiation!(
-        P::Parameters
-    )
+abstract type AbstractLongwave{NF} <: AbstractRadiation{NF} end
 
-Initialise variables and parameters used by the longwave radiation parametrization 
-"""
-function initialize_longwave_radiation!(K::ParameterizationConstants,
-                                        P::Parameters) 
-    radset!(K,P)
+struct NoLongwave{NF} <: AbstractLongwave{NF} end
+NoLongwave(SG::SpectralGrid) = NoLongwave{SG.NF}()
+initialize!(::NoLongwave,::PrimitiveEquation) = nothing
+
+# function barrier for all AbstractLongwave
+function longwave_radiation!(column::ColumnVariables,model::PrimitiveEquation)
+    longwave_radiation!(column,model.longwave_radiation,model)
 end
 
-"""
-    function longwave_radiation!(
-        column::ColumnVariables{NF}, model::PrimitiveEquation
-    )
+longwave_radiation!(::ColumnVariables,::NoLongwave,::PrimitiveEquation) = nothing
 
-Compute air temperature tendencies from longwave radiation for an atmospheric column.
-For more details see [http://users.ictp.it/~kucharsk/speedy_description/km_ver41_appendixA.pdf](http://users.ictp.it/~kucharsk/speedy_description/km_ver41_appendixA.pdf)
-"""
-function longwave_radiation!(   column::ColumnVariables,
-                                model::PrimitiveEquation)
-
-    radlw_down!(column, model)
-    compute_bbe!(column, model)
-    radlw_up!(column, model)
+Base.@kwdef struct UniformCooling{NF} <: AbstractLongwave{NF}
+    time_scale::Second = Hour(16)
 end
 
-"""
-    function radset!(model::PrimitiveEquation) where {NF<:AbstractFloat}
-        
-Compute energy fractions in four longwave bands as a function of temperature.
+UniformCooling(SG::SpectralGrid;kwargs...) = UniformCooling{SG.NF}(;kwargs...)
+initialize!(scheme::UniformCooling,model::PrimitiveEquation) = nothing
 
-"""
-function radset!(   K::ParameterizationConstants,
-                    P::Parameters)
-
-    (; NF, nband ) = P
-    (; epslw ) = P.radiation_coefs
-    (; fband ) = K
-
-    @assert nband == 4 "Only four bands are supported, given $nband"
-
-    eps1 = 1 - epslw
-    for jtemp = 200:320
-        fband[jtemp, 2] = (NF(0.148) - NF(3.0e-6) * (jtemp - 247)^2) * eps1
-        fband[jtemp, 3] = (NF(0.356) - NF(5.2e-6) * (jtemp - 282)^2) * eps1
-        fband[jtemp, 4] = (NF(0.314) + NF(1.0e-5) * (jtemp - 315)^2) * eps1
-        fband[jtemp, 1] = eps1 - (fband[jtemp, 2] + fband[jtemp, 3] + fband[jtemp, 4])
-    end
-    for jb = 1:4
-        for jtemp = 100:199
-            fband[jtemp, jb] = fband[200, jb]
-        end
-        for jtemp = 321:400
-            fband[jtemp, jb] = fband[320, jb]
-        end
-    end
+function longwave_radiation!(
+    column::ColumnVariables,
+    scheme::UniformCooling,
+    model::PrimitiveEquation,
+)
+    longwave_radiation!(column,scheme)
 end
 
-
-"""
-    function radlw_down!(
[…]
