+++ src/physics/define_column.jl
@@ -70,7 +70,7 @@ Base.@kwdef mutable struct ColumnVariables{NF<:AbstractFloat} <: AbstractColumnV
     const geopot::Vector{NF} = zeros(NF,nlev)                   # gepotential height [m]
 
     # CONVECTION AND PRECIPITATION
-    cloup_top::Int = nlev + 1               # layer index k of top-most layer with clouds
+    cloud_top::Int = nlev + 1               # layer index k of top-most layer with clouds
     precip_convection::NF = 0               # Precipitation due to convection [m]
     precip_large_scale::NF = 0              # precipitation due to large-scale condensation [m]
 end
