### 2024-02-15 ad04f41
reset_column without convection fields
