### 2024-02-15 421b3cd
no convection for primitive dry
