### 2024-02-15 ec4563e
uniform cooling different for "stratosphere"
