+++ src/physics/longwave_radiation.jl
@@ -13,6 +13,9 @@ longwave_radiation!(::ColumnVariables,::NoLongwave,::PrimitiveEquation) = nothin
 
 Base.@kwdef struct UniformCooling{NF} <: AbstractLongwave{NF}
     time_scale::Second = Hour(16)
+    temp_min::NF = 207.5
+    temp_stratosphere::NF = 200
+    time_scale_stratosphere::Second = Day(5)
 end
 
 UniformCooling(SG::SpectralGrid;kwargs...) = UniformCooling{SG.NF}(;kwargs...)
@@ -30,9 +33,13 @@ function longwave_radiation!(
     column::ColumnVariables{NF},
     scheme::UniformCooling,
 ) where NF
-    (;temp_tend) = column
+    (;temp, temp_tend) = column
+    (;temp_min, temp_stratosphere) = scheme
+    
     cooling = -inv(convert(NF,scheme.time_scale.value))
+    τ⁻¹ = inv(scheme.time_scale_stratosphere.value)
+
     @inbounds for k in eachlayer(column)
-        temp_tend[k] += cooling
+        temp_tend[k] += temp[k] > temp_min ? cooling : (temp_stratosphere - temp[k])*τ⁻¹
     end
 end
