### 2024-02-16 SpeedyWeather/SpeedyWeather.jl#443 @maximilian-gelbrecht
@milankl Let me know if you want me to do something
