### 2024-02-16 SpeedyWeather/SpeedyWeather.jl#443 @milankl
I can create a type tree for the documentation and outline more clearly what new types within that tree need to fulfill in order to be used properly, e.g. what would a `MyConvection <: AbstractConvection` need to have/do in order to serve as an alternative to our current `SimplifiedBettsMiller` convection?

However, @maximilian-gelbrecht what do you think our community guidelines are or should be? I think "open an issue" is the answer to all of them, but maybe we can just make that clearer in the readme or documentation?
