### 2024-02-19 SpeedyWeather/SpeedyWeather.jl#446 @vavrines
[JOSS Review] questions + suggestions
Hi @milankl
I’m working on the review for JOSS. In addition to the checklist, here are some of my questions and suggestions:
- Three sets of equations are introduced at the beginning of the Summary, where the first two are explained with figures. Maybe it makes sense to do the same for the barotropic vorticity equation?
- In the manuscript, both the physical models, e.g., the atmospheric general circulation model, and the computer programs, e.g., the Dedalus, CLIMLAB, and Oceananigans, are called “models”. The latter may be renamed to avoid ambiguity.
- The performance claims of the software appear to be limited. As the weather forecast is a pretty time-consuming task, the audience would be interested to know about the HPC capacity (distributed computing, GPU, scaling, etc.).
- The parameterization and data-driven parts of the documentation are yet to be completed. Is this intentional?
- It is mentioned that “The primitive equation core (wet or dry) is in development” in the README. How complete is this development?
