### 2024-02-19 #6323 @vavrines
I have submitted the issue in the in the software repo
https://github.com/SpeedyWeather/SpeedyWeather.jl/issues/446
