### 2024-02-19 SpeedyWeather/SpeedyWeather.jl#446 @vavrines
https://github.com/openjournals/joss-reviews/issues/6323
