### 2024-02-21 SpeedyWeather/SpeedyWeather.jl#446 @milankl
Thanks @vavrines, much appreciated, we'll be working on those and let you know when and how they're addressed!
