### 2024-02-21 ccff3f2
initial conditions with some humidity in higher layers
