+++ src/physics/large_scale_condensation.jl
@@ -1,81 +1,3 @@
-"""
-Large scale condensation as in Fortran SPEEDY with default values from therein.
-$(TYPEDFIELDS)"""
-Base.@kwdef struct SpeedyCondensation{NF<:AbstractFloat} <: AbstractCondensation{NF}
-
-    "number of vertical levels"
-    nlev::Int
-
-    "Relative humidity threshold for boundary layer"
-    threshold_boundary_layer::NF = 0.95
-
-    "Vertical range of relative humidity threshold"
-    threshold_range::NF = 0.1
-
-    "Maximum relative humidity threshold [1]"
-    threshold_max::NF = 0.9
-
-    "Relaxation time for humidity [hrs]"
-    time_scale::Second = Hour(4)
-
-    "Flux limiter for humidity tendency"
-    max_flux::NF = 0.1
-
-    "Flux limiter for condensation heating"
-    max_heating::NF = 0.001
-
-    # precomputed arrays
-    n_stratosphere_levels::Base.RefValue{Int} = Ref(0)
-    humid_tend_max::Vector{NF} = zeros(NF,nlev)
-    relative_threshold::Vector{NF} = zeros(NF,nlev)
-    latent_heat_cₚ::Base.RefValue{NF} = Ref(zero(NF))
-end
-
-SpeedyCondensation(SG::SpectralGrid;kwargs...) = SpeedyCondensation{SG.NF}(nlev=SG.nlev;kwargs...)
-
-"""
-$(TYPEDSIGNATURES)
-Initialize the SpeedyCondensation scheme."""
-function initialize!(scheme::SpeedyCondensation,model::PrimitiveEquation)
-
-    (;nlev,threshold_boundary_layer,threshold_range,threshold_max) = scheme
-    (;humid_tend_max, max_flux, relative_threshold, time_scale) = scheme
-    (;σ_levels_full) = model.geometry
-    (;σ_tropopause,σ_boundary_layer) = model.atmosphere
-   
-    n = findlast(σ->σ<=σ_tropopause,σ_levels_full)
-    scheme.n_stratosphere_levels[] = isnothing(n) ? 0 : n
-
-    for k in 1:nlev
-        # the relative humidity threshold above which condensation occurs per layer
-        σₖ² = σ_levels_full[k]^2
-        relative_threshold[k] = threshold_max + threshold_range * (σₖ² - 1)
-        if σ_levels_full[k] >= σ_boundary_layer
-            relative_threshold[k] = max(relative_threshold[k], threshold_boundary_layer)
-        end
-
-        # Impose a maximum heating rate to avoid grid-point storm instability
-        # This formula does not appear in the original SPEEDY documentation
-        # there's a (pres[end]/pres_ref)^2 which we set to 1 for simplicity
-        humid_tend_max[k] = max_flux*σₖ² / time_scale.value
-    end
-
-    # to convert humidity tendency to temperature tendency
-    # Fortran SPEEDY documentation eq. (23)
-    scheme.latent_heat_cₚ[] = model.atmosphere.latent_heat_condensation/    # [J/kg]
-                                model.atmosphere.cₚ     # [J/K/kg] specific heat capacity, const pres
-end
-
-"""
-$(TYPEDSIGNATURES)
-No condensation in a PrimitiveDry model."""
-function large_scale_condensation!( 
-    column::ColumnVariables,
-    model::PrimitiveDry,
-)
-    return nothing
-end
-
 # function barrier for all AbstractCondensation
 function large_scale_condensation!( 
     column::ColumnVariables,
@@ -85,96 +7,25 @@ function large_scale_condensation!(
     large_scale_condensation!(column,model.large_scale_condensation,model)
 end
 
[…]
