### 2024-02-21 5caff4c
primitive dry fallback for large scale condensation
