### 2024-02-22 SpeedyWeather/SpeedyWeather.jl#448 @slayoo
[JOSS review] Zenodo archival
It seems that the Zenodo release archival stopped working at v0.4.0: https://zenodo.org/doi/10.5281/zenodo.6510139
It would be very beneficial to have it up an running again.
