### 2024-02-22 SpeedyWeather/SpeedyWeather.jl#449 @slayoo
[JOSS Review] firstnames vs. initials in .bib file
In some entries, the author first names are abbreviated, while in others not. As a result, these authors are considered to be different persons by BibLaTeX. The output pdf contains malformed reference labels, e.g.: "... can be used with a single code basis (M. Klöwer et al. 2020; Milan Klöwer et al. 2022)"
