### 2024-02-22 SpeedyWeather/SpeedyWeather.jl#449 @slayoo
(in relation to https://github.com/openjournals/joss-reviews/issues/6323)
