### 2024-02-22 SpeedyWeather/SpeedyWeather.jl#443 @navidcy
> We should also add an CONTRIBUTING.md to the repo

Definitely
