### 2024-02-22 SpeedyWeather/SpeedyWeather.jl#448 @navidcy
This coincides with when the repo moved from @milankl's account to SpeedyWeather org. Probably it's related!
I'll try to sort this out.
