### 2024-02-22 SpeedyWeather/SpeedyWeather.jl#448 @navidcy
@milankl I think you need to do it because I don't have admin rights for the organization. We need to give third-party access to Zenodo in the organization; see https://docs.github.com/en/organizations/managing-oauth-access-to-your-organizations-data/approving-oauth-apps-for-your-organization
