+++ docs/joss/paper.bib
@@ -1,6 +1,6 @@
 @article{Amezcua2011,
   title = {The {{Effects}} of the {{RAW Filter}} on the {{Climatology}} and {{Forecast Skill}} of the {{SPEEDY Model}}},
-  author = {Amezcua, Javier and Kalnay, Eugenia and Williams, Paul D.},
+  author = {Amezcua, J. and Kalnay, E. and Williams, P. D.},
   year = {2011},
   month = feb,
   journal = {Monthly Weather Review},
@@ -19,7 +19,7 @@
 @article{Bezanson2017,
   title = {Julia: {{A Fresh Approach}} to {{Numerical Computing}}},
   shorttitle = {Julia},
-  author = {Bezanson, {\relax Jeff}. and Edelman, {\relax Alan}. and Karpinski, {\relax Stefan}. and Shah, Viral B.},
+  author = {Bezanson, J. and Edelman, A. and Karpinski, S. and Shah, V. B.},
   year = {2017},
   month = jan,
   journal = {SIAM Review},
@@ -35,7 +35,7 @@
 
 @article{Bourke1972,
   title = {An {{Efficient}}, {{One-Level}}, {{Primitive-Equation Spectral Model}}},
-  author = {Bourke, William},
+  author = {Bourke, W.},
   year = {1972},
   month = sep,
   journal = {Monthly Weather Review},
@@ -51,7 +51,7 @@
 @article{Burns2020,
   title = {Dedalus: {{A}} Flexible Framework for Numerical Simulations with Spectral Methods},
   shorttitle = {Dedalus},
-  author = {Burns, Keaton J. and Vasil, Geoffrey M. and Oishi, Jeffrey S. and Lecoanet, Daniel and Brown, Benjamin P.},
+  author = {Burns, K. J. and Vasil, G. M. and Oishi, J. S. and Lecoanet, D. and Brown, B. P.},
   year = {2020},
   month = apr,
   journal = {Physical Review Research},
@@ -106,7 +106,7 @@ url = {https://scitools.org.uk/cartopy}
 
 @article{Held1994,
   title = {A {{Proposal}} for the {{Intercomparison}} of the {{Dynamical Cores}} of {{Atmospheric General Circulation Models}}},
-  author = {Held, Isaac M. and Suarez, Max J.},
+  author = {Held, I. M. and Suarez, M. J.},
   year = {1994},
   month = oct,
   journal = {Bulletin of the American Meteorological Society},
@@ -139,7 +139,7 @@ url = {https://scitools.org.uk/cartopy}
 
 @article{Hotta2018,
   title = {A Nestable, Multigrid-Friendly Grid on a Sphere for Global Spectral Models Based on {{Clenshaw}}\textendash{{Curtis}} Quadrature},
-  author = {Hotta, Daisuke and Ujiie, Masashi},
+  author = {Hotta, D. and Ujiie, M.},
   year = {2018},
   journal = {Quarterly Journal of the Royal Meteorological Society},
   volume = {144},
@@ -156,7 +156,7 @@ url = {https://scitools.org.uk/cartopy}
 @article{Hunter2007,
   title = {Matplotlib: {{A 2D Graphics Environment}}},
   shorttitle = {Matplotlib},
-  author = {Hunter, John D.},
+  author = {Hunter, J. D.},
   year = {2007},
   month = may,
   journal = {Computing in Science \& Engineering},
@@ -171,7 +171,7 @@ url = {https://scitools.org.uk/cartopy}
 
 @article{Jablonowski2006,
   title = {A Baroclinic Instability Test Case for Atmospheric Model Dynamical Cores},
-  author = {Jablonowski, Christiane and Williamson, David L.},
[…]
