+++ docs/joss/paper.md
@@ -71,7 +71,7 @@ affiliations:
    index: 7
  - name: University of Melbourne, Australia
    index: 8
- - name: Australian National University, Canberra, Australia
+ - name: The University of Melbourne, Parkville, VIC, Australia
    index: 9
    
 date: 14 September 2023
