### 2024-02-22 SpeedyWeather/SpeedyWeather.jl#448 @milankl
Thanks for spotting this. Yes, I realised this a month ago or so too. It was because of the transition to the SpeedyWeather org as Navid says. I've given zenodo the respective rights to the SpeedyWeather org and this should work now, leaving this issue open now that we can check after the next release though!
