### 2024-02-22 SpeedyWeather/SpeedyWeather.jl#449 @milankl
Thanks Navid, looks like it's resolved now. Weird that sometimes bibtex does abbreviate first names automatically, sometimes not? Anyway.
