### 2024-02-22 SpeedyWeather/SpeedyWeather.jl#448 @navidcy
We might need to update the zenodo doi?
