### 2024-02-22 aa7aa92
Add AquaPlanetMask
