### 2024-02-22 SpeedyWeather/SpeedyWeather.jl#443 @navidcy
Regarding @natgeo-wong comment:

> For some reason, for the Shallow Water w/ Mountains example, the model crashed when I changed the resolution from T63 to T85. Is there a reason? Do I need to specify a smaller timestep for stability? If so, perhaps it would be useful to put this somewhere, maybe we should have a section/subsection on changing the timestep to ensure stability.

@natgeo-wong I understand you are talking about https://speedyweather.github.io/SpeedyWeather.jl/dev/setups/#Shallow-water-with-mountains

Personally I can change the resolution without any issue:

```julia
using SpeedyWeather

for resolution in [63, 85, 127]
    @info "resolution is T"*string(resolution)
    @show spectral_grid = SpectralGrid(trunc=resolution,nlev=1)
    orography = NoOrography(spectral_grid);
    initial_conditions = ZonalJet();
    @show model.time_stepping
    model = ShallowWaterModel(;spectral_grid, orography, initial_conditions);
    simulation = initialize!(model);
    run!(simulation,period=Day(6));
end
```

```Julia
[ Info: resolution is T63
spectral_grid = SpectralGrid(trunc = resolution, nlev = 1) = SpectralGrid:
├ Spectral:   T63 LowerTriangularMatrix{Complex{Float32}}, radius = 6.371e6 m
├ Grid:       96-ring OctahedralGaussianGrid{Float32}, 10944 grid points
├ Resolution: 216km (average), 192km × 207km (Equator)
└ Vertical:   1-level SigmaCoordinates
model.time_stepping = Leapfrog{Float32} <: TimeStepper
├ trunc::Int64 = 127
├ Δt_at_T31::Second = 1800 seconds
├ radius::Float32 = 6.371e6
├ adjust_with_output::Bool = true
├ robert_filter::Float32 = 0.05
├ williams_filter::Float32 = 0.53
├ Δt_millisec::Dates.Millisecond = 450000 milliseconds
├ Δt_sec::Float32 = 450.0
└ Δt::Float32 = 7.063255e-5
Weather is speedy: 100%|██████████████████████████████████████████████████████████| Time: 0:00:00 ( 3.47 millenia/day)
[ Info: resolution is T85
spectral_grid = SpectralGrid(trunc = resolution, nlev = 1) = SpectralGrid:
├ Spectral:   T85 LowerTriangularMatrix{Complex{Float32}}, radius = 6.371e6 m
├ Grid:       128-ring OctahedralGaussianGrid{Float32}, 18688 grid points
├ Resolution: 165km (average), 147km × 156km (Equator)
└ Vertical:   1-level SigmaCoordinates
model.time_stepping = Leapfrog{Float32} <: TimeStepper
├ trunc::Int64 = 63
├ Δt_at_T31::Second = 1800 seconds
├ radius::Float32 = 6.371e6
├ adjust_with_output::Bool = true
├ robert_filter::Float32 = 0.05
├ williams_filter::Float32 = 0.53
├ Δt_millisec::Dates.Millisecond = 900000 milliseconds
├ Δt_sec::Float32 = 900.0
└ Δt::Float32 = 0.0001412651
Weather is speedy: 100%|██████████████████████████████████████████████████████████| Time: 0:00:01 (1278.21 years/day)
[ Info: resolution is T127
spectral_grid = SpectralGrid(trunc = resolution, nlev = 1) = SpectralGrid:
├ Spectral:   T127 LowerTriangularMatrix{Complex{Float32}}, radius = 6.371e6 m
├ Grid:       192-ring OctahedralGaussianGrid{Float32}, 40320 grid points
├ Resolution: 112km (average), 100km × 104km (Equator)
└ Vertical:   1-level SigmaCoordinates
model.time_stepping = Leapfrog{Float32} <: TimeStepper
├ trunc::Int64 = 85
├ Δt_at_T31::Second = 1800 seconds
├ radius::Float32 = 6.371e6
├ adjust_with_output::Bool = true
├ robert_filter::Float32 = 0.05
├ williams_filter::Float32 = 0.53
├ Δt_millisec::Dates.Millisecond = 675000 milliseconds
├ Δt_sec::Float32 = 675.0
└ Δt::Float32 = 0.00010594883
Weather is speedy: 100%|██████████████████████████████████████████████████████████| Time: 0:00:04 (302.81 years/day)
```

Notice that when we construct the model, the tilmestep is different for different resolutions. I believe the CFL-type of instability you are hinting is taken care of by this automatic timestep change. Isn't it @milankl?

  <summary>My machine details</summary>

```julia
julia> versioninfo()
Julia Version 1.10.1
Commit 7790d6f064 (2024-02-13 20:41 UTC)
Build Info:

    Note: This is an unofficial build, please report bugs to the project
    responsible for this build and not to the Julia project unless you can
    reproduce the issue using official builds available at https://julialang.org/downloads

Platform Info:
  OS: macOS (arm64-apple-darwin23.3.0)
  CPU: 10 × Apple M1 Max
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-15.0.7 (ORCJIT, apple-m1)
Threads: 8 default, 0 interactive, 4 GC (on 8 virtual cores)
Environment:
  JULIA_EDITOR = code

julia> using Pkg; Pkg.status()
Project SpeedyWeather v0.8.0
Status `~/Research/SpeedyWeather.jl/Project.toml`
  [621f4979] AbstractFFTs v1.5.0
  [79e6a3ab] Adapt v4.0.1
  [2119f1ac] AssociatedLegendrePolynomials v1.0.1
  [de688a37] BitInformation v0.6.1
  [052768ef] CUDA v5.2.0
⌃ [944b1d66] CodecZlib v0.7.3
  [ffbed154] DocStringExtensions v0.9.3
⌃ [7a1cc6ca] FFTW v1.7.2
  [cc61a311] FLoops v0.2.1
  [442a2c76] FastGaussQuadrature v1.0.2
  [a8297547] GenericFFT v0.1.6
⌃ [033835bb] JLD2 v0.4.44
  [63c18a36] KernelAbstractions v0.9.16
⌃ [85f8d34a] NCDatasets v0.14.0
  [27ebfcd6] Primes v0.5.5
  [92933f4c] ProgressMeter v1.9.0
  [b8865327] UnicodePlots v3.6.3
  [ade2ca70] Dates
  [37e2e46d] LinearAlgebra
  [de0858da] Printf
  [9a3f8284] Random
  [10745b16] Statistics v1.10.0
  [fa267f1f] TOML v1.0.3
Info Packages marked with ⌃ have new versions available and may be upgradable.
```
