### 2024-02-22 a279caf
Implement AquaPlanetMask
