### 2024-02-22 f17f863
same NCDatasets compat as main project
