+++ docs/src/output.md
@@ -154,74 +154,6 @@ Note that some fields are actual options, but others are derived from the option
 arrays/objects the output writer needs, but shouldn't be passed on by the user.
 The actual options are declared as `[OPTION]` in the following
 
-```julia
-help?> OutputWriter
-search: OutputWriter
-
-  NetCDF output writer. Contains all output options and auxiliary fields for
-  output interpolation. To be initialised with
-  OutputWriter(::SpectralGrid,::Type{<:ModelSetup},kwargs...) to pass on the
-  resolution information and the model type which chooses which variables to
-  output. Options include
-
-    •  spectral_grid::SpectralGrid
-
-    •  output::Bool
-
-    •  path::String: [OPTION] path to output folder, run_???? will be
-       created within
-
-    •  id::String: [OPTION] run identification number/string
-
-    •  run_path::String
-
-    •  filename::String: [OPTION] name of the output netcdf file
-
-    •  write_restart::Bool: [OPTION] also write restart file if
-       output==true?
-
-    •  pkg_version::VersionNumber
-
-    •  startdate::DateTime
-
-    •  output_dt::Second: [OPTION] output frequency, time step
-
-    •  output_vars::Vector{Symbol}: [OPTION] which variables to output,
-       u, v, vor, div, pres, temp, humid
-
-    •  missing_value::Union{Float32, Float64}: [OPTION] missing value to
-       be used in netcdf output
-
-    •  compression_level::Int64: [OPTION] lossless compression level;
-       1=low but fast, 9=high but slow
-
-    •  shuffle::Bool: [OPTION] shuffle/bittranspose filter for
-       compression
-
-    •  keepbits::SpeedyWeather.Keepbits: [OPTION] mantissa bits to keep
-       for every variable
-
-    •  output_every_n_steps::Int64
-
-    •  timestep_counter::Int64
-
-    •  output_counter::Int64
-
-    •  netcdf_file::Union{Nothing, NCDatasets.NCDataset}
-
-    •  input_Grid::Type{<:AbstractGrid}
-
-    •  as_matrix::Bool: [OPTION] sort grid points into a matrix
-       (interpolation-free), for OctahedralClenshawGrid, OctaHEALPixGrid
-       only
-
-    •  quadrant_rotation::NTuple{4, Int64}
-
-    •  matrix_quadrant::NTuple{4, Tuple{Int64, Int64}}
-
-    •  output_Grid::Type{<:AbstractFullGrid}: [OPTION] the grid used for
-       output, full grids only
-
-    •  nlat_half::Int64: [OPTION] the resolution of the output grid,
-       default: same nlat_half as in the dynamical core
+```@docs
+SpeedyWeather.OutputWriter
 ```
