### 2024-02-22 a9d27f8
docstring explanation for AbstractLandSeaMask
