### 2024-02-22 8da1bb0
explain also generator function for land sea mask
