+++ src/physics/land_sea_mask.jl
@@ -4,13 +4,24 @@ Abstract type for land-sea masks. Custom land-sea masks have to be defined as
     CustomMask{NF<:AbstractFloat,Grid<:AbstractGrid{NF}} <: AbstractLandSeaMask{NF,Grid}
 
 and need to have at least a field called `land_sea_mask::Grid` that uses a `Grid` as defined
-by the spectral grid object, so of correct size and with the numberformat NF.
+by the spectral grid object, so of correct size and with the number format NF.
+It is therefore recommended (but not required) to write a generator function as follows
+
+    function CustomMask(spectral_grid::SpectralGrid;kwargs...)
+        (;NF, Grid, nlat_half) = spectral_grid
+        land_sea_mask   = zeros(Grid{NF},nlat_half)
+        return LandSeaMask{NF,Grid{NF}}(;land_sea_mask,kwargs...)
+    end
+
+to allow a convenient construction like `mask = CustomMask(spectral_grid,option=argument)`.
 Then the initialize function has to be extended for that new mask
 
     initialize!(mask::CustomMask,model::PrimitiveEquation)
 
 which generally is used to tweak the mask.land_sea_mask grid as you like, using
-(preferrably read-only) anything from `model`.
+any other options you have included in `CustomMask` as fields
+or anything else (preferrably read-only, because this is only to initialize the land-sea mask, nothing else)
+from `model`.
 
 The land-sea mask grid is expected to have values between [0,1] as we use a fractional mask,
 allowing for grid points being, e.g. quarter land and three quarters sea for 0.25
@@ -19,6 +30,12 @@ from sea and land surface temperatures. Note however, that the land-sea mask can
 grid points being (at least partially) ocean even though the sea surface temperatures
 aren't defined (=NaN) in that grid point. In that case, not flux is applied."""
 abstract type AbstractLandSeaMask{NF,Grid} end
+
+function Base.show(io::IO,L::AbstractLandSeaMask)
+    println(io,"$(typeof(L)) <: AbstractLandSeaMask")
+    keys = propertynames(L)
+    print_fields(io,L,keys)
+end
 
 # make available when using SpeedyWeather
 export LandSeaMask, AquaPlanetMask
