### 2024-02-22 f3f5b80
general AbstractLandSeaMask generator via functor
