### 2024-02-22 SpeedyWeather/SpeedyWeather.jl#446 @navidcy
> * Three sets of equations are introduced at the beginning of the Summary, where the first two are explained with figures. Maybe it makes sense to do the same for the barotropic vorticity equation?

@milankl, perhaps a snapshot from the video in the README:

https://github.com/SpeedyWeather/SpeedyWeather.jl/assets/25530332/3d7fccd5-b66d-42e3-9f73-64dcf21d00ee
