### 2024-02-22 7d6285c
NoCallback and GlobalSurfaceTemperatureCallback
