+++ src/abstract_types.jl
@@ -41,6 +41,7 @@ abstract type AbstractEvaporation{NF} <: AbstractParameterization{NF} end
 # INPUT/OUTPUT
 abstract type AbstractFeedback end
 abstract type AbstractOutputWriter end
+abstract type AbstractCallback end
 
 # NUMERICS
 abstract type HorizontalDiffusion{NF} end
