### 2024-02-22 2df0793
NoCallback and GlobalSurfaceTemperatureCallback
