### 2024-02-22 97a370d
export also GlobalSurfaceTemperatureCallback
