### 2024-02-23 #6323 @kthyng
Yes! This is ok. Ultimately when this package is uploaded to e.g. zenodo your username will be included in the author metadata but we ask that the metadata be overwritten to match the JOSS title and author list anyway. Thanks for your effort and enthusiasm!!
