### 2024-02-23 e6d9470
allocate callbacks with empty vector
