### 2024-02-23 7fb2763
add append!(::AbstractVector{<:AbstractCallback},::AbstractCallback) method
