+++ docs/make.jl
@@ -9,19 +9,39 @@ makedocs(
     modules = [SpeedyWeather],
     pages = ["Home"=>"index.md",
             "Installation"=>"installation.md",
-            "How to run SpeedyWeather.jl"=>"how_to_run_speedy.md",
-            "Model setups"=>"setups.md",
-            "Spherical Harmonic Transform"=>"spectral_transform.md",
-            "Grids"=>"grids.md",
-            "Barotropic model"=>"barotropic.md",
-            "Shallow water model"=>"shallowwater.md",
-            "Primitive equation model"=>"primitiveequation.md",
-            "Parameterizations"=>"parameterizations.md",
-            "Extending SpeedyWeather"=>"extending.md",
-            "NetCDF output"=>"output.md",
-            "Submodule: RingGrids"=>"ringgrids.md",
-            "Submodule: LowerTriangularMatrices"=>"lowertriangularmatrices.md",
-            "Submodule: SpeedyTransforms"=>"speedytransforms.md",
+            "Dynamics" => [
+                "Barotropic model"=>"barotropic.md",
+                "Shallow water model"=>"shallowwater.md",
+                "Primitive equation model"=>"primitiveequation.md",
+            ],
+            "Physics" => [
+                "Large-scale condensation"=>"large_scale_condensation.md",
+                "Convection"=>"convection.md",
+                "Radiation"=>"radiation.md",
+                "Vertical diffusion"=>"vertical_diffusion.md",
+                "Surface fluxes"=>"surface_fluxes.md",
+            ],
+            "Discretization" => [
+                "Spherical Harmonic Transform"=>"spectral_transform.md",
+                "Grids"=>"grids.md",
+            ],
+            "Running SpeedyWeather.jl" => [
+                "How to run SpeedyWeather.jl"=>"how_to_run_speedy.md",
+                "Model setups"=>"setups.md",
+                "NetCDF output"=>"output.md",
+            ],
+            "Extending SpeedyWeather" => [
+                "Forcing and drag"=>"forcing_drag.md",
+                "Parameterizations"=>"parameterizations.md",
+                "Orography"=>"orography.md",
+                "Land-Sea Mask"=>"land_sea_mask.md",
+                "Callbacks"=>"callbacks.md",
+            ],
+            "Submodules" => [
+                "RingGrids"=>"ringgrids.md",
+                "LowerTriangularMatrices"=>"lowertriangularmatrices.md",
+                "SpeedyTransforms"=>"speedytransforms.md",
+            ],
             "Function and type index"=>"functions.md"]
 )
 
