### 2024-02-23 SpeedyWeather/SpeedyWeather.jl#443 @milankl
Yes, the timestep is
1. automatically chosen given the resolution. At T31 it currently uses 30min which is the default set in `model.time_stepping.Δt_at_T31`. It then scales ~linearly to your resolution
2. Given the output frequency, the timestep might be slightly adjusted to have e.g. output exactly every hour without interpolation. This rarely influences the time step by more than +- a few percent. `model.time_stepping.adjust_with_output = false` can disable that, also a warning is thrown if that's more than +5% or less than -5%

Overall the time step is part of the numerics that SpeedyWeather should hide from the general user. A general user shouldn't worry about time stepping, scaling, transforms, etc. Sometimes that's in
