+++ src/physics/abstract_types.jl
@@ -1,5 +1,7 @@
+abstract type AbstractParameterization <: AbstractModelComponent end
+
 # print all fields with type <: Number
-function Base.show(io::IO,P::AbstractParameterization)
+function Base.show(io::IO,P::AbstractModelComponent)
     println(io,"$(typeof(P)) <: $(supertype(typeof(P)))")
     keys = propertynames(P)
     print_fields(io,P,keys)

[…]
