### 2024-02-26 7e128d2
zero tendencies at beginning of time step
