### 2024-02-26 94d537a
zero tendencies at beginning of time step
