### 2024-02-27 7043864
add tests for callbacks
