+++ test/runtests.jl
@@ -40,6 +40,7 @@ include("run_speedy.jl")
 
 # EXTENSION
 include("extending.jl")
+include("callbacks.jl")
 
 # OUTPUT 
 include("netcdf_output.jl")
