### 2024-02-27 08a5a64
strip away {NF,D} parameters from models
