+++ src/dynamics/models.jl
@@ -19,30 +19,30 @@ $(SIGNATURES)
 The BarotropicModel struct holds all other structs that contain precalculated constants,
 whether scalars or arrays that do not change throughout model integration.
 $(TYPEDFIELDS)"""
-Base.@kwdef mutable struct BarotropicModel{NF<:AbstractFloat, D<:AbstractDevice} <: Barotropic
+Base.@kwdef mutable struct BarotropicModel <: Barotropic
     spectral_grid::SpectralGrid = SpectralGrid(nlev=1)
 
     # DYNAMICS
     planet::AbstractPlanet = Earth()
     atmosphere::AbstractAtmosphere = EarthAtmosphere()
-    forcing::AbstractForcing{NF} = NoForcing(spectral_grid)
-    drag::AbstractDrag{NF} = NoDrag(spectral_grid)
+    forcing::AbstractForcing = NoForcing(spectral_grid)
+    drag::AbstractDrag = NoDrag(spectral_grid)
     initial_conditions::InitialConditions = StartWithRandomVorticity()
 
     # NUMERICS
-    time_stepping::TimeStepper{NF} = Leapfrog(spectral_grid)
-    spectral_transform::SpectralTransform{NF} = SpectralTransform(spectral_grid)
-    horizontal_diffusion::HorizontalDiffusion{NF} = HyperDiffusion(spectral_grid)
-    implicit::AbstractImplicit{NF} = NoImplicit(spectral_grid)
+    time_stepping::TimeStepper = Leapfrog(spectral_grid)
+    spectral_transform::SpectralTransform = SpectralTransform(spectral_grid)
+    horizontal_diffusion::HorizontalDiffusion = HyperDiffusion(spectral_grid)
+    implicit::AbstractImplicit = NoImplicit(spectral_grid)
 
     # INTERNALS
-    geometry::Geometry{NF} = Geometry(spectral_grid)
-    constants::DynamicsConstants{NF} = DynamicsConstants(spectral_grid,planet,atmosphere,geometry)
-    device_setup::DeviceSetup{D} = DeviceSetup(CPUDevice())
+    geometry::Geometry = Geometry(spectral_grid)
+    constants::DynamicsConstants = DynamicsConstants(spectral_grid,planet,atmosphere,geometry)
+    device_setup::DeviceSetup = DeviceSetup(CPUDevice())
 
     # OUTPUT
     output::AbstractOutputWriter = OutputWriter(spectral_grid,Barotropic)
-    callbacks::AbstractVector{<:AbstractCallback} = Vector{AbstractCallback}()
+    callbacks::Vector{AbstractCallback} = AbstractCallback[]
     feedback::AbstractFeedback = Feedback()
 end
 
@@ -82,31 +82,31 @@ $(SIGNATURES)
 The ShallowWaterModel struct holds all other structs that contain precalculated constants,
 whether scalars or arrays that do not change throughout model integration.
 $(TYPEDFIELDS)"""
-Base.@kwdef mutable struct ShallowWaterModel{NF<:AbstractFloat, D<:AbstractDevice} <: ShallowWater
+Base.@kwdef mutable struct ShallowWaterModel <: ShallowWater
     spectral_grid::SpectralGrid = SpectralGrid(nlev=1)
 
     # DYNAMICS
     planet::AbstractPlanet = Earth()
     atmosphere::AbstractAtmosphere = EarthAtmosphere()
-    forcing::AbstractForcing{NF} = NoForcing(spectral_grid)
-    drag::AbstractDrag{NF} = NoDrag(spectral_grid)
+    forcing::AbstractForcing = NoForcing(spectral_grid)
[…]
