### 2024-02-29 109a681
parametric models: barotropic working
