+++ src/output/feedback.jl
@@ -1,4 +1,3 @@
-abstract type AbstractFeedback end
 export Feedback
 
 """
@@ -140,7 +139,7 @@ end
 """
 $(TYPEDSIGNATURES)
 Finalises the progress meter and the progress txt file."""
-function progress_finish!(F::Feedback)
+function finish!(F::Feedback)
     ProgressMeter.finish!(F.progress_meter)
     
     if F.output     # write final progress to txt file
