+++ src/physics/longwave_radiation.jl
@@ -1,7 +1,7 @@
-abstract type AbstractLongwave{NF} <: AbstractRadiation{NF} end
+abstract type AbstractLongwave <: AbstractRadiation end
 
-struct NoLongwave{NF} <: AbstractLongwave{NF} end
-NoLongwave(SG::SpectralGrid) = NoLongwave{SG.NF}()
+struct NoLongwave <: AbstractLongwave end
+NoLongwave(SG::SpectralGrid) = NoLongwave()
 initialize!(::NoLongwave,::PrimitiveEquation) = nothing
 
 # function barrier for all AbstractLongwave
@@ -11,7 +11,7 @@ end
 
 longwave_radiation!(::ColumnVariables,::NoLongwave,::PrimitiveEquation) = nothing
 
-Base.@kwdef struct UniformCooling{NF} <: AbstractLongwave{NF}
+Base.@kwdef struct UniformCooling{NF} <: AbstractLongwave
     time_scale::Second = Hour(16)
     temp_min::NF = 207.5
     temp_stratosphere::NF = 200
