### 2024-02-29 SpeedyWeather/SpeedyWeather.jl#446 @vavrines
Thanks for the answer.

> some gallery of the repository

This is just a personal advice from the fact that it will be helpful for someone like me who is not in the climate community to understand better. There's no sticking to it. You may choose whatever form you think is appropriate.

> what about I create a simple overview table in the readme of typical speeds

That also sounds good.

Now everything is clear to me.

Lastly, thank you for creating this work. It should be very helpful for both the scientific research and Julia community.
