### 2024-02-29 896291d
make some model components mutable
