+++ src/dynamics/horizontal_diffusion.jl
@@ -11,7 +11,7 @@ layers. Furthermore the power can be decreased above the `tapering_σ` to
 `power_stratosphere` (default 2). For Barotropic, ShallowWater,
 the default non-adaptive constant-time scale hyper diffusion is used. Options are
 $(TYPEDFIELDS)"""
-Base.@kwdef struct HyperDiffusion{NF} <: AbstractHorizontalDiffusion
+Base.@kwdef mutable struct HyperDiffusion{NF} <: AbstractHorizontalDiffusion
     # DIMENSIONS
     "spectral resolution"
     trunc::Int
