### 2024-02-29 a7901aa
parametric ShallowWaterModel
