+++ src/models/shallow_water.jl
@@ -1,36 +1,57 @@
 export ShallowWaterModel
 
 """
-$(SIGNATURES)
-The ShallowWaterModel struct holds all other structs that contain precalculated constants,
-whether scalars or arrays that do not change throughout model integration.
+The ShallowWaterModel contains all model components needed for the simulation of the
+shallow water equations. To be constructed like
+
+    model = ShallowWaterModel(;spectral_grid, kwargs...)
+
+with `spectral_grid::SpectralGrid` used to initalize all non-default components
+passed on as keyword arguments, e.g. `planet=Earth(spectral_grid)`. Fields, representing
+model components, are
 $(TYPEDFIELDS)"""
-Base.@kwdef mutable struct ShallowWaterModel{NF<:AbstractFloat, D<:AbstractDevice} <: ShallowWater
+Base.@kwdef mutable struct ShallowWaterModel{
+    NF<:AbstractFloat,
+    DS<:DeviceSetup,
+    PL<:AbstractPlanet,
+    AT<:AbstractAtmosphere,
+    CO<:AbstractCoriolis,
+    OR<:AbstractOrography,
+    FR<:AbstractForcing,
+    DR<:AbstractDrag,
+    IC<:InitialConditions,
+    TS<:AbstractTimeStepper,
+    ST<:SpectralTransform{NF},
+    IM<:AbstractImplicit,
+    HD<:AbstractHorizontalDiffusion,
+    GE<:AbstractGeometry,
+    OW<:AbstractOutputWriter,
+    FB<:AbstractFeedback,
+} <: ShallowWater
+    
     spectral_grid::SpectralGrid = SpectralGrid(nlev=1)
+    device_setup::DS = DeviceSetup(CPUDevice())
 
     # DYNAMICS
-    planet::AbstractPlanet = Earth()
-    atmosphere::AbstractAtmosphere = EarthAtmosphere()
-    coriolis::AbstractCoriolis = Coriolis(spectral_grid)
-    forcing::AbstractForcing = NoForcing()
-    drag::AbstractDrag = NoDrag()
-    initial_conditions::InitialConditions = ZonalJet()
-    orography::AbstractOrography{NF} = EarthOrography(spectral_grid)
+    planet::PL = Earth(spectral_grid)
+    atmosphere::AT = EarthAtmosphere(spectral_grid)
+    coriolis::CO = Coriolis(spectral_grid)
+    orography::OR = EarthOrography(spectral_grid)
+    forcing::FR = NoForcing()
+    drag::DR = NoDrag()
+    initial_conditions::IC = ZonalJet()
 
     # NUMERICS
-    time_stepping::TimeStepper = Leapfrog(spectral_grid)
-    spectral_transform::SpectralTransform{NF} = SpectralTransform(spectral_grid)
-    horizontal_diffusion::HorizontalDiffusion = HyperDiffusion(spectral_grid)
-    implicit::AbstractImplicit = ImplicitShallowWater(spectral_grid)
-
-    # INTERNALS
-    geometry::AbstractGeometry = Geometry(spectral_grid)
-    constants::DynamicsConstants{NF} = DynamicsConstants(spectral_grid,planet,atmosphere,geometry)
-    device_setup::DeviceSetup{D} = DeviceSetup(CPUDevice())
+    time_stepping::TS = Leapfrog(spectral_grid)
+    spectral_transform::ST = SpectralTransform(spectral_grid)
+    implicit::IM = ImplicitShallowWater(spectral_grid)
+    horizontal_diffusion::HD = HyperDiffusion(spectral_grid)
+    geometry::GE = Geometry(spectral_grid)
 
     # OUTPUT
[…]
