+++ src/physics/abstract_types.jl
@@ -1,8 +1,1 @@
 abstract type AbstractParameterization <: AbstractModelComponent end
-
-# print all fields with type <: Number
-function Base.show(io::IO,P::AbstractModelComponent)
-    println(io,"$(typeof(P)) <: $(supertype(typeof(P)))")
-    keys = propertynames(P)
-    print_fields(io,P,keys)
-end
