### 2024-02-29 fbaca7f
fix docstring rendering
