### 2024-03-01 b2d6251
primitive dry model parametric
