+++ src/SpeedyWeather.jl
@@ -91,6 +91,7 @@ include("dynamics/time_integration.jl")
 include("dynamics/forcing.jl")
 include("dynamics/drag.jl")
 include("dynamics/geopotential.jl")
+include("dynamics/virtual_temperature.jl")
 include("dynamics/initial_conditions.jl")
 include("dynamics/horizontal_diffusion.jl")
 include("dynamics/vertical_advection.jl")
@@ -128,6 +129,6 @@ include("output/callbacks.jl")
 include("models/simulation.jl")
 include("models/barotropic.jl")
 include("models/shallow_water.jl")
-# include("models/primitive_dry.jl")
+include("models/primitive_dry.jl")
 # include("models/primitive_wet.jl")
 end
