+++ src/models/barotropic.jl
@@ -28,9 +28,10 @@ Base.@kwdef mutable struct BarotropicModel{
     FB<:AbstractFeedback,
 } <: Barotropic
     
+    # GRID
     spectral_grid::SpectralGrid = SpectralGrid(nlev=1)
-    device_setup::DS = DeviceSetup(CPUDevice())
-
+    geometry::GE = Geometry(spectral_grid)
+    
     # DYNAMICS
     planet::PL = Earth(spectral_grid)
     atmosphere::AT = EarthAtmosphere(spectral_grid)
@@ -38,13 +39,13 @@ Base.@kwdef mutable struct BarotropicModel{
     forcing::FR = NoForcing()
     drag::DR = NoDrag()
     initial_conditions::IC = StartWithRandomVorticity()
-
+    
     # NUMERICS
+    device_setup::DS = DeviceSetup(CPUDevice())
     time_stepping::TS = Leapfrog(spectral_grid)
     spectral_transform::ST = SpectralTransform(spectral_grid)
     implicit::IM = NoImplicit(spectral_grid)
     horizontal_diffusion::HD = HyperDiffusion(spectral_grid)
-    geometry::GE = Geometry(spectral_grid)
 
     # OUTPUT
     output::OW = OutputWriter(spectral_grid,Barotropic)
