### 2024-03-01 63e3d53
PrimitiveDryModel parametric (+prev unsaved changes)
