### 2024-03-01 d023a03
PrimitiveWet also parametric and so 20% faster!!
