+++ src/models/primitive_wet.jl
@@ -5,59 +5,96 @@ $(SIGNATURES)
 The PrimitiveDryModel struct holds all other structs that contain precalculated constants,
 whether scalars or arrays that do not change throughout model integration.
 $(TYPEDFIELDS)"""
-Base.@kwdef mutable struct PrimitiveWetModel{NF<:AbstractFloat, D<:AbstractDevice} <: PrimitiveWet
-    spectral_grid::SpectralGrid = SpectralGrid()
+Base.@kwdef mutable struct PrimitiveWetModel{
+    NF<:AbstractFloat,
+    DS<:DeviceSetup,
+    PL<:AbstractPlanet,
+    AT<:AbstractAtmosphere,
+    CO<:AbstractCoriolis,
+    GO<:AbstractGeopotential,
+    OR<:AbstractOrography,
+    AC<:AbstractAdiabaticConversion,
+    IC<:InitialConditions,
+    LS<:AbstractLandSeaMask,
+    OC<:AbstractOcean,
+    LA<:AbstractLand,
+    ZE<:AbstractZenith,
+    SO<:AbstractSoil,
+    VG<:AbstractVegetation,
+    CC<:AbstractClausiusClapeyron,
+    BL<:AbstractBoundaryLayer,
+    TR<:AbstractTemperatureRelaxation,
+    SE<:AbstractVerticalDiffusion,
+    HU<:AbstractVerticalDiffusion,
+    SUT<:AbstractSurfaceThermodynamics,
+    SUW<:AbstractSurfaceWind,
+    SH<:AbstractSurfaceHeat,
+    EV<:AbstractEvaporation,
+    LSC<:AbstractCondensation,
+    CV<:AbstractConvection,
+    SW<:AbstractShortwave,
+    LW<:AbstractLongwave,
+    TS<:AbstractTimeStepper,
+    ST<:SpectralTransform{NF},
+    IM<:AbstractImplicit,
+    HD<:AbstractHorizontalDiffusion,
+    VA<:AbstractVerticalAdvection,
+    HF<:AbstractHoleFilling,
+    GE<:AbstractGeometry,
+    OW<:AbstractOutputWriter,
+    FB<:AbstractFeedback,
+} <: PrimitiveWet
 
+    spectral_grid::SpectralGrid = SpectralGrid()
+    geometry::GE = Geometry(spectral_grid)
+    
     # DYNAMICS
     dynamics::Bool = true
-    planet::AbstractPlanet = Earth()
-    atmosphere::AbstractAtmosphere = EarthAtmosphere()
-    coriolis::AbstractCoriolis = Coriolis(spectral_grid)
-    initial_conditions::InitialConditions = ZonalWind()
-    orography::AbstractOrography{NF} = EarthOrography(spectral_grid)
-    geopotential::AbstractGeopotential = Geopotential(spectral_grid)
-    adiabatic_conversion::AbstractAdiabaticConversion = AdiabaticConversion(spectral_grid)
-
+    planet::PL = Earth(spectral_grid)
+    atmosphere::AT = EarthAtmosphere(spectral_grid)
+    coriolis::CO = Coriolis(spectral_grid)
+    geopotential::GO = Geopotential(spectral_grid)
+    adiabatic_conversion::AC = AdiabaticConversion(spectral_grid)
+    initial_conditions::IC = ZonalWind()
+    
     # BOUNDARY CONDITIONS
-    land_sea_mask::AbstractLandSeaMask{NF} = LandSeaMask(spectral_grid)
-    ocean::AbstractOcean{NF} = SeasonalOceanClimatology(spectral_grid)
-    land::AbstractLand{NF} = SeasonalLandTemperature(spectral_grid)
[…]
