+++ src/SpeedyWeather.jl
@@ -75,8 +75,8 @@ include("dynamics/vertical_coordinates.jl")
 include("dynamics/spectral_grid.jl")
 include("dynamics/geometry.jl")
 include("dynamics/coriolis.jl")
-include("dynamics/planets.jl")
-include("dynamics/atmospheres.jl")
+include("dynamics/planet.jl")
+include("dynamics/atmosphere.jl")
 include("dynamics/adiabatic_conversion.jl")
 include("dynamics/orography.jl")
 include("physics/land_sea_mask.jl")
@@ -130,5 +130,5 @@ include("models/simulation.jl")
 include("models/barotropic.jl")
 include("models/shallow_water.jl")
 include("models/primitive_dry.jl")
-# include("models/primitive_wet.jl")
+include("models/primitive_wet.jl")
 end
