+++ docs/src/forcing_drag.md
@@ -37,7 +37,7 @@ To define a new forcing type, at the most basic level you would do
 ```@example extending
 using SpeedyWeather
 
-struct MyForcing{NF} <: AbstractForcing{NF}
+struct MyForcing{NF} <: SpeedyWeather.AbstractForcing
     # define some parameters and work arrays here
     a::NF
     v::Vector{NF}
@@ -47,7 +47,7 @@ In Julia this introduces a new (so-called compound) type that is a subtype of `A
 we have a bunch of these abstract super types defined and you want to
 piggy-back on them because of multiple-dispatch. This new type could also be 
 a `mutable struct`, could have keywords defined with `Base.@kwdef` and can
-also be parametric with respect to the number format or grid, but let's skip
+also be parametric with respect to the number format `NF` or grid, but let's skip
 those details for now. Conceptually you include into the type any parameters
 (example the float `a` here) that you may need and especially those that you want to change.
 This type will get a user-facing interface so that one can quickly create a new
@@ -163,7 +163,7 @@ will explain the details in second
 ```@example extend
 using SpeedyWeather, Dates
 
-Base.@kwdef struct StochasticStirring{NF} <: AbstractForcing{NF}
+Base.@kwdef struct StochasticStirring{NF} <: SpeedyWeather.AbstractForcing
 
     # DIMENSIONS from SpectralGrid
     "Spectral resolution as max degree of spherical harmonics"
@@ -175,16 +175,16 @@ Base.@kwdef struct StochasticStirring{NF} <: AbstractForcing{NF}
 
     # OPTIONS
     "Decorrelation time scale τ [days]"
-    decorrelation_time::Float64 = 2
+    decorrelation_time::Second = Day(2)
 
     "Stirring strength A [1/s²]"
-    strength::Float64 = 1e-11
+    strength::NF = 7e-11
 
     "Stirring latitude [˚N]"
-    latitude::Float64 = 45
+    latitude::NF = 45
 
     "Stirring width [˚]"
-    width::Float64 = 24
+    width::NF = 24
 
 
     # TO BE INITIALISED
@@ -202,9 +202,10 @@ Base.@kwdef struct StochasticStirring{NF} <: AbstractForcing{NF}
 end
 ```
 
-So, first the scalar parameters, are added as fields of type `Float64` with some default values
-as suggested in the [Vallis et al., 2004](https://doi.org/10.1175/1520-0469(2004)061%3C0264:AMASDM%3E2.0.CO;2)
-paper. In order to be able to define the default values, we add the `Base.@kwdef` macro
+So, first the scalar parameters, are added as fields of type `NF` (you could harcode `Float64` too)
+with some default values as suggested in the
+[Vallis et al., 2004](https://doi.org/10.1175/1520-0469(2004)061%3C0264:AMASDM%3E2.0.CO;2) paper.
+In order to be able to define the default values, we add the `Base.@kwdef` macro
 before the `struct` definition. Then we need the term `S` as coefficients of the spherical harmonics, which
 is a `LowerTriangularMatrix`, however we want its elements to be of number format `NF`,
 which is also the parametric type of `StochasticStirring{NF}`, this is done because it will
@@ -236,33 +237,29 @@ zero. For this we want to define a latitudinal mask `lat_mask` that is a vector 
 the number of latitude rings. Similar to `S`, we want to allocate it with zeros (or any other
 value for that matter), but then precompute this mask in the `initialize!` step. For this
 we need to know `nlat` at creation time meaning we add this field similar as to how we added
-`trunc`. This mask requires the parameters `latitude` and a `width` which are therefore also
-added to the definition of `StochasticStirring`.
+`trunc`. This mask requires the parameters `latitude` (it's position) and a `width` which
[…]
