### 2024-03-02 9ed6be2
Callbacks implemented as dictionary
