+++ src/models/primitive_dry.jl
@@ -77,7 +77,7 @@ Base.@kwdef mutable struct PrimitiveDryModel{
     
     # OUTPUT
     output::OW = OutputWriter(spectral_grid, PrimitiveDry)
-    callbacks::Vector{AbstractCallback} = AbstractCallback[]
+    callbacks::Dict{Symbol,AbstractCallback} = Dict{Symbol,AbstractCallback}()
     feedback::FB = Feedback()
 end
 
