+++ src/SpeedyWeather.jl
@@ -20,6 +20,7 @@ import Adapt: Adapt, adapt, adapt_structure
 import TOML
 import Dates: Dates, DateTime, Period, Millisecond, Second, Minute, Hour, Day
 import Printf: Printf, @sprintf
+import Random: randstring
 import NCDatasets: NCDatasets, NCDataset, defDim, defVar
 import JLD2: jldopen
 import CodecZlib
