+++ test/callbacks.jl
@@ -7,12 +7,16 @@
     @test CallbackDict(:a => NoCallback()) isa SpeedyWeather.CALLBACK_DICT
 
     d = CallbackDict()
-    add!(d,NoCallback())
-    add!(d,:my_callback,NoCallback())
-    add!(d,"my_callback",NoCallback())
+    add!(d, NoCallback())
+    add!(d, NoCallback(), NoCallback())
+    add!(d, :my_callback, NoCallback())
+    add!(d, "my_callback2", NoCallback())
+    delete!(d, :my_callback)
+
+    add!(d,:another_callback => NoCallback())
+    add!(d,:another_callback1 => NoCallback(), :another_callback2 => NoCallback())
     @test d isa SpeedyWeather.CALLBACK_DICT
-    delete!(d,:my_callback)
-    @test d isa SpeedyWeather.CALLBACK_DICT
+    @test length(d) == 7
 end
 
 @testset "Callbacks interface" begin
@@ -79,11 +83,11 @@ end
     
     storm_chaser = StormChaser(spectral_grid)
     key = :storm_chaser
-    add!(model.callbacks, key, storm_chaser)  # with :storm_chaser key
-    add!(model.callbacks, NoCallback())                 # add dummy too 
+    add!(model.callbacks, key => storm_chaser)      # with :storm_chaser key
+    add!(model.callbacks, NoCallback())             # add dummy too 
 
     simulation = initialize!(model)
-    run!(simulation,period=Day(3))
+    run!(simulation,period=Day(1))
 
     # maximum wind speed should always be non-negative
     @test all(model.callbacks[key].maximum_surface_wind_speed .>= 0)
