### 2024-03-02 adc6e08
no need for SpeedyWeather. in extending.jl tests
