### 2024-03-02 310fa3b
Remove also unnecessary SpeedyWeather. in forcing_drag.md
