+++ src/SpeedyWeather.jl
@@ -83,6 +83,7 @@ include("dynamics/orography.jl")
 include("physics/land_sea_mask.jl")
 
 # VARIABLES
+include("dynamics/particles.jl")
 include("dynamics/prognostic_variables.jl")
 include("physics/define_column.jl")
 include("dynamics/diagnostic_variables.jl")
