+++ src/dynamics/particles.jl
@@ -1,35 +1,91 @@
 abstract type AbstractParticle{NF} end
 
 export Particle
-Base.@kwdef struct Particle{NF<:AbstractFloat} <: AbstractParticle{NF}
-    lat::NF = 0         # [-90˚,90˚]
-    lon::NF = 0         # [0,360˚]
-    σ::NF   = 0         # [0,1]
-    active::Bool = true # deactivated particles don't move
+struct Particle{
+    NF<:AbstractFloat,  # number format of coordinates
+    isactive,           # bool indicating whether particle is active
+} <: AbstractParticle{NF}
+    lon::NF             # longitude [0,360˚]
+    lat::NF             # latitude [-90˚,90˚]
+    σ::NF               # vertical sigma coordinate [0,1] (top to surface)
 end
 
-function Base.show(io::IO,p::Particle)
+# particle is by default active, of number format DEFAULT_NF and at 0˚N, 0˚E, σ=0
+
+# keyword constructors
+Particle{NF}(;lon,lat,σ=0) where NF = Particle{NF}(lon,lat,σ)
+Particle{NF,isactive}(;lon,lat,σ=0) where {NF,isactive} = Particle{NF,isactive}(lon,lat,σ)
+Particle(;lon,lat,σ=0) = Particle(lon,lat,σ)
+Particle(;lon,lat,σ=0) = Particle(lon,lat,σ)
+
+# empty constructor map to zero
+Particle() = zero(Particle)
+Particle{NF}() where NF = zero(Particle{NF})
+Particle{NF,isactive}() where {NF,isactive} = zero(Particle{NF,isactive})
+Particle{NF}(args...) where NF = Particle{NF,true}(args...)
+
+# promotion of arguments
+Particle(lon,lat) = Particle(lon,lat,0)
+Particle(lon::Integer,lat::Integer) = Particle(lon,lat,0)
+Particle(args::Integer...) = Particle{DEFAULT_NF,true}(args...)
+Particle(args...) = Particle{promote_type(typeof.(args)...),true}(args...)
+
+# zero generators
+Base.zero(::Type{Particle}) = Particle{DEFAULT_NF,true}(0,0,0)
+Base.zero(::Type{Particle{NF}}) where NF = Particle{NF,true}(0,0,0)
+Base.zero(::Type{Particle{NF,isactive}}) where {NF,isactive} = Particle{NF,isactive}(0,0,0)
+Base.zero(::P) where {P<:Particle} = zero(P)
+
+Base.rand(rng::Random.AbstractRNG, ::Random.Sampler{Particle}) = rand(rng,Particle{DEFAULT_NF,true})
+Base.rand(rng::Random.AbstractRNG, ::Random.Sampler{Particle{NF}}) where NF = rand(rng,Particle{NF,true})
+function Base.rand(rng::Random.AbstractRNG, ::Random.Sampler{Particle{NF,isactive}}) where {NF,isactive}
+    lon = 360*rand(rng,NF)
+    yπ = convert(NF,π)*(2rand(rng,NF)-1)    # yπ ∈ [-1,1]*π
+    lat = sign(yπ)*acosd((1 + cos(yπ))/2)   # cos-distributed latitude
+    σ = rand(rng,NF)
+    return Particle{NF,isactive}(lon,lat,σ)
+end
+
+
+# equality with same location and same activity
+function Base.:(==)(p1::Particle{NF1,active1},p2::Particle{NF2,active2}) where {NF1,active1,NF2,active2}
+    return (active1==active2) && (p1.lon == p2.lon) && (p1.lat == p2.lat) && (p1.σ == p2.σ)
+end
+
[…]
