### 2024-03-05 0aa03fa
Initial conditions modular for variables
