+++ src/dynamics/initial_conditions.jl
@@ -1,33 +1,55 @@
-abstract type InitialConditions end
+abstract type AbstractInitialConditions <: AbstractModelComponent end
 
-# EXPORT INITIAL CONDITIONS
-export  StartFromFile,
-        StartFromRest,
-        ZonalJet,
-        ZonalWind,
-        StartWithRandomVorticity
-
-Base.@kwdef struct StartFromRest <: InitialConditions 
-    pressure_on_orography::Bool = false
+export InitialConditions
+Base.@kwdef struct InitialConditions{V,P,T,H} <: AbstractInitialConditions
+    vordiv::V = ZeroInitially()
+    pres::P = ZeroInitially()
+    temp::T = ZeroInitially()
+    humid::H = ZeroInitially()
 end
 
-function initialize!(   progn::PrognosticVariables,
-                        initial_conditions::StartFromRest,
-                        model::ModelSetup)
-    return nothing          # everything remains zero 
+function initialize!(
+    progn::PrognosticVariables,
+    IC::InitialConditions,
+    model::ModelSetup
+)
+    has(model,:vor) && initialize!(progn,IC.vordiv,model)
+    has(model,:pres) && initialize!(progn,IC.pres,model)
+    has(model,:temp) && initialize!(progn,IC.temp,model)
+    has(model,:humid) && initialize!(progn,IC.humid,model)
 end
 
-function initialize!(   progn::PrognosticVariables,
-                        initial_conditions::StartFromRest,
-                        model::PrimitiveEquation)
-    homogeneous_temperature!(progn,model)
-    initial_conditions.pressure_on_orography && pressure_on_orography!(progn,model)
-    # TODO initialise humidity
+InitialConditions(::Type{<:Barotropic}) = InitialConditions(;vordiv = StartWithRandomVorticity())
+InitialConditions(::Type{<:ShallowWater}) = InitialConditions(;vordiv = ZonalJet())
+function InitialConditions(::Type{<:PrimitiveDry})
+    vordiv = ZonalWind()
+    pres = PressureOnOrography()
+    temp = JablonowskiTemperature()
+    return InitialConditions(;vordiv,pres,temp)
 end
+
+function InitialConditions(::Type{<:PrimitiveWet})
+    vordiv = ZonalWind()
+    pres = PressureOnOrography()
+    temp = JablonowskiTemperature()
+    humid = ConstantRelativeHumidity()
+    return InitialConditions(vordiv,pres,temp,humid)
+end
+
+export ZeroInitially
+struct ZeroInitially <: AbstractInitialConditions end
+initialize!(::PrognosticVariables,::ZeroInitially,::ModelSetup) = nothing
+
+# to avoid a breaking change, like ZeroInitially
+export StartFromRest
+struct StartFromRest <: AbstractInitialConditions end
+initialize!(::PrognosticVariables,::StartFromRest,::ModelSetup) = nothing
+
+export StartWithRandomVorticity
 
 """Start with random vorticity as initial conditions
 $(TYPEDFIELDS)"""
-Base.@kwdef mutable struct StartWithRandomVorticity <: InitialConditions
+Base.@kwdef mutable struct StartWithRandomVorticity <: AbstractInitialConditions
     "Power of the spectral distribution k^power"
     power::Float64 = -3
 
@@ -57,12 +79,13 @@ function initialize!(   progn::PrognosticVariables{NF},
     end
 end
 
+export ZonalJet
+
 """
-$(TYPEDSIGNATURES)
-Create a struct that contains all parameters for the Galewsky et al, 2004 zonal jet
[…]
