### 2024-03-05 9184f68
Remove DEFAULT_NF generators of Earth and EarthAtmosphere
