+++ src/dynamics/planet.jl
@@ -42,5 +42,4 @@ Base.@kwdef mutable struct Earth{NF<:AbstractFloat} <: AbstractPlanet
     solar_constant::NF = 1365
 end
 
-Earth(;kwargs...) = Earth{DEFAULT_NF}(;kwargs...)
 Earth(SG::SpectralGrid;kwargs...) = Earth{SG.NF}(;kwargs...)
