+++ .github/workflows/julianightly.yml
@@ -13,7 +13,7 @@ jobs:
         with:
           version: nightly
           arch: x64
-      - uses: actions/cache@v3
+      - uses: actions/cache@v4
         env:
           cache-name: cache-artifacts
         with:
