### 2024-03-06 #6323 @vavrines
Everything will be fine with me once https://github.com/SpeedyWeather/SpeedyWeather.jl/issues/446 gets closed.
