+++ src/dynamics/implicit.jl
@@ -4,8 +4,8 @@ abstract type AbstractImplicit <: AbstractModelComponent end
 export NoImplicit
 struct NoImplicit <: AbstractImplicit end
 NoImplicit(SG::SpectralGrid) = NoImplicit()
-initialize!(::NoImplicit,args...) = nothing
-implicit_correction!(::DiagnosticVariables,::PrognosticVariables,::NoImplicit) = nothing
+initialize!(::NoImplicit, args...) = nothing
+implicit_correction!(::DiagnosticVariables, ::PrognosticVariables, ::NoImplicit) = nothing
 
 # SHALLOW WATER MODEL
 export ImplicitShallowWater
@@ -25,21 +25,21 @@ Base.@kwdef struct ImplicitShallowWater{NF<:AbstractFloat} <: AbstractImplicit
     # PRECOMPUTED ARRAYS, to be initiliased with initialize!
     H::Base.RefValue{NF} = Ref(zero(NF))    # layer_thickness
     ξH::Base.RefValue{NF} = Ref(zero(NF))   # = 2αΔt*layer_thickness, store in RefValue for mutability
-    g∇²::Vector{NF} = zeros(NF,trunc+2)     # = gravity*eigenvalues
-    ξg∇²::Vector{NF} = zeros(NF,trunc+2)    # = 2αΔt*gravity*eigenvalues
-    S⁻¹::Vector{NF} = zeros(NF,trunc+2)     # = 1 / (1-ξH*ξg∇²), implicit operator
+    g∇²::Vector{NF} = zeros(NF, trunc+2)     # = gravity*eigenvalues
+    ξg∇²::Vector{NF} = zeros(NF, trunc+2)    # = 2αΔt*gravity*eigenvalues
+    S⁻¹::Vector{NF} = zeros(NF, trunc+2)     # = 1 / (1-ξH*ξg∇²), implicit operator
 end
 
 """
 $(TYPEDSIGNATURES)
 Generator using the resolution from `spectral_grid`."""
 function ImplicitShallowWater(spectral_grid::SpectralGrid;kwargs...) 
-    (;NF,trunc) = spectral_grid
-    return ImplicitShallowWater{NF}(;trunc,kwargs...)
+    (;NF, trunc) = spectral_grid
+    return ImplicitShallowWater{NF}(;trunc, kwargs...)
 end
 
 # function barrier to unpack the constants struct for shallow water
-function initialize!(I::ImplicitShallowWater,dt::Real,::DiagnosticVariables,model::ShallowWater)
+function initialize!(I::ImplicitShallowWater, dt::Real, ::DiagnosticVariables, model::ShallowWater)
     initialize!(I, dt, model.planet, model.atmosphere)
 end
 
@@ -61,7 +61,7 @@ function initialize!(
     # α = 0   means the gravity wave terms are evaluated at i-1 (forward)
     # α = 0.5 evaluates at i+1 and i-1 (centered implicit)
     # α = 1   evaluates at i+1 (backward implicit)
-    # α ∈ [0.5,1] are also possible which controls the strength of the gravity wave dampening.
+    # α ∈ [0.5, 1] are also possible which controls the strength of the gravity wave dampening.
     # α = 0.5 slows gravity waves and prevents them from amplifying
     # α > 0.5 will dampen the gravity waves within days to a few timesteps (α=1)
 
@@ -70,7 +70,7 @@ function initialize!(
     ξH[] = ξ*layer_thickness   # update ξ*H with new ξ, in RefValue for mutability
 
     # loop over degree l of the harmonics (implicit terms are independent of order m)
-    @inbounds for l in eachindex(g∇²,ξg∇²,S⁻¹)
[…]
