+++ src/SpeedyTransforms/spectral_transform.jl
@@ -10,8 +10,8 @@ struct SpectralTransform{NF<:AbstractFloat}
     nlat_half::Int              # resolution parameter of grid (# of latitudes on one hemisphere, Eq incl)
 
     # SPECTRAL RESOLUTION
-    lmax::Int               # Maximum degree l=[0,lmax] of spherical harmonics
-    mmax::Int               # Maximum order m=[0,l] of spherical harmonics
+    lmax::Int               # Maximum degree l=[0, lmax] of spherical harmonics
+    mmax::Int               # Maximum order m=[0, l] of spherical harmonics
     nfreq_max::Int          # Maximum (at Equator) number of Fourier frequencies (real FFT)
     m_truncs::Vector{Int}   # Maximum order m to retain per latitude ring
 
@@ -21,13 +21,13 @@ struct SpectralTransform{NF<:AbstractFloat}
     nlat::Int               # Number of latitude rings
 
     # CORRESPONDING GRID VECTORS
-    colat::Vector{NF}                   # Gaussian colatitudes (0,π) North to South Pole 
+    colat::Vector{NF}                   # Gaussian colatitudes (0, π) North to South Pole 
     cos_colat::Vector{NF}               # Cosine of colatitudes
     sin_colat::Vector{NF}               # Sine of colatitudes
     lon_offsets::Matrix{Complex{NF}}    # Offset of first lon per ring from prime meridian
 
     # NORMALIZATION
-    norm_sphere::NF         # normalization of the l=0,m=0 mode
+    norm_sphere::NF         # normalization of the l=0, m=0 mode
 
     # FFT plans, one plan for each latitude ring
     rfft_plans::Vector{AbstractFFTs.Plan}     # FFT plan for grid to spectral transform
@@ -50,11 +50,11 @@ struct SpectralTransform{NF<:AbstractFloat}
     grad_y1::LowerTriangularMatrix{NF}      # precomputed meridional gradient factors, term 1
     grad_y2::LowerTriangularMatrix{NF}      # term 2
 
-    # GRADIENT MATRICES FOR U,V -> Vorticity,Divergence
+    # GRADIENT MATRICES FOR U, V -> Vorticity, Divergence
     grad_y_vordiv1::LowerTriangularMatrix{NF}
     grad_y_vordiv2::LowerTriangularMatrix{NF}
 
-    # GRADIENT MATRICES FOR Vorticity,Divergence -> U,V
+    # GRADIENT MATRICES FOR Vorticity, Divergence -> U, V
     vordiv_to_uv_x::LowerTriangularMatrix{NF}
     vordiv_to_uv1::LowerTriangularMatrix{NF}
     vordiv_to_uv2::LowerTriangularMatrix{NF}
@@ -67,7 +67,7 @@ end
 """
 $(TYPEDSIGNATURES)
 Generator function for a SpectralTransform struct. With `NF` the number format,
-`Grid` the grid type `<:AbstractGrid` and spectral truncation `lmax,mmax` this function sets up
+`Grid` the grid type `<:AbstractGrid` and spectral truncation `lmax, mmax` this function sets up
 necessary constants for the spetral transform. Also plans the Fourier transforms, retrieves the colatitudes,
 and preallocates the Legendre polynomials (if recompute_legendre == false) and quadrature weights."""
 function SpectralTransform( ::Type{NF},                     # Number format NF
@@ -80,61 +80,61 @@ function SpectralTransform( ::Type{NF},                     # Number format NF
                             ) where NF
 
[…]
