### 2024-03-06 08896fa
fix geostrophy example + execute via docs
