### 2024-03-06 4c8f40f
fix docstring rendering
