+++ src/output/callbacks.jl
@@ -48,8 +48,8 @@ export add!
 $(TYPEDSIGNATURES)
 Add a or several callbacks to a Dict{String,AbstractCallback} dictionary. To be used like
 
-add!(model.callbacks,:my_callback => callback)
-add!(model.callbacks,:my_callback1 => callback, :my_callback2 => other_callback)
+    add!(model.callbacks,:my_callback => callback)
+    add!(model.callbacks,:my_callback1 => callback, :my_callback2 => other_callback)
 """
 function add!(D::CALLBACK_DICT, key_callbacks::Pair{Symbol, <:AbstractCallback}...)
     for key_callback in key_callbacks
