### 2024-03-06 #6323 @natgeo-wong
I would refer @kthyng to https://github.com/SpeedyWeather/SpeedyWeather.jl/issues/443, there's still quite a bit of work to be done especially re. documentation.
