### 2024-03-06 1857358
some code alignment + delete unnecessary spaces
