+++ src/dynamics/planet.jl
@@ -42,4 +42,4 @@ Base.@kwdef mutable struct Earth{NF<:AbstractFloat} <: AbstractPlanet
     solar_constant::NF = 1365
 end
 
-Earth(SG::SpectralGrid;kwargs...) = Earth{SG.NF}(;kwargs...)
+Earth(SG::SpectralGrid; kwargs...) = Earth{SG.NF}(; kwargs...)
