+++ src/physics/thermodynamics.jl
@@ -40,8 +40,8 @@ end
 
 # generator function
 function ClausiusClapeyron(SG::SpectralGrid, atm::AbstractAtmosphere; kwargs...)
-    (;R_dry, R_vapour, latent_heat_condensation, heat_capacity) = atm
-    return ClausiusClapeyron{SG.NF}(;Lᵥ=latent_heat_condensation, R_dry, R_vapour, cₚ=heat_capacity, kwargs...)
+    (; R_dry, R_vapour, latent_heat_condensation, heat_capacity) = atm
+    return ClausiusClapeyron{SG.NF}(; Lᵥ=latent_heat_condensation, R_dry, R_vapour, cₚ=heat_capacity, kwargs...)
 end
 
 """
@@ -54,7 +54,7 @@ Clausius-Clapeyron equation,
 where T is in Kelvin, Lᵥ the the latent heat of vaporization and Rᵥ the gas constant
 of water vapour, T₀ is 0˚C in Kelvin."""
 function (CC::ClausiusClapeyron{NF})(temp_kelvin::NF) where NF
-    (;e₀, T₀⁻¹, Lᵥ_Rᵥ) = CC
+    (; e₀, T₀⁻¹, Lᵥ_Rᵥ) = CC
     return e₀ * exp(Lᵥ_Rᵥ*(T₀⁻¹ - inv(temp_kelvin)))
 end
 
@@ -104,7 +104,7 @@ function saturation_humidity(
     clausius_clapeyron::AbstractClausiusClapeyron,
 ) where NF
     sat_vap_pres = clausius_clapeyron(temp_kelvin)
-    return saturation_humidity(sat_vap_pres, pres;mol_ratio=clausius_clapeyron.mol_ratio)
+    return saturation_humidity(sat_vap_pres, pres; mol_ratio=clausius_clapeyron.mol_ratio)
 end
 
 """
@@ -132,7 +132,7 @@ function dry_static_energy!(
     atmosphere::AbstractAtmosphere
 )
     cₚ = atmosphere.heat_capacity
-    (;dry_static_energy, geopot, temp) = column
+    (; dry_static_energy, geopot, temp) = column
 
     @inbounds for k in eachlayer(column)
         dry_static_energy[k] = cₚ * temp[k] + geopot[k]
@@ -146,7 +146,7 @@ function bulk_richardson!(
     atmosphere::AbstractAtmosphere,
 )
     cₚ = atmosphere.heat_capacity
-    (;u, v, geopot, temp_virt, nlev, bulk_richardson) = column
+    (; u, v, geopot, temp_virt, nlev, bulk_richardson) = column
 
     V² = u[nlev]^2 + v[nlev]^2
     Θ₀ = cₚ*temp_virt[nlev]
@@ -169,7 +169,7 @@ function saturation_humidity!(
     column::ColumnVariables,
     clausius_clapeyron::AbstractClausiusClapeyron,
 )
-    (;sat_humid, pres, temp) = column
+    (; sat_humid, pres, temp) = column
 
     for k in eachlayer(column)
         sat_humid[k] = saturation_humidity(temp[k], pres[k], clausius_clapeyron)
@@ -188,9 +188,9 @@ function moist_static_energy!(
     column::ColumnVariables,
     clausius_clapeyron::AbstractClausiusClapeyron
 )
-    (;Lᵥ) = clausius_clapeyron      # latent heat of vaporization
-    (;sat_moist_static_energy, moist_static_energy, dry_static_energy) = column
-    (;humid, sat_humid) = column
+    (; Lᵥ) = clausius_clapeyron      # latent heat of vaporization
+    (; sat_moist_static_energy, moist_static_energy, dry_static_energy) = column
[…]
