### 2024-03-07 ce74382
Clock with start and counter
