+++ src/dynamics/planet.jl
@@ -44,3 +44,4 @@ end
 
 Earth(;kwargs...) = Earth{DEFAULT_NF}(;kwargs...)
 Earth(SG::SpectralGrid;kwargs...) = Earth{SG.NF}(;kwargs...)
+Earth(::Type{NF};kwargs...) where NF = Earth{NF}(;kwargs...)
