### 2024-03-07 a24ccb0
Particle advection in the ShallowWaterModel too
