+++ src/models/primitive_dry.jl
@@ -119,7 +119,8 @@ function initialize!(model::PrimitiveDry; time::DateTime = DEFAULT_DATE)
     prognostic_variables = PrognosticVariables(spectral_grid, model)
     initialize!(prognostic_variables, model.initial_conditions, model)
     (;clock) = prognostic_variables
-    clock.time = time       # set the time
+    clock.time = time       # set the current time
+    clock.start = time      # and store the start time
 
     # initialize ocean and land and synchronize clocks
     initialize!(prognostic_variables.ocean, clock.time, model)
