+++ src/output/feedback.jl
@@ -72,11 +72,6 @@ $(TYPEDSIGNATURES)
 Initializes the a `Feedback` struct."""
 function initialize!(feedback::Feedback,clock::Clock,model::ModelSetup)
 
-    # reinitalize progress meter, minus one to exclude first_timesteps! which contain compilation
-    (;showspeed, desc) = feedback.progress_meter
-    (;verbose) = feedback
-    feedback.progress_meter = ProgressMeter.Progress(clock.n_timesteps-1;enabled=verbose, showspeed, desc)
-    
     # set to false to recheck for NaRs
     feedback.nars_detected = false
 
@@ -101,6 +96,12 @@ function initialize!(feedback::Feedback,clock::Clock,model::ModelSetup)
         write(progress_txt,"\nAll data will be stored in $run_path\n")
         feedback.progress_txt = progress_txt
     end
+
+    # reinitalize progress meter, minus one to exclude first_timesteps! which contain compilation
+    # only do now for benchmark accuracy
+    (;showspeed, desc) = feedback.progress_meter
+    (;verbose) = feedback
+    feedback.progress_meter = ProgressMeter.Progress(clock.n_timesteps-1;enabled=verbose, showspeed, desc)
 end
 
 """
