### 2024-03-07 466a0b9
ShallowWaterModel with AbstractInitialConditions
