+++ src/dynamics/abstract_types.jl
@@ -1,1 +1,5 @@
 abstract type AbstractTimeStepper <: AbstractModelComponent end
+
+abstract type AbstractVariables end
+abstract type AbstractPrognosticVariables <: AbstractVariables end
+abstract type AbstractDiagnosticVariables <: AbstractVariables end
