+++ src/SpeedyWeather.jl
@@ -101,6 +101,7 @@ include("dynamics/implicit.jl")
 include("dynamics/scaling.jl")
 include("dynamics/tendencies.jl")
 include("dynamics/hole_filling.jl")
+include("dynamics/particle_advection.jl")
 
 # PARAMETERIZATIONS
 include("physics/albedo.jl")
