### 2024-03-08 6d249c0
Particle docstring and tests
