+++ src/dynamics/particles.jl
@@ -1,13 +1,24 @@
 abstract type AbstractParticle{NF} end
 
 export Particle
+
+"""
+Particle with location lon (longitude), lat (latitude) and σ (vertical coordinate).
+Longitude is assumed to be in [0,360˚E), latitude in [-90˚,90˚N] and σ in [0,1] but
+not strictly enforced at creation, see `mod(::Particle)` and `ismod(::Particle)`.
+A particle is either active or inactive, determined by the Boolean in it's 2nd
+type parameter.
+$(TYPEDFIELDS)"""
 struct Particle{
     NF<:AbstractFloat,  # number format of coordinates
     isactive,           # bool indicating whether particle is active
 } <: AbstractParticle{NF}
-    lon::NF             # longitude [0,360˚]
-    lat::NF             # latitude [-90˚,90˚]
-    σ::NF               # vertical sigma coordinate [0,1] (top to surface)
+    "longitude in [0,360˚]"
+    lon::NF
+    "latitude [-90˚,90˚]"
+    lat::NF
+    "vertical sigma coordinate [0,1] (top to surface)"
+    σ::NF
 end
 
 # particle is by default active, of number format DEFAULT_NF and at 0˚N, 0˚E, σ=0
@@ -37,6 +48,8 @@ Base.zero(::P) where {P<:Particle} = zero(P)
 
 Base.rand(rng::Random.AbstractRNG, ::Random.Sampler{Particle}) = rand(rng,Particle{DEFAULT_NF,true})
 Base.rand(rng::Random.AbstractRNG, ::Random.Sampler{Particle{NF}}) where NF = rand(rng,Particle{NF,true})
+
+# rand uniformly distributed over the globe with cos-distribution for poles
 function Base.rand(rng::Random.AbstractRNG, ::Random.Sampler{Particle{NF,isactive}}) where {NF,isactive}
     lon = 360*rand(rng,NF)
     yπ = convert(NF,π)*(2rand(rng,NF)-1)    # yπ ∈ [-1,1]*π
@@ -47,7 +60,19 @@ end
 
 # equality with same location and same activity
 function Base.:(==)(p1::Particle{NF1,active1},p2::Particle{NF2,active2}) where {NF1,active1,NF2,active2}
-    return (active1==active2) && (p1.lon == p2.lon) && (p1.lat == p2.lat) && (p1.σ == p2.σ)
+    return  (active1 == active2) &&     # both active or both inactive
+            (p1.lat == p2.lat) &&       # same latitude
+            (p1.σ == p2.σ) &&           # same elevation
+            ((p1.lon == p2.lon) ||      # same longitude OR
+            (p1.lat*p2.lat == 8100))    # both at the north/south pole, because 90˚N, 0˚E == 90˚N, 10˚E
+end
+
+function Base.isapprox(p1::Particle{NF1,active1},p2::Particle{NF2,active2}) where {NF1,active1,NF2,active2}
+    return  (active1 == active2) &&     # both active or both inactive
+            (p1.lat ≈ p2.lat) &&       # same latitude
+            (p1.σ ≈ p2.σ) &&           # same elevation
+            ((p1.lon ≈ p2.lon) ||      # same longitude OR
+            (p1.lat*p2.lat == 8100))    # both at the north/south pole, because 90˚N, 0˚E == 90˚N, 10˚E
 end
 
 function Base.show(io::IO,p::Particle{NF,isactive}) where {NF,isactive}
@@ -59,29 +84,61 @@ function Base.show(io::IO,p::Particle{NF,isactive}) where {NF,isactive}
 end
 
 export move
+
[…]
