### 2024-03-09 9176a9a
No move tests for Float16 because of rounding errors
