+++ test/runtests.jl
@@ -21,6 +21,7 @@ include("diffusion.jl")
 include("time_stepping.jl")
 include("vertical_advection.jl")
 include("particles.jl")
+include("particle_advection.jl")
 
 # VERTICAL LEVELS
 include("vertical_levels.jl")
