### 2024-03-09 b85ec3d
Higher error tolerance for particles moved 4x around globe
