### 2024-03-09 SpeedyWeather/SpeedyWeather.jl#480 @slayoo
JOSS review: text comments
reading through the JOSS paper text, I've noted down the following suggestions/comments:
- [x] the "monolithic" statements call for reference[s]
- [x] ditto for "modern trend sees simulations in Fortran and data analysis in Python" - please add reference[s]
- [x] what is "large-scale condensation"? (and how would it be different from other-scale condensation?)
- [x] why is Cartopy cited as "2010-2015"?
- [x] bib: Malardel et al. 2016 - DOI is missing: https://doi.org/10.21957/zwdu9u5i
- [x] bib: Moses & Churavy 2020 - add clickable link (https://proceedings.neurips.cc/paper_files/paper/2020/file/9332c513ef44b682e9347822c2e457ac-Paper.pdf or https://doi.org/10.48550/arXiv.2010.01709)
- [x] bib: Rasp et al. 2018 - DOI is missing: https://doi.org/10.1073/pnas.1810286115
- [x] bib: Stompor 2011 - add clickable URL: https://ascl.net/1110.013
- [x] bib: Willmert 2020 - use permalink: http://web.archive.org/web/*/https://justinwillmert.com/articles/2020/notes-on-calculating-the-spherical-harmonics/
- [x] paper mentions that "simple parallelization across vertical layers is supported" but README informs also about "grid point-wise [multi-threading] for physics", and CUDA is among the dependencies - please clarify and elaborate
- [x] if distributed-memory parallelisation (e.g., MPI) is not supported, please mention it explicitly

HTH
