### 2024-03-09 #6323 @slayoo
@natgeo-wong, I've posted my comments in the following issues:
- https://github.com/SpeedyWeather/SpeedyWeather.jl/issues/448
- https://github.com/SpeedyWeather/SpeedyWeather.jl/issues/449
- https://github.com/SpeedyWeather/SpeedyWeather.jl/issues/480
