### 2024-03-11 89d4958
Particle empty constructors removed
