+++ test/particles.jl
@@ -1,11 +1,12 @@
 @testset "Create and compare particles" begin
-    @test Particle() == Particle(0,0) == Particle(0,0,0)
-    @test Particle(0,0,0) == zero(Particle) == zero(Particle{Float32}) == zero(Particle{Float32}())
-    @test Particle() == Particle(0.0,0.0) == Particle(0f0,0f0,0.0) == Particle(0,Float16(0))
-    @test Particle{Float32}() == Particle{Float32,true}()
-    @test Particle{Float32}() != Particle{Float32,false}()
-    @test deactivate(Particle()) != Particle()
-    @test activate(Particle()) == Particle()
+    @test Particle(0, 0) == Particle(0, 0, 0)
+    @test Particle(0, 0, 0) == zero(Particle) == zero(Particle{Float32}) == zero(Particle{Float32, true})
+    @test Particle(0.0, 0.0) == Particle(0f0, 0f0, 0.0) == Particle(0, Float16(0))
+    @test zero(Particle{Float32}) == Particle{Float32,true}(0, 0)
+    @test zero(Particle{Float32}) != Particle{Float32,false}(0, 0)
+    @test deactivate(zero(Particle)) != zero(Particle)
+    @test activate(zero(Particle)) == zero(Particle)
+
     @test Particle(1,2) != Particle(2,1)
     @test Particle(lon=1,lat=2) == Particle(lat=2,lon=1.0)
     @test Particle(lon=1,lat=2,σ=0) == Particle(lat=2,lon=1.0)
