### 2024-03-11 6b1e680
SpeedyWeather. for ismod missing
