### 2024-03-11 0acce20
Add warning for 0-particle particle advection
