### 2024-03-11 fb2489d
particle advection 1/(coslat+eps) to prevent division by zero for advection on the pole
