+++ docs/src/particles.md
@@ -1,0 +1,244 @@
+# Particle advection
+
+All SpeedyWeather.jl models support particle advection. Particles are objects without mass or volume
+at a location ``\mathbf{x} = (\lambda, \theta, \sigma)`` (longitude ``\lambda``, latitude ``\theta``,
+vertical sigma coordinate ``\sigma``, see [Sigma coordinates](@ref)) that are moved with the wind
+``\mathbf{u}(\mathbf{x})`` at that location. The location of the ``p``-th particle changes as follows
+
+```math
+\frac{d \mathbf{x}_p}{d t} = \mathbf{u}(\mathbf{x}_p)
+```
+
+This equation applies in 2D, i.e. ``\mathbf{x} = (\lambda, \theta)`` and ``\mathbf{u} = (u, v)`` or
+in 3D, but at the moment only 2D advection is supported. In the [Primitive equation model](@ref)
+the vertical layer on which the advection takes place has to be specified. It is therefore not
+advected with the vertical velocity but maintains a constant pressure ratio compared to the
+surface pressure (``\sigma`` is constant).
+
+## Discretization of particle advection
+
+The particle advection equation has to be discretized to be numerically solved. While the
+particle location can generally be anywhere on the sphere, the velocity ``\mathbf{u}``
+is only available on the discrete grid points of the simulation, such that
+``\mathbf{u}(\mathbf{x}_p)`` requires an interpolation in order to obtain a velocity
+at the particles location ``\mathbf{x}_p`` to move it around. Dropping the subscript
+``p`` in favour a subscript ``i`` denoting the time step, with Euler forward 
+the equation can be discretized as 
+
+```math
+\mathbf{x}_{i+1} = \mathbf{x}_i + \Delta t~\mathbf{u}_i (\mathbf{x}_i)
+```
+
+Meaning we have used the velocity field at both departure time ``i`` and departure
+location ``\mathbf{x}_i`` to update a particle's location which makes this scheme
+first order accurate. But only a single interpolation of the velocity field,
+which, in fact, is one per dimension, is necessary. Note that the time step
+``\Delta t`` here and the time step to solve the dynamics do not have to be
+identical. We could use a larger time step for the particle advection then to
+solve the dynamics inside the model, and because the stability criteria for these
+equations are different, one is encouraged to do so. Also because the particles
+are considered passive, meaning that their location does not influence the
+other prognostic variables.
+
+We can write down a more accurate scheme at the cost of a second interpolation step.
+The Heun method, also called predictor-corrector is 2nd order accurate and uses an average of the
+velocity at departure time ``i`` and location ``\mathbf{x}_i`` and at a
+(_predicted_ meaning preliminary) arrival point ``x^\star_{i+1}`` and arrival time ``i+1``.
+
+```math
+\begin{aligned}
+\mathbf{x}^\star_{i+1} &= \mathbf{x}_i + \Delta t~\mathbf{u}_i (\mathbf{x}_i) \\
+\mathbf{x}_{i+1} &= \mathbf{x}_i + \frac{\Delta t}{2}~\left(
+    \mathbf{u}_i (\mathbf{x}_i) + \mathbf{u}_{i+1} (\mathbf{x}^\star_{i+1})\right)
+\end{aligned}
+```
+
+Because we don't have ``\mathbf{u}_{i+1}`` available at time ``i``, we perform this
+integration retrospectively, i.e. if the other model dynamics have reached time ``i+1``
+then we let the particle advection catch up by integrating them from ``i`` to ``i+1``.
+This, however, requires some storage of the velocity ``\mathbf{u}_i`` at the previous
+advection time step. Remember that this does not need to be the time step for the
+momentum equations and could be much further in the past. We could either store
[…]
