### 2024-03-11 801da6b
use custom model ref names for robustness
this way links won't break if we change the model page titles
