### 2024-03-11 191b41b
resume CI on julia v1.8+1.9
