+++ .github/workflows/CI.yml
@@ -13,8 +13,8 @@ jobs:
     strategy:
       matrix:
         version:
-          # - '1.8'
-          # - '1.9'
+          - '1.8'
+          - '1.9'
           - '1.10'
         os:
           - ubuntu-latest
