### 2024-03-11 db45b9a
try skipping set_vars tests
