+++ test/runtests.jl
@@ -7,7 +7,7 @@ include("dates.jl")
 include("lower_triangular_matrix.jl")
 include("grids.jl")
 include("interpolation.jl")
-include("set_vars.jl")
+# include("set_vars.jl")
 
 # GPU/KERNELABSTRACTIONS
 include("kernelabstractions.jl")
