+++ README.md
@@ -8,30 +8,74 @@
 SpeedyWeather.jl is a global spectral atmospheric model with simple physics which is developed as a research playground
 with an everything-flexible attitude as long as it is speedy. With minimal code redundancies it supports
 
-- different physical models (barotropic vorticity, shallow water, primitive equations dry and wet core)
-- physics parameterizations for precipitation, boundary layer, etc; new ones are passed as argument to the model constructor
-- different spatial grids (full and octahedral grids, Gaussian and Clenshaw-Curtis, HEALPix, OctaHEALPix)
-- different horizontal resolutions (T31 to T1023 and higher, i.e. 400km to 10km using linear, quadratic or cubic truncation)
-- 32-bit single and 64-bit double precision (16 bits in development)
-- multi-threading, layer-wise for dynamics, grid point-wise for physics (multi-processing in development)
-- different architectures (x86 and arm, GPUs in development)
+**Dynamics and physics**
+- Different physical equations (barotropic vorticity, shallow water, primitive equations)
+- Particle advection in 2D for all equations
+- Physics parameterizations for convection, precipitation, boundary layer, etc.
+
+**Numerics and computing**
+- Different spatial grids (full and octahedral grids, Gaussian and Clenshaw-Curtis, HEALPix, OctaHEALPix)
+- Different resolutions (T31 to T1023 and higher, i.e. 400km to 10km using linear, quadratic or cubic truncation)
+- Different arithmetics: Float32 (default), Float64, and (experimental) BFloat16, stochastic rounding
+- multi-threading, layer-wise for dynamics, grid point-wise for physics
+
+**User interface**
+- Extensibility: New model components (incl. parameterizations) can be externally defined
+- Modularity: Models are constructed from its components, non-defaults are passed on as argument
+- Interactivity: SpeedyWeather.jl runs in a notebook or in the REPL as well as from scripts
+- Callbacks can be used to inject any piece of code after every time step, e.g. custom output, event handling, changing the model while it's running
 
 and Julia will compile to these choices just-in-time.
 
 For an overview of the functionality and explanation see the
 [documentation](https://speedyweather.github.io/SpeedyWeather.jl/dev).
+But as the documentation always lags behind our full functionality you are encouraged
+to [raise an issue](https://github.com/SpeedyWeather/SpeedyWeather.jl/issues) describing
+what you'd like to use SpeedyWeather for.
+
+## Vision and roadmap
+
+Why another model? You may ask. We believe that most currently available are stiff, difficult to use
+and extend, and therefore slow down research whereas a modern code in a modern language wouldn't have to.
+We decided to use Julia because it combines the best of Fortran and Python: Within a single language
+we can interactively run SpeedyWeather but also extend it, inspect its components, evaluate
+individual terms of the equations, and analyse and visualise output on the fly.
+
+We do not aim to make SpeedyWeather an atmospheric model similar to the production-ready models used
+in weather forecasting, at least not at the cost of our current level of interactivity and ease of
+use or extensibility. If someone wants to implement a cloud parameterization that is very complicated
+and expensive to run then they are more than encouraged to do so, but it will probably live in
+its own repository and we are happy to provide a general interface to do so. But SpeedyWeather's
+defaults should be balanced: Physically accurate yet general; as independently as possible from other
+components and parameter choices; not too complicated to implement and understand; and computationally cheap.
+Finding a good balance is difficult but we try our best. 
+
[…]
