### 2024-03-13 aa3bbe5
Event and periodic schedule simultaneously
