+++ docs/src/callbacks.md
@@ -251,3 +251,152 @@ the simulation while running or to interfere with a simulation. We therefore
 encourage users to use callbacks as widely as possible, but if you run
 into any issues please open an issue in the repository and explain what
 you'd like to achieve and which errors you are facing. We are happy to help.
+
+## Schedules
+
+For convenience, SpeedyWeather.jl implements a [`Schedule`](@ref) which helps to
+schedule when callbacks are called. Because in many situations you don't want to call
+them on every time step but only periodically, say once a day, or only on specific
+dates and times, e.g. Jan 1 at noon. Several examples how to create schedules
+
+```@example schedule
+using SpeedyWeather
+
+# execute on timestep at or after Jan 2 2000
+event_schedule = Schedule(DateTime(2000,1,2))   
+
+# several events scheduled
+events = (DateTime(2000,1,3), DateTime(2000,1,5,12))
+several_events_schedule = Schedule(events...)
+
+# provided as Vector{DateTime} with times= keyword
+always_at_noon = [DateTime(2000,1,i,12) for i in 1:10]
+noon_schedule = Schedule(times=always_at_noon)
+
+# or using every= for periodic execution, here once a day
+periodic_schedule = Schedule(every=Day(1))
+```
+
+A `Schedule` has 5 fields, see [`Schedule`](@ref). `every` is an option
+to create a periodic schedule to execute every time that indicated
+period has passed. `steps` and `counter` will let you know how many
+callback execution steps there are and count them up. `times` is a 
+`Vector{DateTime}` containing scheduled events. `schedule` is the
+actual schedule inside a `Schedule`, implemented as `BitVector`
+indicating whether to execute on a given time step (`true`) or not
+(`false`).
+
+Let's show how to use a `Schedule` inside a callback
+
+```@example schedule
+struct MyScheduledCallback <: SpeedyWeather.AbstractCallback
+    schedule::Schedule
+    # add other fields here that you need
+end
+
+function SpeedyWeather.initialize!(
+    callback::MyScheduledCallback,
+    progn::PrognosticVariables,
+    args...
+)
+    # when initializing a scheduled callback also initialize its schedule!
+    initialize!(callback.schedule, progn.clock)
+
+    # initialize other things in your callback here
+end
+
+function SpeedyWeather.callback!(
+    callback::MyScheduledCallback,
+    progn::PrognosticVariables,
+    diagn::DiagnosticVariables,
+    model::ModelSetup,
+)
+    # scheduled callbacks start with this line to execute only when scheduled!
+    # else escape immediately
+    SpeedyWeather.isscheduled(callback.schedule, progn.clock) || return nothing
+
+    # Just print the North Pole surface temperature to screen
+    (;time) = progn.clock
+    temp_at_north_pole = diagn.layers[end].grid_variables.temp_grid[1]
+    @info "North pole has a temperature of $temp_at_north_pole on $time."
+end
+
+# nothing needs to be done when finishing
+SpeedyWeather.finish!(::MyScheduledCallback, args...) = nothing
+```
+
+So in summary
+- add a field `schedule::Schedule` to your callback
+- add the line `initialize!(callback.schedule, progn.clock)` when initializing your callback
+- start your `callback!` method with `isscheduled(callback.schedule, progn.clock) || return nothing` to execute only when scheduled
+
+A `Schedule` is a field inside a callback as this allows you the set the callbacks
+desired schedule when creating it. In the example above we can create our callback
+that is supposed to print the North Pole's temperature like so
+```@example schedule
+north_pole_temp_at_noon_jan9 = MyScheduledCallback(Schedule(DateTime(2000,1,9,12)))
+```
+The default for `every` is `typemax(Int)` indicating "never". This just means that there
+is no periodically reoccuring schedule, only `schedule.times` would include some times
[…]
