### 2024-03-13 c6ba834
rm SpeedyWeather. for issscheduled
