+++ src/SpeedyWeather.jl
@@ -137,4 +137,5 @@ include("models/barotropic.jl")
 include("models/shallow_water.jl")
 include("models/primitive_dry.jl")
 include("models/primitive_wet.jl")
+include("models/tree.jl")
 end
