### 2024-03-15 39b4372
Update docs/src/structure.md
