+++ src/dynamics/variables_3D_4D.jl
@@ -1,0 +1,172 @@
+## GRID 3D
+export GridVariable3D
+
+"""
+A horizontal grid::AbstractGrid with a third dimension (time or vertical).
+As the grid is a AbstractVector (with a single index only) a GridVariable3D
+is a subtype of AbstractMatrix (and not AbstractArray{T,3}) with the horizontal
+dimension unravelled into its first dimension. To be generated and indexed like
+
+    nlat_half = 24      # resolution parameter of the horizontal grid
+    nlayer = 8          # number of vertical layers (or timesteps)
+    G = zeros(GridVariable3D, OctahedralGaussianGrid{Float32}, nlat_half, nlayer)
+    G[50, 3]            # the 50th grid point on the 3rd layer
+
+Fields are
+$(TYPEDFIELDS)"""
+struct GridVariable3D{NF, Grid<:AbstractGrid{NF}} <: AbstractMatrix{NF}
+    data::Vector{Grid}
+    npoints2D::Int
+end
+
+function Base.zeros(
+    ::Type{GridVariable3D},
+    Grid::Type{<:AbstractGrid{NF}},
+    nlat_half::Integer,
+    k::Integer,
+) where NF
+    npoints2D = RingGrids.get_npoints(Grid, nlat_half)
+    G3D = [zeros(Grid, nlat_half) for _ in 1:k]
+    GridVariable3D{NF, Grid}(G3D, npoints2D)
+end
+
+Base.size(G3D::GridVariable3D) = (G3D.npoints2D, length(G3D.data))
+
+@inline function Base.getindex(G::GridVariable3D, ij::Integer, k::Integer)
+    @boundscheck 0 < k <= length(G.data) || throw(BoundsError(G, (ij, k)))
+    @boundscheck 0 < ij <= G.npoints2D || throw(BoundsError(G, (ij, k)))
+    @inbounds r = G.data[k][ij]
+end
+
+Base.@propagate_inbounds function Base.setindex!(G::GridVariable3D, x, ij::Integer, k::Integer)
+    @boundscheck 0 < k <= length(G.data) || throw(BoundsError(G, (ij, k)))
+    @boundscheck 0 < ij <= G.npoints2D || throw(BoundsError(G, (ij, k)))
+    @inbounds G.data[k][ij] = x
+end
+
+## GRID 4D
+export GridVariable4D
+
+"""
+A horizontal grid::AbstractGrid with a third and forth dimension (vertical and time).
+As the grid is a AbstractVector (with a single index only) a GridVariable3D
+is a subtype of AbstractArray{T, 3} (and not 4) with the horizontal dimension
+unravelled into its first dimension. To be generated and indexed like
+
+    nlat_half = 24      # resolution parameter of the horizontal grid
+    nlayer = 8          # number of vertical layers
+    ntimesteps = 2      # number of time steps
+    G = zeros(GridVariable4D, OctahedralGaussianGrid{Float32}, nlat_half, nlayer, ntimesteps)
+    G[50, 3, 1]         # the 50th grid point on the 3rd layer and the 1st timestep
+
+Fields are
+$(TYPEDFIELDS)"""
+struct GridVariable4D{NF, Grid<:AbstractGrid{NF}} <: AbstractArray{NF, 3}
+    data::Matrix{Grid}
+    npoints2D::Int
+end
+
+function Base.zeros(
+    ::Type{GridVariable4D},
+    Grid::Type{<:AbstractGrid{NF}},
+    nlat_half::Integer,
+    k::Integer,
+    l::Integer,
+) where NF
+    npoints2D = RingGrids.get_npoints(Grid, nlat_half)
[…]
