### 2024-03-15 a6eeeeb
boundschecks in underlying vectors/matrices only
