+++ src/dynamics/variables_3D_4D.jl
@@ -32,16 +32,9 @@ end
 
 Base.size(G3D::GridVariable3D) = (G3D.npoints2D, length(G3D.data))
 
-@inline function Base.getindex(G::GridVariable3D, ij::Integer, k::Integer)
-    @boundscheck 0 < k <= length(G.data) || throw(BoundsError(G, (ij, k)))
-    @boundscheck 0 < ij <= G.npoints2D || throw(BoundsError(G, (ij, k)))
-    @inbounds r = G.data[k][ij]
-end
-
+Base.@propagate_inbounds Base.getindex(G::GridVariable3D, ij::Integer, k::Integer) = G.data[k][ij]
 Base.@propagate_inbounds function Base.setindex!(G::GridVariable3D, x, ij::Integer, k::Integer)
-    @boundscheck 0 < k <= length(G.data) || throw(BoundsError(G, (ij, k)))
-    @boundscheck 0 < ij <= G.npoints2D || throw(BoundsError(G, (ij, k)))
-    @inbounds G.data[k][ij] = x
+    G.data[k][ij] = x
 end
 
 ## GRID 4D
@@ -80,18 +73,9 @@ end
 
 Base.size(G4D::GridVariable4D) = (G4D.npoints2D, size(G4D.data)...)
 
-@inline function Base.getindex(G::GridVariable4D, ij::Integer, k::Integer, l::Integer)
-    @boundscheck 0 < k <= size(G.data,1) || throw(BoundsError(G, (ij, k, l)))
-    @boundscheck 0 < l <= size(G.data,2) || throw(BoundsError(G, (ij, k, l)))
-    @boundscheck 0 < ij <= G.npoints2D || throw(BoundsError(G, (ij, k, l)))
-    @inbounds G.data[k, l][ij]
-end
-
+Base.@propagate_inbounds Base.getindex(G::GridVariable4D, ij::Integer, k::Integer, l::Integer) = G.data[k, l][ij]
 Base.@propagate_inbounds function Base.setindex!(G::GridVariable4D, x, ij::Integer, k::Integer, l::Integer)
-    @boundscheck 0 < k <= size(G.data,1) || throw(BoundsError(G, (ij, k, l)))
-    @boundscheck 0 < l <= size(G.data,2) || throw(BoundsError(G, (ij, k, l)))
-    @boundscheck 0 < ij <= G.npoints2D || throw(BoundsError(G, (ij, k, l)))
-    @inbounds G.data[k, l][ij] = x
+    G.data[k, l][ij] = x
 end
 
 ## SPECTRAL 3D
@@ -99,17 +83,17 @@ export SpectralVariable3D
 
 """
 A horizontal LowerTriangularMatrix for spherical harmonic coefficients with a third
-dimension (time or vertical). Uses either (1) two indices lm, k with lm being the unravelled
-single index for the LowerTriangularMatrix (skipping the zeros in the upper triangle)
-and k the vertical layer or the timestep. Or (2) three indices l, m, k with l, m
-for the (1-indexed) spherical harmonic of degree l and order m.
+dimension (time or vertical). Uses either (1) two indices ij, k with ij being the unravelled
+single index for the LowerTriangularMatrix (skipping the zeros in the upper triangle, diagonal excl)
+and k the vertical layer or the timestep. Or (2) three indices i, j, k with i, j
+for the (1-indexed) spherical harmonic of degree i and order j.
 To be generated and indexed like
 
     m, n = 4, 4         # size of LowerTriangularMatrix
     nlayer = 8          # number of vertical layers (or timesteps)
     S = zeros(SpectralVariable3D{Float32}, m, n, nlayer)
[…]
