### 2024-03-15 e5558f9
name harmonics l,m consistently i, j for 3D 4D vars
