+++ src/SpeedyWeather.jl
@@ -40,6 +40,7 @@ include("utility_functions.jl")
 export LowerTriangularMatrices, LowerTriangularMatrix
 include("LowerTriangularMatrices/LowerTriangularMatrices.jl")
 using .LowerTriangularMatrices
+const LTM = LowerTriangularMatrix       # for convenience
 
 # RingGrids
 export RingGrids
@@ -65,13 +66,13 @@ using .SpeedyTransforms
 # Utility for GPU / KernelAbstractions
 include("gpu.jl")                               
 
-# abstract types
+# abstract types
 include("models/abstract_models.jl")
 include("dynamics/abstract_types.jl")
 include("output/abstract_types.jl")
 include("physics/abstract_types.jl")
 
-# GEOMETRY CONSTANTS ETC
+# GEOMETRY CONSTANTS ETC
 include("dynamics/vertical_coordinates.jl")
 include("dynamics/spectral_grid.jl")
 include("dynamics/geometry.jl")
@@ -82,7 +83,7 @@ include("dynamics/adiabatic_conversion.jl")
 include("dynamics/orography.jl")
 include("physics/land_sea_mask.jl")
 
-# VARIABLES
+# VARIABLES
 include("dynamics/variables_3D_4D.jl")
 include("dynamics/particles.jl")
 include("dynamics/clock.jl")
