### 2024-03-19 e9b5078
add! callbacks directly to model too
