### 2024-03-19 a4b05ba
tree with `with_size` using Base.summarysize
