+++ src/models/tree.jl
@@ -5,9 +5,16 @@ $(TYPEDSIGNATURES)
 Create a tree of fields inside a Simulation instance and fields within these fields
 as long as they are defined within the modules argument (default SpeedyWeather).
 Other keyword arguments are `max_level::Integer=10`, `with_types::Bool=false`."""
-function tree(S::Simulation{M}; modules=SpeedyWeather, kwargs...) where M
-    println("Simulation{$(model_type(M))}")
-    _tree(S, modules; kwargs...)
+function tree(
+    S::Simulation{M};
+    modules=SpeedyWeather,
+    with_size::Bool = false,
+    kwargs...
+) where M
+    s = "Simulation{$(model_type(M))}"
+    s = ~with_size ? s : s*" ($(prettymemory(Base.summarysize(S))))"
+    println(s)
+    _tree(S, modules; with_size, kwargs...)
 end
 
 """
@@ -15,9 +22,16 @@ $(TYPEDSIGNATURES)
 Create a tree of fields inside a model and fields within these fields
 as long as they are defined within the modules argument (default SpeedyWeather).
 Other keyword arguments are `max_level::Integer=10`, `with_types::Bool=false`."""
-function tree(M::ModelSetup; modules=SpeedyWeather, kwargs...)
-    println("$(model_type(M)){...}")
-    _tree(M, modules; kwargs...)
+function tree(
+    M::ModelSetup;
+    modules=SpeedyWeather,
+    with_size::Bool = false,
+    kwargs...
+)
+    s = "$(model_type(M)){...}"
+    s = ~with_size ? s : s*" ($(prettymemory(Base.summarysize(S))))"
+    println(s)
+    _tree(M, modules; with_size, kwargs...)
 end
 
 """
@@ -25,39 +39,43 @@ $(TYPEDSIGNATURES)
 Create a tree of fields inside S and fields within these fields
 as long as they are defined within the modules argument (default SpeedyWeather).
 Other keyword arguments are `max_level::Integer=10`, `with_types::Bool=false`."""
-function tree(S; modules=SpeedyWeather, kwargs...)
-    println("$(typeof(S))")
-    _tree(S, modules; kwargs...)
+function tree(S; modules=SpeedyWeather, with_size::Bool = false, kwargs...)
+    s = "$(typeof(S))"
+    s = ~with_size ? s : s*" ($(prettymemory(Base.summarysize(S))))"
+    println(s)
+    _tree(S, modules; with_size, kwargs...)
 end
 
 function _tree(
     S,
     modules::Module...;         # fields within the modules are further inspected
-    max_level::Integer = 10,    # depth of the tree
+    max_level::Integer = 10,    # depth of the tree
     with_types::Bool = false,   # print also the types of the fields?
+    with_size::Bool = false,    # print also the Base.summarysize?
 )
-    level = 0                   # starting level of depth tree
-    prevs = falses(max_level+2) # determine whether there's still a branch levels up
-                                # needed for │ printing of upper levels that continue
+    level = 0                   # starting level of depth tree
+    prevs = falses(max_level+2) # determine whether there's still a branch levels up
+                                # needed for │ printing of upper levels that continue
     property_names = propertynames(S)
     n_properties = length(property_names)
 
     for (i,property_name) in enumerate(property_names)
         last = i == n_properties    # last elements in branches are printed with └ not ├
         prevs[level+1] = ~last
-        print_branch(property_name, S, level, max_level, last, prevs, with_types, modules...)
+        print_branch(property_name, S, level, max_level, last, prevs, with_types, with_size, modules...)
     end
 end
 
 function print_branch(
     property_name::Symbol,  # name of current field
-    S,                      # its parent struct
[…]
