+++ src/physics/ocean.jl
@@ -1,4 +1,4 @@
-abstract type AbstractOcean{NF, Grid} end
+abstract type AbstractOcean end
 
 function Base.show(io::IO, O::AbstractOcean)
     println(io, "$(typeof(O)) <: AbstractOcean")
@@ -6,32 +6,48 @@ function Base.show(io::IO, O::AbstractOcean)
     print_fields(io, O, keys)
 end
 
+# function barrier for all oceans
+function initialize!(   ocean::PrognosticVariablesOcean,
+                        time::DateTime,
+                        model::PrimitiveEquation)
+    initialize!(ocean, time, model.ocean, model)
+end
+
+# function barrier for all oceans
+function ocean_timestep!(   ocean::PrognosticVariablesOcean,
+                            time::DateTime,
+                            model::PrimitiveEquation)
+    ocean_timestep!(ocean, time, model.ocean)
+end
+
+
+## SEASONAL OCEAN CLIMATOLOGY
+
 export SeasonalOceanClimatology
-Base.@kwdef struct SeasonalOceanClimatology{NF, Grid<:AbstractGrid{NF}} <: AbstractOcean{NF, Grid}
+Base.@kwdef struct SeasonalOceanClimatology{NF, Grid<:AbstractGrid{NF}} <: AbstractOcean
 
     "number of latitudes on one hemisphere, Equator included"
     nlat_half::Int
 
-    # OPTIONS
-    "Time step used to update sea surface temperatures"
+    "[OPTION] Time step used to update sea surface temperatures"
     Δt::Dates.Day = Dates.Day(3)
 
-    "path to the folder containing the land-sea mask file, pkg path default"
+    "[OPTION] Path to the folder containing the sea surface temperatures, pkg path default"
     path::String = "SpeedyWeather.jl/input_data"
 
-    "filename of sea surface temperatures"
+    "[OPTION] Filename of sea surface temperatures"
     file::String = "sea_surface_temperature.nc"
 
-    "Variable name in netcdf file"
+    "[OPTION] Variable name in netcdf file"
     varname::String = "sst"
 
-    "Grid the sea surface temperature file comes on"
+    "[OPTION] Grid the sea surface temperature file comes on"
     file_Grid::Type{<:AbstractGrid} = FullGaussianGrid
 
-    "The missing value in the data respresenting land"
+    "[OPTION] The missing value in the data respresenting land"
     missing_value::NF = NF(NaN)
 
-    # to be filled from file
+    # to be filled from file
     "Monthly sea surface temperatures [K], interpolated onto Grid"
     monthly_temperature::Vector{Grid} = [zeros(Grid, nlat_half) for _ in 1:12]
 end
@@ -83,41 +99,150 @@ function load_monthly_climatology!(
     return nothing
 end
 
-function initialize!(   ocean::PrognosticVariablesOcean,
-                        time::DateTime,
-                        model::PrimitiveEquation)
-    ocean_timestep!(ocean, time, model, initialize=true)
+function initialize!(   
+    ocean::PrognosticVariablesOcean,
+    time::DateTime,
+    ocean_model::SeasonalOceanClimatology,
+    model::PrimitiveEquation,
+)
+    ocean.time = time   # set initial time
+    interpolate_monthly!(   ocean.sea_surface_temperature,
+                            ocean_model.monthly_temperature,
+                            time)
 end
-
-# function barrier
+    
 function ocean_timestep!(   ocean::PrognosticVariablesOcean,
                             time::DateTime,
-                            model::PrimitiveEquation;
-                            initialize::Bool = false)
[…]
