+++ test/runtests.jl
@@ -31,6 +31,7 @@ include("geopotential.jl")
 include("column_variables.jl")
 include("orography.jl")
 include("land_sea_mask.jl")
+include("ocean.jl")
 # include("thermodynamics.jl")
 # include("large_scale_condensation.jl")
 # include("convection.jl")
