### 2024-03-19 ed63f7e
Ocean model docstrings added
