### 2024-03-19 1038693
Aquaplanet example model setup
