### 2024-03-19 056796f
Aquaplanet model setup in docs described
