### 2024-03-19 947486b
rename mask inside land_sea_mask simply to mask
