+++ docs/src/land_sea_mask.md
@@ -1,0 +1,175 @@
+# Land-sea mask
+
+The following describes how a custom land-sea mask can be defined.
+SpeedyWeather uses a _fractional_ land-sea mask, i.e. for every grid-point
+
+- 1 indicates land
+- 0 indicates ocean
+- (0,1) indicates a grid-cell partially covered by ocean and land
+
+The land-sea mask determines solely how to weight the surface fluxes
+coming from land or from the ocean. For the sensible heat fluxes this uses
+land and sea surface temperatures and weights the respective fluxes
+proportional to the fractional mask. Similar for evaporation.
+You can therefore define an ocean on top of a mountain, or a land without
+heat fluxes when the land-surface temperature is not defined, i.e. `NaN`.
+Let ``F_L, F_S`` be the fluxes coming from land and sea, respectively.
+Then the land-sea mask ``a \in [0,1]`` weights them as follows for the
+total flux ``F``
+
+```math
+F = aF_L + (1-a)F_S
+```
+
+but ``F=F_L`` if the sea flux is NaN (because the ocean temperature is not defined)
+and ``F=F_S`` if the land flux is NaN (because the land temperature or soil moisture
+is not defined, for sensible heat fluxes or evaporation), and ``F=0`` if both fluxes
+are NaN.
+
+Setting the land-sea mask to ocean therefore will disable any fluxes that
+may come from land, and vice versa. However, with an ocean-everywhere land-sea mask
+you must also define sea surface temperatures everywhere, otherwise the fluxes
+in those regions will be zero.
+
+## Manual land-sea mask
+
+You can create the default land-sea mask as follows
+
+```@example landseamask
+using SpeedyWeather
+spectral_grid = SpectralGrid(trunc=31, nlev=8)
+land_sea_mask = LandSeaMask(spectral_grid)
+```
+
+which will automatically interpolate the land-sea mask onto grid and resolution
+as defined in `spectral_grid` at initialization. The actual mask is in
+`land_sea_mask.land_sea_mask` and you can visualise it with
+
+```@example landseamask
+model = PrimitiveWetModel(;spectral_grid, land_sea_mask)
+simulation = initialize!(model)     # triggers also initialization of model.land_sea_mask
+plot(land_sea_mask.mask)
+```
+
+Now before you run a simulation you could manually change the land-sea mask by
+
+```@example landseamask
+# unpack, this is a flat copy, changing it will also change the mask inside model
+(; mask) = land_sea_mask
+
+# ocean everywhere, or
+mask .= 0    
+
+# random land-sea mask, or
+for i in eachindex(mask)
+    mask[i] = rand()     
+end
+
+# ocean only between 10˚S and 10˚N
+for (j, ring) in enumerate(RingGrids.eachring(mask))
+    for ij in ring
+        mask[ij] = abs(model.geometry.latd[j]) > 10 ? 1 : 0
+    end
+end
+```
+
+And now you can run the simulation as usual with `run!(simulation)`. Most useful
+for the generation of custom land-sea masks in this manual way is probably the
+`model.geometry` component which has all sorts of coordinates like `latd`
+(latitudes in degrees on rings) or `latds, londs` (latitude, longitude in degrees
+for every grid point).
+
+## Earth's land-sea mask
+
+The Earth's [`LandSeaMask`](@ref) has itself the option to load another
+land-sea mask from file, but you also have to specify the grid that mask
+from files comes on. It will then attempt to read it via `NCDatasets`
+and interpolate onto the model grid.
+
+## AquaPlanetMask
+
+Predefined is also the [`AquaPlanetMask`](@ref) which can be created as
+```@example landseamask
+land_sea_mask = AquaPlanetMask(spectral_grid)
+```
[…]
