### 2024-03-19 55bf602
PythonPlot colorbar requires mappable
