+++ docs/src/surface_fluxes.md
@@ -1,3 +1,3 @@
-# Surface fluxes
+# Surface fluxes
 
 More to follow ... 
