### 2024-03-19 c0a46ce
initialize callback schedule for MilleniumFlood
