### 2024-03-20 SpeedyWeather/SpeedyWeather.jl#480 @navidcy
@milankl shall open a PR from a new branch -> #387 to try to address the above?
