### 2024-03-20 eaa5061
Add analysis.md to docs
