+++ docs/src/analysis.md
@@ -1,0 +1,233 @@
+# Analysing a simulation
+
+While you can analyze a SpeedyWeather simulation through its [NetCDF output](@ref),
+as most users will be used to with other models, you can also reuse a lot of
+functionality from SpeedyWeather interactively for analysis. This makes
+SpeedyWeather beyond being a model for the atmospheric general circulation
+also a library with many functions for the analysis of simulations.
+Often this also avoids the two language problem that you will face if you
+run a simulation with a model in one language but then do the data
+analysis in another, treating the model as a blackbox although it likely
+has many of the functions you will need for analysis already defined.
+With SpeedyWeather we try to avoid this and are working towards a 
+more unified approach in atmospheric modelling where simulation
+and analysis are done interactively with the same library: SpeedyWeather.jl.
+
+## Advantages of online analysis
+
+Now you could run a SpeedyWeather simulation, and analyse the [NetCDF output](@ref)
+but that comes with several issues related to accuracy
+
+- If you use a reduced grid for the simulation, then the output will (by default) be
+interpolated on a full grid. This interpolation comes introduces an error.
+- Computing integrals over gridded data on the sphere by weighting every grid point
+according to its area is not the most accurate numerical integration.
+- Computing gradients over gridded data comes with similar issues. While our
+[RingGrids](@ref) are always equidistant in longitude, they are not necessarily
+in latitude.
+
+The first point you can avoid by running a simulation on one of the full grids that
+are implemented, see [SpectralGrid](@ref). But that also impacts the simulation
+and for various reasons we don't run on full grids by default.
+The second point you can address by defining a more advanced numerical integration
+scheme, but that likely requires you to depend on external libraries and then,
+well, you could also just depend on SpeedyWeather.jl directly, because we
+have to do these computations internally anyway. Similar for the third point,
+many gradients have to be computed on every time step and we do that with
+spectral transforms to reduce the discretization error.
+
+The following contains a (hopefully growing) list of examples
+of how a simulation can be analysed interactively. We call this
+_online_ analysis because you are directly using the functionality from
+the model as if it was a library.
+
+## Mass conservation
+
+In the absence of sources and sinks for the interface displacement ``\eta``
+in the [shallow water equations](@ref shallow_water_model), total mass
+(or equivalently volume as density is constant) is conserved.
+The total volume is defined as the integral of the dynamic layer thickness
+``h = \eta + H - H_b`` (``H`` is the layer thickness at rest, ``H_b`` is orography)
+over the surface ``A`` of the sphere
+
+```math
+\iint h dA = \iint \eta dA + \iint H dA - \\int H_b dA
+```
+
+to check for conservation we want to assess that
+
+```math
+\frac{\partial}{\partial t} \\int h dA = 0
+```
+
+And because ``V = \iint H dA - \\int H_b dA``, the total volume at rest,
+is a constant (``H`` is a global constant, the orography ``H_b`` does not change with time)
+we can just check whether ``\\int \eta dA`` changes over time.
+Instead of computing this integral in grid-point space, we use the spectral
+``\eta`` whereby the coefficient of the first spherical harmonic (the ``l = m = 0`` mode,
+see [Spherical Harmonic Transform](@ref) encodes the global average.
+
+```@example analysis
+using SpeedyWeather
+spectral_grid = SpectralGrid(trunc=31, nlev=1)
[…]
