### 2024-03-20 083a869
More formatting issues in analysis.md
