### 2024-03-20 98a6ac5
Convection with work arrays for thread safety
