+++ src/physics/convection.jl
@@ -21,16 +21,10 @@ Base.@kwdef struct SimplifiedBettsMiller{NF} <: AbstractConvection
 
     "[OPTION] Relative humidity for reference profile"
     relative_humidity::NF = 0.7
-
-    "temperature [K] reference profile to adjust to"
-    temp_ref_profile::Vector{NF} = zeros(NF, nlev)
-
-    "specific humidity [kg/kg] profile to adjust to"
-    humid_ref_profile::Vector{NF} = zeros(NF, nlev)
 end
 
 SimplifiedBettsMiller(SG::SpectralGrid; kwargs...) = SimplifiedBettsMiller{SG.NF}(nlev=SG.nlev; kwargs...)
-initialize!(::SimplifiedBettsMiller, ::PrimitiveWet) = nothing
+initialize!(::SimplifiedBettsMiller, ::PrimitiveEquation) = nothing
 
 # function barrier for all AbstractConvection
 function convection!(
@@ -44,7 +38,7 @@ end
 function convection!(
     column::ColumnVariables,
     scheme::SimplifiedBettsMiller,
-    model::PrimitiveEquation,
+    model::PrimitiveWet,
 )
     convection!(column, scheme, model.clausius_clapeyron,
                     model.geometry, model.planet, model.atmosphere, model.time_stepping)
@@ -70,11 +64,14 @@ function convection!(
     σ = geometry.σ_levels_full
     σ_half = geometry.σ_levels_half
     Δσ = geometry.σ_levels_thick
-    (; temp_ref_profile, humid_ref_profile) = SBM
     (; geopot, nlev, temp, temp_virt, humid, temp_tend, humid_tend) = column
     pₛ = column.pres[end]
     (; Lᵥ, cₚ) = clausius_clapeyron
-
+    
+    # use work arrays for temp_ref_profile, humid_ref_profile
+    temp_ref_profile = column.a     # temperature [K] reference profile to adjust to
+    humid_ref_profile = column.b    # specific humidity [kg/kg] profile to adjust to
+    
     # CONVECTIVE CRITERIA AND FIRST GUESS RELAXATION
     temp_parcel = temp[nlev]
     humid_parcel = humid[nlev]
@@ -88,7 +85,7 @@ function convection!(
         humid_ref_profile[k] = qsat*SBM.relative_humidity
     end
 
-    local Pq::NF = 0        # precipitation due to drying
+    local Pq::NF = 0        # precipitation due to drying
     local PT::NF = 0        # precipitation due to cooling
     local ΔT::NF = 0        # vertically uniform temperature profile adjustment
     local Qref::NF = 0      # = ∫_pₛ^p_LZB -humid_ref_profile dp
@@ -111,7 +108,7 @@ function convection!(
     deep_convection = Pq > 0 && PT > 0
     shallow_convection = Pq <= 0 && PT > 0
 
-    # escape immediately for no convection
+    # escape immediately for no convection
     no_convection = !(deep_convection || shallow_convection)
     no_convection && return nothing
 
@@ -123,24 +120,24 @@ function convection!(
         ΔT = (PT - Pq*Lᵥ/cₚ)/Δσ_lzb         # eq (5) but reusing PT, Pq, and /cₚ already included
 
         for k in level_zero_buoyancy:nlev
-            temp_ref_profile[k] -= ΔT       # equation (6)
+            temp_ref_profile[k] -= ΔT       # equation (6)
         end
     
     elseif shallow_convection
         
         # FRIERSON'S QREF SCHEME
         # "changing the reference profiles for both temperature and humidity so the
-        # precipitation is zero.
+        # precipitation is zero.
 
[…]
