+++ test/convection.jl
@@ -8,10 +8,14 @@
         for Model in (  PrimitiveDryModel,
                         PrimitiveWetModel)
 
-            convection = Convection(spectral_grid)
-            model = Model(;spectral_grid, convection)
-            simulation = initialize!(model)
-            run!(simulation, period=Day(5))
+            # that combination is not defined
+            if ~(Convection == SimplifiedBettsMiller && Model == PrimitiveDryModel)
+                convection = Convection(spectral_grid)
+                model = Model(;spectral_grid, convection)
+                model.feedback.verbose = false
+                simulation = initialize!(model)
+                run!(simulation, period=Day(1))
+            end
         end
     end
 end
