### 2024-03-20 c4e7e27
Bulk Richardson-based vertical diffusion
