+++ src/dynamics/hole_filling.jl
@@ -1,10 +1,14 @@
 abstract type AbstractHoleFilling end
 
+export NoHoleFilling
+struct NoHoleFilling <: AbstractHoleFilling end
+NoHoleFilling(SG::SpectralGrid) = NoHoleFilling()
+initialize!(::NoHoleFilling, ::PrimitiveWet) = nothing
+hole_filling!(::AbstractGrid,::NoHoleFilling,::PrimitiveWet) = nothing
+
 export ClipNegatives
 struct ClipNegatives <: AbstractHoleFilling end
 ClipNegatives(SG::SpectralGrid) = ClipNegatives()
-
-# nothing to initialize
 initialize!(::ClipNegatives, ::PrimitiveWet) = nothing
 
 # function barrier
@@ -16,7 +20,7 @@ function hole_filling!(
     hole_filling!(A, H)
 end
 
-function hole_filling!(A::AbstractGrid, ::ClipNegatives)
+function hole_filling!(A::AbstractGrid,::ClipNegatives)
     @inbounds for ij in eachgridpoint(A)
         A[ij] = max(A[ij], 0)
     end
