+++ src/physics/vertical_diffusion.jl
@@ -1,179 +1,118 @@
 abstract type AbstractVerticalDiffusion <: AbstractParameterization end
 
+## FUNCTION BARRIERS for all AbstractVerticalDiffusion
+function vertical_diffusion!(   column::ColumnVariables,
+                                model::PrimitiveEquation)
+    vertical_diffusion!(column, model.vertical_diffusion, model)
+end
+
+## NO VERTICAL DIFFUSION
 export NoVerticalDiffusion
 struct NoVerticalDiffusion <: AbstractVerticalDiffusion end
 NoVerticalDiffusion(SG::SpectralGrid) = NoVerticalDiffusion()
 
-function initialize!(   scheme::NoVerticalDiffusion,
-                        model::PrimitiveEquation)
+# define dummy functions
+initialize!(::NoVerticalDiffusion,::PrimitiveEquation) = nothing
+vertical_diffusion!(::ColumnVariables, ::NoVerticalDiffusion, ::PrimitiveEquation) = nothing
+
+
+Base.@kwdef struct BulkRichardsonDiffusion{NF} <: AbstractVerticalDiffusion
+    nlev::Int
+
+    "[OPTION] von Kármán constant [1]"
+    κ::NF = 0.4
+
+    "[OPTION] roughness length [m]"
+    z₀::NF = 3.21e-5
+
+    "[OPTION] Critical Richardson number for stable mixing cutoff [1]"
+    Ri_c::NF = 1
+
+    "[OPTION] Surface layer fraction"
+    fb::NF = 0.1
+
+    "[OPTION] diffuse static energy?"
+    diffuse_static_energy::Bool = true
+
+    "[OPTION] diffuse momentum?"
+    diffuse_momentum::Bool = true
+
+    "[OPTION] diffuse humidity? Ignored for PrimitiveDryModels"
+    diffuse_humidity::Bool = true
+
+    # precomputed operators
+    ∇²_above::Vector{NF} = zeros(NF, nlev-1)
+    ∇²_below::Vector{NF} = zeros(NF, nlev-1)
+    Δσ::Vector{NF} = zeros(NF, nlev-1)
+end
+
+BulkRichardsonDiffusion(SG::SpectralGrid; kwargs...) = BulkRichardsonDiffusion{SG.NF}(; nlev=SG.nlev, kwargs...)
+
+function initialize!(scheme::BulkRichardsonDiffusion, model::PrimitiveEquation)
+    Δσ = model.geometry.σ_levels_thick
+    
+    # 1. thickness Δσ of half levels
+    # 2. 1/Δσ² but for variable Δσ on half levels = 1/(1/2*Δσₖ(Δσ_k-1 + Δσₖ))
+    # 3. as 2. but for half levels below = 1/(1/2*Δσₖ(Δσ_k+1 + Δσₖ))
+    @. scheme.Δσ = 1/2*(Δσ[2:end] + Δσ[1:end-1])
+    @. scheme.∇²_above = inv(Δσ[2:end]*scheme.Δσ)
+    @. scheme.∇²_below = inv(Δσ[1:end-1]*scheme.Δσ)
+
     return nothing
-end 
-
-# function barrier
-function static_energy_diffusion!(  column::ColumnVariables,
-                                    model::PrimitiveEquation)
-    static_energy_diffusion!(column, model.static_energy_diffusion)
 end
 
-function static_energy_diffusion!(  column::ColumnVariables,
-                                    scheme::NoVerticalDiffusion)
+function vertical_diffusion!(
+    column::ColumnVariables,
+    scheme::BulkRichardsonDiffusion,
+    model::PrimitiveEquation)
+
+    (; diffuse_momentum, diffuse_static_energy, diffuse_humidity) = scheme
[…]
