### 2024-03-21 8092643
Vertical diffusion operator streamlined
