+++ src/physics/vertical_diffusion.jl
@@ -40,23 +40,46 @@ Base.@kwdef struct BulkRichardsonDiffusion{NF} <: AbstractVerticalDiffusion
     "[OPTION] diffuse humidity? Ignored for PrimitiveDryModels"
     diffuse_humidity::Bool = true
 
+    # precomputed constants
+    κZsqrtC_g::Base.RefValue{NF} = Ref(zero(NF))
+
     # precomputed operators
-    ∇²_above::Vector{NF} = zeros(NF, nlev-1)
-    ∇²_below::Vector{NF} = zeros(NF, nlev-1)
-    Δσ::Vector{NF} = zeros(NF, nlev-1)
+    ∇²_above::Vector{NF} = zeros(NF, nlev)
+    ∇²_below::Vector{NF} = zeros(NF, nlev)
 end
 
 BulkRichardsonDiffusion(SG::SpectralGrid; kwargs...) = BulkRichardsonDiffusion{SG.NF}(; nlev=SG.nlev, kwargs...)
 
 function initialize!(scheme::BulkRichardsonDiffusion, model::PrimitiveEquation)
-    Δσ = model.geometry.σ_levels_thick
     
-    # 1. thickness Δσ of half levels
-    # 2. 1/Δσ² but for variable Δσ on half levels = 1/(1/2*Δσₖ(Δσ_k-1 + Δσₖ))
-    # 3. as 2. but for half levels below = 1/(1/2*Δσₖ(Δσ_k+1 + Δσₖ))
-    @. scheme.Δσ = 1/2*(Δσ[2:end] + Δσ[1:end-1])
-    @. scheme.∇²_above = inv(Δσ[2:end]*scheme.Δσ)
-    @. scheme.∇²_below = inv(Δσ[1:end-1]*scheme.Δσ)
+    (; nlev) = model.geometry
+    nlev == 1 && return nothing     # no diffusion for 1-layer model
+
+    # ∇² operator on σ levels like 1/Δσ² but for variable Δσ
+    # also includes a 1/2 so that the diffusion coefficients on full levels can be added
+    # which is equivalent to interpolating them on half levels for a ∂σ (K ∂σ) formulation
+    # with σ-dependent diffusion coefficient K
+    σ = model.geometry.σ_levels_full
+    σ_half = model.geometry.σ_levels_half
+
+    for k in 1:nlev
+        k₋ = max(k, 1)      # sets the gradient across surface and top to 0
+        k₊ = min(k, nlev)   # = no flux boundary conditions
+        scheme.∇²_above[k] = inv(2*(σ[k] - σ[k₋]) * (σ_half[k+1] - σ_half[k]))
+        scheme.∇²_below[k] = inv(2*(σ[k₊] - σ[k]) * (σ_half[k+1] - σ_half[k]))
+    end
+
+    # Typical height Z of lowermost layer from geopotential of reference surface temperature
+    # minus surface geopotential (orography * gravity)
+    (; temp_ref) = model.atmosphere
+    (; gravity) = model.planet
+    (; Δp_geopot_full) = model.geopotential
+    Z = temp_ref * Δp_geopot_full[end] / gravity
+    
+    # maximum drag Cmax from that height, stable conditions would decrease Cmax towards 0
+    # Frierson 2006, eq (12)
+    (; κ, z₀) = scheme
+    scheme.κZsqrtC_g[] = Z*κ^2/(log(Z/z₀)*gravity)
 
     return nothing
 end
@@ -67,52 +90,102 @@ function vertical_diffusion!(
     model::PrimitiveEquation)
 
     (; diffuse_momentum, diffuse_static_energy, diffuse_humidity) = scheme
[…]
