### 2024-03-21 ee4b065
static energy diffusion divided by heat capacity
