+++ src/LowerTriangularMatrices/lower_triangular_matrix.jl
@@ -1,47 +1,54 @@
+import LinearAlgebra: tril!
+
 """
-    L = LowerTriangularMatrix{T}(v::Vector{T}, m::Int, n::Int)
+    L = LowerTriangularArray{T, N, ArrayType}(v::AbstractArray{T,N}, m::Int, n::Int)
 
 A lower triangular matrix implementation that only stores the non-zero entries explicitly.
 `L<:AbstractMatrix` although in general we have `L[i] != Matrix(L)[i]`, the former skips
 zero entries, tha latter includes them."""
-struct LowerTriangularMatrix{T} <: AbstractMatrix{T}
-    data::Vector{T}     # non-zero elements unravelled into a vector
+struct LowerTriangularArray{T, N, ArrayType <: AbstractArray{T}} <: AbstractArray{T,N}
+    data::ArrayType     # non-zero elements unravelled into an array with one dimension less N-1
     m::Int              # number of rows
     n::Int              # number of columns
 
-    LowerTriangularMatrix{T}(data, m, n) where T = length(data) == nonzeros(m, n) ?
-        new(data, m, n) : error("$(length(data))-element Vector{$(eltype(data))} cannot be used to create a "*
-            "$(m)x$(n) LowerTriangularMatrix{$T} with $(nonzeros(m, n)) non-zero entries.")
+    LowerTriangularArray{T, N, ArrayType}(data, m, n) where {T, N, ArrayType<:AbstractArray} = length(data) == prod(size(data)[2:end]) * nonzeros(m, n) ? new(data, m, n) : error("$(length(data))-element Vector{$(eltype(data))} cannot be used to create a $(m)x$(n) LowerTriangularArray{$T,$N,$ArrayType} with $(nonzeros(m, n)) non-zero entries.")  #TODO: adjust error string
 end
 
-LowerTriangularMatrix(data::AbstractVector{T}, m::Integer, n::Integer) where T = LowerTriangularMatrix{T}(data, m, n)
+const LowerTriangularMatrix{T, ArrayType} = LowerTriangularArray{T, 2, ArrayType}
+
+LowerTriangularArray(data::ArrayType, m::Integer, n::Integer) where {T, N, ArrayType<:AbstractArray{T,N}} = LowerTriangularArray{T, N+1, ArrayType}(data, m, n)
+LowerTriangularMatrix(data::ArrayType, m::Integer, n::Integer) where {T, ArrayType<:AbstractVector{T}} = LowerTriangularMatrix{T, ArrayType}(data, m, n)
 
 # SIZE ETC
-Base.length(L::LowerTriangularMatrix) = length(L.data)  # define length as number of non-zero elements
-Base.size(L::LowerTriangularMatrix) = (L.m, L.n)         # define size as matrix size
-Base.sizeof(L::LowerTriangularMatrix) = sizeof(L.data)  # sizeof the underlying data vector
+Base.length(L::LowerTriangularArray) = length(L.data)  # define length as number of non-zero elements
+Base.size(L::LowerTriangularArray) = (L.m, L.n, size(L.data)[2:end]...)  # define size as matrix size (doesnt give an error, even for LowerTriangularMatrix due to tuple indexing logic)
+Base.sizeof(L::LowerTriangularArray) = sizeof(L.data)  # sizeof the underlying data vector
 
 # CREATE INSTANCES (ZEROS, ONES, UNDEF)
 for zeros_or_ones in (:zeros, :ones)
     @eval begin
-        function Base.$zeros_or_ones(::Type{LowerTriangularMatrix{T}}, m::Integer, n::Integer) where T
[…]
