+++ test/lower_triangular_matrix.jl
@@ -1,8 +1,11 @@
+import JLArrays
+
 @testset "LowerTriangularMatrix" begin
     @testset for NF in (Float32, Float64)
         mmax = 32
         @testset for lmax = (mmax, mmax+1)
             A = randn(Complex{NF}, lmax, mmax)
+
             SpeedyWeather.spectral_truncation!(A)
 
             L = SpeedyWeather.LowerTriangularMatrix(A)
@@ -25,6 +28,41 @@
         end
     end
 end
+
+@testset "LowerTriangularArray: N-dim" begin 
+    @testset for NF in (Float32, Float64)
+        mmax = 32
+        @testset for idims = ((), (5,), (5,5))
+            @testset for lmax = (mmax, mmax+1)
+                A = randn(Complex{NF}, lmax, mmax, idims...)
+
+                # replaces spectraltrunction! here, we just set the elements zero manually
+                ind = SpeedyWeather.LowerTriangularMatrices.lowertriangle_indices(A)
+                ind = @. ~(ind)
+                A[ind] .= zero(Complex{NF})
+
+                L = SpeedyWeather.LowerTriangularArray(A)
+
+                @test size(L) == size(A)
+
+                for m in 1:mmax
+                    for l in 1:lmax
+                        @test A[l, m, [Colon() for i=1:length(idims)]...] == L[l, m, [Colon() for i=1:length(idims)]...]
+                    end
+                end
+
+                @test_throws BoundsError L[lmax+1, mmax, [1 for i=1:length(idims)]...]
+                @test_throws BoundsError L[lmax, mmax+1, [1 for i=1:length(idims)]...]
+                
+                @test_throws BoundsError L[1, 2, [1 for i=1:length(idims)]...] = 1
+                @test_throws BoundsError L[lmax*mmax+1, [1 for i=1:length(idims)]...] = 1
+
+                @test Array(L) == A
+            end
+        end
+    end
+
+end 
 
 @testset "LowerTriangularMatrix: @inbounds" begin
     A = randn(LowerTriangularMatrix, 33, 32)
@@ -56,34 +94,91 @@ end
     end
 end
 
+@testset "LowerTriangularArray: @inbounds" begin
+    A = randn(LowerTriangularArray, 33, 32, 1, 1)
+    
+    @testset "getindex" begin
+        @test_throws BoundsError A[34, 32, 1, 1]   # outside of i, j range
+        @test_throws BoundsError A[561, 1, 1]     # outside of k range
+        @test_throws BoundsError A[33, 32, 2, 1]     # outside of 3rd dim range
+        @test_throws BoundsError A[33, 32, 1, 2]     # outside of 4th dim range
+    end
+
+    @testset "setindex!" begin
+        @test_throws BoundsError A[34, 32, 1, 1] = 1
+        @test_throws BoundsError A[561, 1, 1] = 1
+        @test_throws BoundsError A[33, 32, 2, 1] = 1 
+        @test_throws BoundsError A[33, 32, 1, 2] = 1 
+    end
+end
+
+@testset "LowerTriangularArray: fill, copy, randn, convert" begin
+    @testset for NF in (Float32, Float64)
+        mmax = 32
+        @testset for idims = ((), (5,), (5,5))
+            @testset for lmax = (mmax, mmax+1)
+                A = randn(Complex{NF}, lmax, mmax, idims...)
+
+                # replaces spectraltrunction! here, we just set the elements zero manually
+                ind = SpeedyWeather.LowerTriangularMatrices.lowertriangle_indices(A)
+                ind = @. ~(ind)
+                A[ind] .= zero(Complex{NF})
+
[…]
