### 2024-03-22 e17cafb
Formulas as in Frierson, no shortcuts
