+++ src/physics/vertical_diffusion.jl
@@ -28,6 +28,9 @@ Base.@kwdef struct BulkRichardsonDiffusion{NF} <: AbstractVerticalDiffusion
     "[OPTION] Critical Richardson number for stable mixing cutoff [1]"
     Ri_c::NF = 1
 
+    "[OPTION] Fraction of surface boundary layer"
+    fb::NF = 0.1
+
     "[OPTION] diffuse static energy?"
     diffuse_static_energy::Bool = true
 
@@ -38,7 +41,7 @@ Base.@kwdef struct BulkRichardsonDiffusion{NF} <: AbstractVerticalDiffusion
     diffuse_humidity::Bool = true
 
     # precomputed constants
-    κZsqrtC_g::Base.RefValue{NF} = Ref(zero(NF))
+    sqrtC_max::Base.RefValue{NF} = Ref(zero(NF))
 
     # precomputed operators
     ∇²_above::Vector{NF} = zeros(NF, nlev)
@@ -67,7 +70,9 @@ function initialize!(scheme::BulkRichardsonDiffusion, model::PrimitiveEquation)
     end
 
     # Typical height Z of lowermost layer from geopotential of reference surface temperature
-    # minus surface geopotential (orography * gravity)
+    # minus surface geopotential (orography * gravity), simplification compared to
+    # Frierson to reduce the number of expensive log calls given that z ≈ Z for most
+    # surface temperature variations
     (; temp_ref) = model.atmosphere
     (; gravity) = model.planet
     (; Δp_geopot_full) = model.geopotential
@@ -76,7 +81,7 @@ function initialize!(scheme::BulkRichardsonDiffusion, model::PrimitiveEquation)
     # maximum drag Cmax from that height, stable conditions would decrease Cmax towards 0
     # Frierson 2006, eq (12)
     (; κ, z₀) = scheme
-    scheme.κZsqrtC_g[] = Z*κ^2/(log(Z/z₀)*gravity)
+    scheme.sqrtC_max[] = κ/log(Z/z₀)
 
     return nothing
 end
@@ -91,15 +96,14 @@ function vertical_diffusion!(
     # escape immediately if all diffusions disabled
     any((diffuse_momentum, diffuse_static_energy, diffuse_humidity)) || return nothing
     
-    K, h = get_diffusion_coefficients!(column, scheme, model.atmosphere)
+    K, kₕ = get_diffusion_coefficients!(column, scheme, model.atmosphere, model.planet)
 
     (; u, u_tend, v, v_tend, dry_static_energy, temp_tend, humid, humid_tend) = column
-    diffuse_momentum                            && vertical_diffusion!(u_tend, u, K, h, scheme)
-    diffuse_momentum                            && vertical_diffusion!(v_tend, v, K, h, scheme)
-    model isa PrimitiveWet && diffuse_humidity  && vertical_diffusion!(humid_tend, humid, K, h, scheme)
-    
-    K ./= model.atmosphere.heat_capacity
-    diffuse_static_energy                       && vertical_diffusion!(temp_tend, dry_static_energy, K, h, scheme)
+    diffuse_momentum                            && vertical_diffusion!(u_tend, u, K, kₕ, scheme)
+    diffuse_momentum                            && vertical_diffusion!(v_tend, v, K, kₕ, scheme)
+    model isa PrimitiveWet && diffuse_humidity  && vertical_diffusion!(humid_tend, humid, K, kₕ, scheme)
+    K ./= model.atmosphere.heat_capacity        # put dry static energy => temperature conversion into K
[…]
