### 2024-03-22 8c8a5f6
height z at least roughness length to avoid DomainError with log
