### 2024-03-22 a197408
new getindex/setindex, some tests fixed
