### 2024-03-23 7c4c466
boundary layer height always positive to avoid log DomainError
