### 2024-03-23 36eb37a
Precomputed log(Z/z₀)
