### 2024-03-24 79554a9
Jeevanjee longwave radiation
