+++ src/physics/longwave_radiation.jl
@@ -4,13 +4,14 @@ export NoLongwave
 struct NoLongwave <: AbstractLongwave end
 NoLongwave(SG::SpectralGrid) = NoLongwave()
 initialize!(::NoLongwave, ::PrimitiveEquation) = nothing
+longwave_radiation!(::ColumnVariables, ::NoLongwave, ::PrimitiveEquation) = nothing
 
 # function barrier for all AbstractLongwave
 function longwave_radiation!(column::ColumnVariables, model::PrimitiveEquation)
     longwave_radiation!(column, model.longwave_radiation, model)
 end
 
-longwave_radiation!(::ColumnVariables, ::NoLongwave, ::PrimitiveEquation) = nothing
+## UNIFORM COOLING
 
 export UniformCooling
 Base.@kwdef struct UniformCooling{NF} <: AbstractLongwave
@@ -44,4 +45,67 @@ function longwave_radiation!(
     @inbounds for k in eachlayer(column)
         temp_tend[k] += temp[k] > temp_min ? cooling : (temp_stratosphere - temp[k])*τ⁻¹
     end
+end
+
+## JEEVANJEE TEMPERATURE FLUX
+
+export JeevanjeeRadiation
+"""
+Temperature flux longwave radiation from Jeevanjee and Romps, 2018,
+following Seeley and Wordsworth, 2023, eq (1)
+
+    dF/dT = α*(T_t - T)
+
+with `F` the upward temperature flux between two layers with temperature
+difference `dT`, `α = 0.025 W/m²/K²` and `T_t = 200K` a prescribed tropopause
+temperature. Flux into the lowermost layer is 0. Flux out of uppermost
+layer also 0, but `dT/dt = (T_t - T)/τ` is added to relax the uppermost
+layer towards the tropopause temperature `T_t` with time scale `τ = 24h`
+(Seeley and Wordsworth, 2023 use 6h, which is unstable a low resolutions here).
+Fields are
+$(TYPEDFIELDS)"""
+Base.@kwdef struct JeevanjeeRadiation{NF} <: AbstractLongwave
+    "Radiative forcing constant (W/m²/K²)"
+    α::NF = 0.025
+
+    "Tropopause temperature [K]"
+    temp_tropopause::NF = 200
+
+    "Tropopause relaxation time scale to temp_tropopause"
+    time_scale::Second = Hour(24)
+end
+
+JeevanjeeRadiation(SG::SpectralGrid; kwargs...) = JeevanjeeRadiation{SG.NF}(; kwargs...)
+initialize!(scheme::JeevanjeeRadiation, model::PrimitiveEquation) = nothing
+
+# function barrier
+function longwave_radiation!(
+    column::ColumnVariables,
+    scheme::JeevanjeeRadiation,
+    model::PrimitiveEquation,
+)
+    longwave_radiation!(column, scheme)
+end
+
+function longwave_radiation!(
+    column::ColumnVariables{NF},
+    scheme::JeevanjeeRadiation,
+) where NF
+
+    (; nlev, temp_tend) = column
+    T = column.temp                 # to match Seeley, 2023 notation
+    F = column.flux_temp_upward
+    (; α, time_scale) = scheme
+    Tₜ = scheme.temp_tropopause
+    
+    Fₖ::NF = 0                      # flux into lowermost layer from below
+
+    @inbounds for k in nlev-1:-1:1
+        # Seeley and Wordsworth, 2023 eq (1)
+        Fₖ = Fₖ + (T[k] - T[k+1])*α*(Tₜ - T[k+1])   # upward flux from layer k+1 into k
+        F[k+1] += Fₖ                                # flux vector uses k+1 to store, accumulate
+    end
+
[…]
