### 2024-03-24 d1158b8
Move cos_zenith to diagnostic variables
