+++ src/physics/column_variables.jl
@@ -55,6 +55,7 @@ function get_column!(   C::ColumnVariables,
     C.skin_temperature_sea = P.ocean.sea_surface_temperature[ij]
     C.skin_temperature_land = P.land.land_surface_temperature[ij]
     C.soil_moisture_availability = D.surface.soil_moisture_availability[ij]
+    C.cos_zenith = D.surface.cos_zenith[ij]
 end
 
 """Recalculate ring index if not provided."""
