+++ src/physics/zenith.jl
@@ -1,6 +1,6 @@
 abstract type AbstractSolarDeclination end
 abstract type AbstractSolarTimeCorrection end
-abstract type AbstractZenith{NF, Grid} end
+abstract type AbstractZenith end
 
 """Coefficients to calculate the solar declination angle δ [radians] based on a simple
 sine function, with Earth's axial tilt as amplitude, equinox as phase shift.
@@ -109,24 +109,25 @@ given the parameters in model.planet. In both cases the seasonal cycle can
 be disabled, calculating the solar declination from the initial time
 instead of current time."""
 function WhichZenith(SG::SpectralGrid, P::AbstractPlanet; kwargs...)
-    (; NF, Grid, nlat_half) = SG
+    (; NF) = SG
     (; daily_cycle, seasonal_cycle, length_of_day, length_of_year) = P
     solar_declination = SinSolarDeclination(SG, P)
 
     if daily_cycle
-        return SolarZenith{NF, Grid{NF}}(;
-            nlat_half, length_of_day, length_of_year, solar_declination, seasonal_cycle, kwargs...)
+        return SolarZenith{NF}(;
+            length_of_day, length_of_year, solar_declination, seasonal_cycle, kwargs...)
 
     else
-        return SolarZenithSeason{NF, Grid{NF}}(;
-            nlat_half, length_of_day, length_of_year, solar_declination, seasonal_cycle, kwargs...)
+        return SolarZenithSeason{NF}(;
+            length_of_day, length_of_year, solar_declination, seasonal_cycle, kwargs...)
     end
 end
 
 # function barrier
-function cos_zenith!(time::DateTime, model::PrimitiveEquation)
+function cos_zenith!(diagn::DiagnosticVariables, time::DateTime, model::PrimitiveEquation)
     (; solar_zenith, geometry) = model
-    cos_zenith!(solar_zenith, time, geometry)
+    (; cos_zenith) = diagn.surface
+    cos_zenith!(cos_zenith, solar_zenith, time, geometry)
 end
 
 function Base.show(io::IO, L::AbstractZenith)
@@ -139,10 +140,7 @@ export SolarZenith
 
 """Solar zenith angle varying with daily and seasonal cycle.
 $(TYPEDFIELDS)"""
-Base.@kwdef struct SolarZenith{NF<:AbstractFloat, Grid<:AbstractGrid{NF}} <: AbstractZenith{NF, Grid}
-    # DIMENSIONS
-    nlat_half::Int
-
+Base.@kwdef struct SolarZenith{NF<:AbstractFloat} <: AbstractZenith
     # OPTIONS
     length_of_day::Second = Hour(24)
     length_of_year::Second = Day(365.25)
@@ -153,9 +151,7 @@ Base.@kwdef struct SolarZenith{NF<:AbstractFloat, Grid<:AbstractGrid{NF}} <: Abs
     solar_declination::SinSolarDeclination{NF} = SinSolarDeclination{NF}()
     time_correction::SolarTimeCorrection{NF} = SolarTimeCorrection{NF}()
 
-    # WORK ARRAYS
     initial_time::Base.RefValue{DateTime} = Ref(DEFAULT_DATE)
-    cos_zenith::Grid = zeros(Grid, nlat_half)
 end
 
 function initialize!(
@@ -164,7 +160,6 @@ function initialize!(
     model::ModelSetup
 )
     S.initial_time[] = initial_time     # to fix the season if no seasonal cycle
-    cos_zenith!(S, initial_time, model.geometry)
 end
 
 """
[…]
