### 2024-03-25 8e07518
reduce solar constant temporarily
