### 2024-03-25 5e56070
Implement AlbedoClimatology
