### 2024-03-25 346f691
TransparentShortwave with spatially varying albedo
