### 2024-03-25 f5ce4b6
eachharmonic non-allocating, copyto! fixed(?)
