### 2024-03-25 17c898b
Radiation also default in PrimitiveDry
