+++ src/LowerTriangularMatrices/lower_triangular_matrix.jl
@@ -1,11 +1,16 @@
 import LinearAlgebra: tril!
+import GPUArrays
 
 """
-    L = LowerTriangularArray{T, N, ArrayType}(v::AbstractArray{T,N}, m::Int, n::Int)
-
-A lower triangular matrix implementation that only stores the non-zero entries explicitly.
+$(TYPEDSIGNATURES)
+A lower triangular array implementation that only stores the non-zero entries explicitly.
 `L<:AbstractMatrix` although in general we have `L[i] != Matrix(L)[i]`, the former skips
-zero entries, tha latter includes them."""
+zero entries, tha latter includes them. 
+
+Supports n-dimensional lower triangular arrays, so that for all trailing dimensions `L[:,:,..]`
+is a matrix in lower triangular form, e.g. a (5x5x3)-LowerTriangularArray would hold 3 lower
+triangular matrices.
+"""
 struct LowerTriangularArray{T, N, ArrayType <: AbstractArray{T}} <: AbstractArray{T,N}
     data::ArrayType     # non-zero elements unravelled into an array with one dimension less N-1
     m::Int              # number of rows
@@ -71,10 +76,7 @@ Base.rand(::Type{LowerTriangularMatrix{T}}, m::Integer, n::Integer) where T = Lo
 
 # INDEXING
 """
-    k = ij2k(   i::Integer,     # row index of matrix
-                j::Integer,     # column index of matrix
-                m::Integer)     # number of rows in matrix
-
+$(TYPEDSIGNATURES)
 Converts the index pair `i, j` of an `m`x`n` LowerTriangularMatrix `L` to a single
 index `k` that indexes the same element in the corresponding vector that stores
 only the lower triangle (the non-zero entries) of `L`."""
@@ -117,6 +119,7 @@ end
 @inline Base.getindex(L::LowerTriangularArray, r::AbstractRange) = getindex(L.data,r)
 @inline Base.getindex(L::LowerTriangularArray, r::AbstractRange, I...) = getindex(L.data, r, I...)
 @inline Base.getindex(L::LowerTriangularArray, i::Integer) = getindex(L.data,i)
+@inline Base.getindex(L::LowerTriangularArray, I::CartesianIndex) = getindex(L, Tuple(I)...)
 
 Base.@propagate_inbounds function Base.setindex!(L::LowerTriangularArray{T,N}, x, I::Vararg{Any, M}) where {T,N,M} 
     @boundscheck M == N-1 || throw(BoundsError(L, I))
@@ -130,22 +133,22 @@ Base.@propagate_inbounds function Base.setindex!(L::LowerTriangularArray{T,N}, x
     setindex!(L.data, x, k, I[3:end]...)
 end
 
-@inline Base.setindex!(L::LowerTriangularMatrix, x::AbstractVector, r::AbstractRange, I...) = setindex!(L.data, x, r, I...)
+@inline Base.setindex!(L::LowerTriangularArray, x::AbstractVector, r::AbstractRange, I...) = setindex!(L.data, x, r, I...)
+@inline Base.setindex!(L::LowerTriangularArray, x, i::Integer) = setindex!(L.data, x, i)
+@inline Base.setindex!(L::LowerTriangularArray, x, I::CartesianIndex) = setindex!(L, x, Tuple(I)...)
 
 # propagate index to data vector
 Base.eachindex(L ::LowerTriangularArray)    = eachindex(L.data)
 Base.eachindex(Ls::LowerTriangularArray...) = eachindex((L.data for L in Ls)...)
 
 """
[…]
