### 2024-03-27 def116b
broadcast for cpu n-dim working
