### 2024-03-27 d7bc572
add k2ij function and turn tests back on
