+++ versions.js
@@ -1,0 +1,12 @@
+var DOC_VERSIONS = [
+  "stable",
+  "v0.9.0",
+  "v0.8.0",
+  "v0.7.1",
+  "v0.7.0",
+  "v0.6.0",
+  "v0.5.0",
+  "dev",
+];
+var DOCUMENTER_NEWEST = "v0.9.0";
+var DOCUMENTER_STABLE = "stable";
