+++ previews/PR387/run_0009/progress.txt
@@ -1,0 +1,33 @@
+Starting SpeedyWeather.jl run 0009 on Wed, 20 Mar 2024 14:48:56
+Integrating:
+SpectralGrid:
+├ Spectral:   T31 LowerTriangularMatrix{Complex{Float32}}, radius = 6.371e6 m
+├ Grid:       48-ring OctahedralGaussianGrid{Float32}, 3168 grid points
+├ Resolution: 401km (average)
+└ Vertical:   5-level SigmaCoordinates
+Time: 50.0 days at Δt = 1800.0s
+
+All data will be stored in /home/runner/work/SpeedyWeather.jl/SpeedyWeather.jl/docs/build/run_0009
+
+  5%, ETA: 0:00:12,   Inf millenia/day
+ 10%, ETA: 0:00:11,   Inf millenia/day
+ 15%, ETA: 0:00:11,   Inf millenia/day
+ 20%, ETA: 0:00:10,   Inf millenia/day
+ 25%, ETA: 0:00:10,   Inf millenia/day
+ 30%, ETA: 0:00:09,   Inf millenia/day
+ 35%, ETA: 0:00:09,   Inf millenia/day
+ 40%, ETA: 0:00:08,   Inf millenia/day
+ 45%, ETA: 0:00:07,   Inf millenia/day
+ 50%, ETA: 0:00:07,   Inf millenia/day
+ 55%, ETA: 0:00:06,   Inf millenia/day
+ 60%, ETA: 0:00:05,   Inf millenia/day
+ 65%, ETA: 0:00:05,   Inf millenia/day
+ 70%, ETA: 0:00:04,   Inf millenia/day
+ 75%, ETA: 0:00:03,   Inf millenia/day
+ 80%, ETA: 0:00:03,   Inf millenia/day
+ 85%, ETA: 0:00:02,   Inf millenia/day
+ 90%, ETA: 0:00:01,   Inf millenia/day
+ 95%, ETA: 0:00:01,   Inf millenia/day
+100%, ETA: 0:00:00,   Inf millenia/day
+
+Integration done in empty period.
