+++ previews/PR483/run_0008/parameters.txt
@@ -1,0 +1,197 @@
+SpectralGrid:
+├ Spectral:   T31 LowerTriangularMatrix{Complex{Float32}}, radius = 6.371e6 m
+├ Grid:       64-ring FullGaussianGrid{Float32}, 8192 grid points
+├ Resolution: 250km (average)
+└ Vertical:   8-level SigmaCoordinates
+Earth{Float32} <: SpeedyWeather.AbstractPlanet
+├ rotation::Float32 = 7.29e-5
+├ gravity::Float32 = 9.81
+├ daily_cycle::Bool = true
+├ length_of_day::Second = 86400 seconds
+├ seasonal_cycle::Bool = true
+├ length_of_year::Second = 31557600 seconds
+├ equinox::DateTime = 2000-03-20T00:00:00
+├ axial_tilt::Float32 = 23.4
+└ solar_constant::Float32 = 341.25
+EarthAtmosphere{Float32} <: SpeedyWeather.AbstractAtmosphere
+├ mol_mass_dry_air::Float32 = 28.9649
+├ mol_mass_vapour::Float32 = 18.0153
+├ heat_capacity::Float32 = 1004.0
+├ R_gas::Float32 = 8.3145
+├ R_dry::Float32 = 287.05432
+├ R_vapour::Float32 = 461.52438
+├ mol_ratio::Float32 = 0.62197006
+├ μ_virt_temp::Float32 = 0.60779446
+├ κ::Float32 = 0.2859107
+├ water_density::Float32 = 1000.0
+├ latent_heat_condensation::Float32 = 2.501e6
+├ latent_heat_sublimation::Float32 = 2.801e6
+├ stefan_boltzmann::Float32 = 5.67e-8
+├ pres_ref::Float32 = 100000.0
+├ temp_ref::Float32 = 288.0
+├ moist_lapse_rate::Float32 = 0.005
+├ dry_lapse_rate::Float32 = 0.0098
+└ layer_thickness::Float32 = 8500.0
+Leapfrog{Float32} <: SpeedyWeather.AbstractTimeStepper
+├ trunc::Int64 = 31
+├ Δt_at_T31::Second = 1800 seconds
+├ radius::Float32 = 6.371e6
+├ adjust_with_output::Bool = true
+├ robert_filter::Float32 = 0.05
+├ williams_filter::Float32 = 0.53
+├ Δt_millisec::Dates.Millisecond = 1800000 milliseconds
+├ Δt_sec::Float32 = 1800.0
+└ Δt::Float32 = 0.0002825302
+OutputWriter{Float32, PrimitiveDry}
+├ spectral_grid::SpectralGrid = SpectralGrid:
+├ Spectral:   T31 LowerTriangularMatrix{Complex{Float32}}, radius = 6.371e6 m
+├ Grid:       64-ring FullGaussianGrid{Float32}, 8192 grid points
+├ Resolution: 250km (average)
+└ Vertical:   8-level SigmaCoordinates
+├ output::Bool = true
+├ path::String = /home/runner/work/SpeedyWeather.jl/SpeedyWeather.jl/docs/build
+├ id::String = 0008
+├ run_path::String = /home/runner/work/SpeedyWeather.jl/SpeedyWeather.jl/docs/build/run_0008
+├ filename::String = output.nc
+├ write_restart::Bool = true
+├ pkg_version::VersionNumber = 0.9.0
+├ startdate::DateTime = 2000-01-01T00:00:00
+├ output_dt::Second = 21600 seconds
+├ missing_value::Float32 = NaN
+├ compression_level::Int64 = 3
+├ shuffle::Bool = true
[…]
