+++ previews/PR483/run_0003/progress.txt
@@ -1,33 +1,0 @@
-Starting SpeedyWeather.jl run 0003 on Mon, 11 Mar 2024 18:10:29
-Integrating:
-SpectralGrid:
-├ Spectral:   T63 LowerTriangularMatrix{Complex{Float32}}, radius = 6.371e6 m
-├ Grid:       96-ring OctahedralGaussianGrid{Float32}, 10944 grid points
-├ Resolution: 216km (average)
-└ Vertical:   1-level SigmaCoordinates
-Time: 6.0 days at Δt = 900.0s
-
-All data will be stored in /home/runner/work/SpeedyWeather.jl/SpeedyWeather.jl/docs/build/run_0003
-
-  5%, ETA: 0:00:01,   Inf millenia/day
- 10%, ETA: 0:00:01,   Inf millenia/day
- 15%, ETA: 0:00:01, 1834.16 years/day
- 20%, ETA: 0:00:01,  2.43 millenia/day
- 25%, ETA: 0:00:01,  3.05 millenia/day
- 30%, ETA: 0:00:01, 1955.86 years/day
- 35%, ETA: 0:00:01,  2.29 millenia/day
- 40%, ETA: 0:00:00, 1777.97 years/day
- 45%, ETA: 0:00:00,  2.00 millenia/day
- 50%, ETA: 0:00:00,  2.23 millenia/day
- 55%, ETA: 0:00:00, 1860.95 years/day
- 60%, ETA: 0:00:00,  2.03 millenia/day
- 65%, ETA: 0:00:00,  2.20 millenia/day
- 70%, ETA: 0:00:00, 1908.67 years/day
- 75%, ETA: 0:00:00,  2.05 millenia/day
- 80%, ETA: 0:00:00, 1798.42 years/day
- 85%, ETA: 0:00:00, 1912.05 years/day
- 90%, ETA: 0:00:00,  2.03 millenia/day
- 95%, ETA: 0:00:00, 1833.60 years/day
-100%, ETA: 0:00:00, 1927.63 years/day
-
-Integration done in 730 milliseconds.
