+++ previews/PR483/run_0002/progress.txt
@@ -1,4 +1,4 @@
-Starting SpeedyWeather.jl run 0002 on Mon, 11 Mar 2024 18:10:17
+Starting SpeedyWeather.jl run 0002 on Wed, 27 Mar 2024 21:14:22
 Integrating:
 SpectralGrid:
 ├ Spectral:   T42 LowerTriangularMatrix{Complex{Float32}}, radius = 6.371e6 m

[…]
