### 2024-03-27 SpeedyWeather/SpeedyWeather.jl#480 @milankl
1. > the "monolithic" statements call for reference[s]

I'm adding [this](https://ieeexplore.ieee.org/abstract/document/8029803) as a reference here. "Monolithic architecture" or "application" is widely used in general software engineering not so much in scientific modelling so thanks for pointing it out. With the reference I believe this should be easy to understand, but I extended the sentence a bit to make it clearer

> A monolithic interface [@Mazlami2017], controlling most of the model's functionality
through arguments of a single function or through parameter files (often called namelists in Fortran), is avoided in favor of a library-style interface.

2. > ditto for "modern trend sees simulations in Fortran and data analysis in Python" - please add reference[s]

Can I push back here? Almost all weather/ocean/climate models are written in Fortran ([ICON](https://gitlab.dkrz.de/icon/icon-model), [IFS](https://github.com/ecmwf-ifs),[ MITgcm](https://github.com/MITgcm/MITgcm), [NEMO](https://forge.nemo-ocean.eu/nemo/nemo), [GFS/FV3](https://github.com/NOAA-GFDL/GFDL_atmos_cubed_sphere/), exceptions I know, using C++, are [SCREAM](https://dl.acm.org/doi/10.1145/3581784.3627044) and [COSMO](https://gmd.copernicus.org/articles/11/1665/2018/)) and many people use python for data analysis and visualisation (see also packages in those organisations github/gitlab profiles). I am not aware of any meta study trying to quantify how much in this field is done in which language and I cannot find one searching for it. If you know a reference to add here please let us know!
