### 2024-03-27 f09bab1
First points of slayoo addressed
