+++ docs/joss/paper.bib
@@ -243,6 +243,19 @@ year = {2020}
   journal = {ECMWF Newsletter},
   abstract = {ECMWF will implement a resolution upgrade for high-resolution forecasts (HRES) and ensemble forecasts (ENS) in spring 2016. HRES will then be run on a grid with a grid-point distance between neighbouring points of approximately 9 km instead of 16 km in the current configuration.},
   howpublished = {https://www.ecmwf.int/node/15041},
+  doi = {10.21957/zwdu9u5i},
+}
+
+@INPROCEEDINGS{Mazlami2017,
+  author={Mazlami, Genc and Cito, Jürgen and Leitner, Philipp},
+  booktitle={2017 IEEE International Conference on Web Services (ICWS)}, 
+  title={Extraction of Microservices from Monolithic Software Architectures}, 
+  year={2017},
+  volume={},
+  number={},
+  pages={524-531},
+  keywords={Couplings;Clustering algorithms;Computer architecture;Cloud computing;Industries;Tools;microservices;extraction;coupling;graph-based clustering},
+  doi={10.1109/ICWS.2017.61},
 }
 
 @article{Molteni2003,
