### 2024-03-27 SpeedyWeather/SpeedyWeather.jl#480 @milankl
5. >  bib: Malardel et al. 2016 - DOI is missing: https://doi.org/10.21957/zwdu9u5i

Added.
