+++ previews/PR387/run_0002/progress.txt
@@ -1,4 +1,4 @@
-Starting SpeedyWeather.jl run 0002 on Wed, 20 Mar 2024 14:47:59
+Starting SpeedyWeather.jl run 0002 on Wed, 27 Mar 2024 22:16:53
 Integrating:
 SpectralGrid:
 ├ Spectral:   T42 LowerTriangularMatrix{Complex{Float32}}, radius = 6.371e6 m
@@ -9,11 +9,11 @@ Time: 1.0 days at Δt = 1200.0s
 
 All data will be stored in /home/runner/work/SpeedyWeather.jl/SpeedyWeather.jl/docs/build/run_0002
 
- 25%, ETA: 0:00:00,   Inf millenia/day
- 35%, ETA: 0:00:00,   Inf millenia/day
- 45%, ETA: 0:00:00,   Inf millenia/day
- 80%, ETA: 0:00:00,   Inf millenia/day
- 90%, ETA: 0:00:00,   Inf millenia/day
-100%, ETA: 0:00:00,   Inf millenia/day
+ 25%, ETA: 0:00:00, 46884.27 millenia/day
+ 35%, ETA: 0:00:00, 66189.56 millenia/day
+ 45%, ETA: 0:00:00, 85494.85 millenia/day
+ 80%, ETA: 0:00:00, 154442.32 millenia/day
+ 90%, ETA: 0:00:00, 173747.61 millenia/day
+100%, ETA: 0:00:00, 193052.90 millenia/day
 
 Integration done in empty period.
