+++ previews/PR387/run_0004/progress.txt
@@ -1,4 +1,4 @@
-Starting SpeedyWeather.jl run 0004 on Wed, 20 Mar 2024 14:48:23
+Starting SpeedyWeather.jl run 0004 on Wed, 27 Mar 2024 22:17:15
 Integrating:
 SpectralGrid:
 ├ Spectral:   T63 LowerTriangularMatrix{Complex{Float32}}, radius = 6.371e6 m
@@ -11,23 +11,23 @@ All data will be stored in /home/runner/work/SpeedyWeather.jl/SpeedyWeather.jl/d
 
   5%, ETA: 0:00:01,   Inf millenia/day
  10%, ETA: 0:00:01,   Inf millenia/day
- 15%, ETA: 0:00:01,  2.08 millenia/day
- 20%, ETA: 0:00:01,  2.76 millenia/day
- 25%, ETA: 0:00:01,  3.46 millenia/day
- 30%, ETA: 0:00:01,  2.10 millenia/day
- 35%, ETA: 0:00:00,  2.45 millenia/day
- 40%, ETA: 0:00:00, 1861.77 years/day
- 45%, ETA: 0:00:00,  2.10 millenia/day
- 50%, ETA: 0:00:00,  2.33 millenia/day
- 55%, ETA: 0:00:00, 1925.08 years/day
- 60%, ETA: 0:00:00,  2.10 millenia/day
- 65%, ETA: 0:00:00,  2.27 millenia/day
- 70%, ETA: 0:00:00, 1959.72 years/day
- 75%, ETA: 0:00:00,  2.10 millenia/day
- 80%, ETA: 0:00:00, 1833.05 years/day
- 85%, ETA: 0:00:00, 1948.86 years/day
+ 15%, ETA: 0:00:01,  2.06 millenia/day
+ 20%, ETA: 0:00:01,  2.73 millenia/day
+ 25%, ETA: 0:00:01,  3.42 millenia/day
+ 30%, ETA: 0:00:01,  2.09 millenia/day
+ 35%, ETA: 0:00:01,  2.44 millenia/day
+ 40%, ETA: 0:00:00,  2.78 millenia/day
+ 45%, ETA: 0:00:00,  2.03 millenia/day
+ 50%, ETA: 0:00:00,  2.26 millenia/day
+ 55%, ETA: 0:00:00, 1881.00 years/day
+ 60%, ETA: 0:00:00,  2.05 millenia/day
+ 65%, ETA: 0:00:00,  2.22 millenia/day
+ 70%, ETA: 0:00:00, 1915.31 years/day
+ 75%, ETA: 0:00:00,  2.05 millenia/day
+ 80%, ETA: 0:00:00,  2.19 millenia/day
+ 85%, ETA: 0:00:00, 1945.98 years/day
  90%, ETA: 0:00:00,  2.06 millenia/day
- 95%, ETA: 0:00:00, 1871.48 years/day
-100%, ETA: 0:00:00, 1967.46 years/day
+ 95%, ETA: 0:00:00, 1868.36 years/day
+100%, ETA: 0:00:00, 1964.17 years/day
 
 Integration done in 720 milliseconds.
