+++ previews/PR387/run_0001/progress.txt
@@ -1,4 +1,4 @@
-Starting SpeedyWeather.jl run 0001 on Wed, 20 Mar 2024 14:47:59
+Starting SpeedyWeather.jl run 0001 on Wed, 27 Mar 2024 22:16:53
 Integrating:
 SpectralGrid:
 ├ Spectral:   T42 LowerTriangularMatrix{Complex{Float32}}, radius = 6.371e6 m
