+++ previews/PR387/run_0005/parameters.txt
@@ -12,7 +12,7 @@ Earth{Float32} <: SpeedyWeather.AbstractPlanet
 ├ length_of_year::Second = 31557600 seconds
 ├ equinox::DateTime = 2000-03-20T00:00:00
 ├ axial_tilt::Float32 = 23.4
-└ solar_constant::Float32 = 1365.0
+└ solar_constant::Float32 = 341.25
 EarthAtmosphere{Float32} <: SpeedyWeather.AbstractAtmosphere
 ├ mol_mass_dry_air::Float32 = 28.9649
 ├ mol_mass_vapour::Float32 = 18.0153
@@ -182,7 +182,7 @@ EarthOrography{Float32, OctahedralGaussianGrid{Float32}} <: SpeedyWeather.Abstra
 ├ file::String = orography.nc
 ├ file_Grid::UnionAll = FullGaussianGrid
 ├ scale::Float64 = 1.0
-├ smoothing::Bool = true
+├ smoothing::Bool = false
 ├ smoothing_power::Float64 = 1.0
 ├ smoothing_strength::Float64 = 0.1
 ├ smoothing_truncation::Int64 = 85
