+++ docs/src/analysis.md
@@ -1,11 +1,12 @@
 # Analysing a simulation
 
-While you can analyze a SpeedyWeather simulation through its [NetCDF output](@ref),
+While you can analyze a SpeedyWeather simulation through its [NetCDF output](@ref)
+(i.e., _offline_ analysis),
 as most users will be used to with other models, you can also reuse a lot of
 functionality from SpeedyWeather interactively for analysis. This makes
 SpeedyWeather beyond being a model for the atmospheric general circulation
 also a library with many functions for the analysis of simulations.
-Often this also avoids the two language problem that you will face if you
+Often this also avoids the two-language problem that you will face if you
 run a simulation with a model in one language but then do the data
 analysis in another, treating the model as a blackbox although it likely
 has many of the functions you will need for analysis already defined.
@@ -18,13 +19,14 @@ and analysis are done interactively with the same library: SpeedyWeather.jl.
 Now you could run a SpeedyWeather simulation, and analyse the [NetCDF output](@ref)
 but that comes with several issues related to accuracy
 
-- If you use a reduced grid for the simulation, then the output will (by default) be interpolated on a full grid. This interpolation comes introduces an error.
+- If you use a reduced grid for the simulation, then the output will (by default) be interpolated on a full grid. This interpolation introduces an error.
 - Computing integrals over gridded data on the sphere by weighting every grid point according to its area is not the most accurate numerical integration.
 - Computing gradients over gridded data comes with similar issues. While our [RingGrids](@ref) are always equidistant in longitude, they are not necessarily in latitude.
 
 The first point you can avoid by running a simulation on one of the full grids that
 are implemented, see [SpectralGrid](@ref). But that also impacts the simulation
 and for various reasons we don't run on full grids by default.
+
 The second point you can address by defining a more advanced numerical integration
 scheme, but that likely requires you to depend on external libraries and then,
 well, you could also just depend on SpeedyWeather.jl directly, because we
@@ -35,7 +37,7 @@ spectral transforms to reduce the discretization error.
 The following contains a (hopefully growing) list of examples
 of how a simulation can be analysed interactively. We call this
 _online_ analysis because you are directly using the functionality from
-the model as if it was a library.
+the model as if it was a library.  For simplicity, we use [shallow-water model](@ref shallow_water_model) to demonstrate this.
 
 ## Mass conservation
 
@@ -47,13 +49,13 @@ The total volume is defined as the integral of the dynamic layer thickness
 over the surface ``A`` of the sphere
 
 ```math
-\iint h dA = \iint \eta dA + \iint H dA - \iint H_b dA
+M = \iint h dA = \iint \left(\eta + H - H_b\right) dA
 ```
 
 to check for conservation we want to assess that
 
 ```math
-\frac{\partial}{\partial t} \iint h dA = 0
+\frac{\partial M}{\partial t} = 0
 ```
 
 And because ``V = \iint H dA - \iint H_b dA``, the total volume at rest,
@@ -112,14 +114,16 @@ Calculating the divergence in spherical harmonics always sets the ``l=m=0`` mode
[…]
