### 2024-03-28 129b11c
remove copy, inline triangle number
