### 2024-03-28 SpeedyWeather/SpeedyWeather.jl#480 @navidcy
> 2. > ditto for "modern trend sees simulations in Fortran and data analysis in Python" - please add reference[s]
>
> Can I push back here? Almost all weather/ocean/climate models are written in Fortran ([ICON](https://gitlab.dkrz.de/icon/icon-model), [IFS](https://github.com/ecmwf-ifs),[ MITgcm](https://github.com/MITgcm/MITgcm), [NEMO](https://forge.nemo-ocean.eu/nemo/nemo), [GFS/FV3](https://github.com/NOAA-GFDL/GFDL_atmos_cubed_sphere/), exceptions I know, using C++, are [SCREAM](https://dl.acm.org/doi/10.1145/3581784.3627044) and [COSMO](https://gmd.copernicus.org/articles/11/1665/2018/)) and many people use python for data analysis and visualisation (see also packages in those organisations github/gitlab profiles). I am not aware of any meta study trying to quantify how much in this field is done in which language and I cannot find one searching for it. If you know a reference to add here please let us know!
>
An alternative is to write something along the lines of, e.g.,

> modern trend sees simulations in Fortran (see, e.g., `here citations for 3-4 FORTRAN models`) and data analysis in Python (`cite here xarray, dask, perhaps some other package`).

?
