### 2024-03-28 eaf566a
global diagnostics plot created in docs
