+++ docs/src/analysis.md
@@ -37,7 +37,8 @@ spectral transforms to reduce the discretization error.
 The following contains a (hopefully growing) list of examples
 of how a simulation can be analysed interactively. We call this
 _online_ analysis because you are directly using the functionality from
-the model as if it was a library.  For simplicity, we use [shallow-water model](@ref shallow_water_model) to demonstrate this.
+the model as if it was a library.  For simplicity, we use the
+[shallow water model](@ref shallow_water_model) to demonstrate this.
 
 ## Mass conservation
 
@@ -95,8 +96,9 @@ So the initial conditions in this simulation are such that the global mean inter
 is that value in meters. You would need to multiply by the area of the sphere
 ``4\pi r^2`` (radius ``r``) to get the actual integral from above, but because that doesn't
 change with time either, we just want to check that `η_mean` doesn't change with time.
-Which is equivalent to ``\partial_t \iint \eta dA = 0`` and so volume conservation and because density is constant
-also mass conservation. Let's check what happens after the simulation ran for some days
+Which is equivalent to ``\partial_t \iint \eta dA = 0`` and so volume conservation and
+because density is constant also mass conservation.
+Let's check what happens after the simulation ran for some days
 
 ```@example analysis
 model.feedback.verbose = false # hide
@@ -110,13 +112,13 @@ which is _exactly_ the same. So mass is conserved, woohoo.
 
 Insight from a numerical perspective: The tendency of ``\eta`` is
 ``\partial_t \eta = -\nabla \cdot (\mathbf{u} h)`` which is a divergence of a flux.
-Calculating the divergence in spherical harmonics always sets the ``l=m=0`` mode to zero exactly
-(the gradient of a periodic variable has zero mean) so the time integration here is always with
-an exactly zero tendency.
+Calculating the divergence in spherical harmonics always sets the ``l=m=0`` mode
+to zero exactly (the gradient of a periodic variable has zero mean) so the
+time integration here is always with an exactly zero tendency.
 
 ## Energy
 
-The total energy in the shallow water equation is the sum of the kinetic energy density
+The total energy in the [shallow water equations](@ref shallow_water_model) is the sum of the kinetic energy density
 ``\frac{1}{2}(u^2 + v^2)`` and the potential energy density ``gz`` integrated over the
 total volume (times ``h=\eta+H-H_b`` for the vertical then integrated over the sphere
 ``\iint dA``).
@@ -128,8 +130,8 @@ E = \iint \left[ \int_{H_b}^{\eta}\frac{1}{2}\left(u^2 + v^2 + gz\right)dz \righ
 
 In contrast to the [Mass conservation](@ref) which, with respect to the
 spectral transform, is a linear calculation, here we need to multiply
-variables, which is done in grid-point space. Then we can transform to spectral
-space for the global integral. Let us define a `total_energy` function as
+variables, which has to be done in grid-point space. Then we can transform to spectral
+space for the global integral as before. Let us define a `total_energy` function as
 
 ```@example analysis
 using SpeedyWeather
@@ -150,7 +152,7 @@ end
 ```
 
 So at the current state of our simulation we have a total energy
-(per square meter as we haven't multiplied by the surface area of the planet)
+(per square meter as we haven't multiplied by the surface area of the planet).
 
 ```@example analysis
[…]
