### 2024-03-28 658f70f
spectral_grid.radius not planet.radius
