### 2024-03-28 93bd766
rename callback instance to diagnostics_recorder
