### 2024-03-29 SpeedyWeather/SpeedyWeather.jl#480 @slayoo
@milankl @navidcy thanks for the follow ups.

As @navidcy comments, indeed my point was to avoid unevidenced/jargon statements such as "Most weather, ocean and climate models" (-> which models?), "From this tradition" (-> what do you mean by tradition?), "specific programming style" (-> what style?). In principle, this is a software engineering journal - many readers might not be acquainted with this particular niche, but would be interested to understand what is the programming style of weather forecasting systems. Similarly, newcomers to weather forecasting would benefit from these sentences having concrete examples provided.
