+++ docs/src/analysis.md
@@ -124,8 +124,10 @@ total volume (times ``h=\eta+H-H_b`` for the vertical then integrated over the s
 ``\iint dA``).
 
 ```math
-E = \iint \left[ \int_{H_b}^{\eta}\frac{1}{2}\left(u^2 + v^2 + gz\right)dz \right]dA \\
-  = \iint \frac{1}{2}\left(u^2 + v^2 + gh\right)h dA
+\begin{align}
+E & = \iint \left[ \int_{H_b}^{\eta}\frac{1}{2}\left(u^2 + v^2 + gz\right)dz \right]dA \\
+  & = \iint \frac{1}{2}\left(u^2 + v^2 + gh\right)h dA
+\end{align}
 ```
 
 In contrast to the [Mass conservation](@ref) which, with respect to the
@@ -136,8 +138,9 @@ space for the global integral as before. Let us define a `total_energy` function
 ```@example analysis
 using SpeedyWeather
 function total_energy(u, v, η, model)
+    # allocate grid variable
     h = zero(u)
-    E = zero(u)                             # allocate grid variable
+    E = zero(u)
     
     H = model.atmosphere.layer_thickness
     Hb = model.orography.orography
@@ -147,7 +150,7 @@ function total_energy(u, v, η, model)
     @. E = h/2*(u^2 + v^2) + g*h^2 # vertically-integrated mechanical energy
 
     # transform to spectral, take l=m=0 mode at [1] and normalize for mean
-    E_mean = real(spectral(E)[1]) / model.spectral_transform.norm_sphere
+    return E_mean = real(spectral(E)[1]) / model.spectral_transform.norm_sphere
 end
 ```
 
@@ -164,7 +167,7 @@ TE = total_energy(u, v, η, model)
 ```
 
 with units of ``m^3 s^{-2}`` (multiplying by surface area of the sphere
-and density of the fluid would turn it into joule = ``kg m^2 s^{-2}``).
+and density of the fluid would turn it into joule = ``kg \, m^2 s^{-2}``).
 To know in general where to find the respective variables ``u, v, \eta`` inside our
 simulation object see [Prognostic variables](@ref) and [Diagnostic variables](@ref).
 Now let us continue the simulation
@@ -254,8 +257,9 @@ Following previous examples, let us define a `total_angular_momentum` function a
 using SpeedyWeather
 
 function total_angular_momentum(u, η, model)
+    # allocate grid variable
     h = zero(u)
-    Λ = zero(u)                             # allocate grid variable
+    Λ = zero(u)
     
     H = model.atmosphere.layer_thickness
     Hb = model.orography.orography
@@ -268,7 +272,7 @@ function total_angular_momentum(u, η, model)
     @. Λ = (u*r + Ω*r^2) * h    # vertically-integrated AAM
 
     # transform to spectral, take l=m=0 mode at [1] and normalize for mean
-    Λ_mean = real(spectral(Λ)[1]) / model.spectral_transform.norm_sphere
+    return Λ_mean = real(spectral(Λ)[1]) / model.spectral_transform.norm_sphere
 end
 ```
 
@@ -298,21 +302,21 @@ Total circulation is defined as the area-integrated absolute vorticity:
 C = \iint \left(\zeta + f\right) dA
 ```
 
-Following previous fashion, we define a function ``total_circulation`` for this:
+Following previous fashion, we define a function `total_circulation` for this:
 
 ```@example analysis
 function total_circulation(ζ, model)
     f = coriolis(ζ)  # create f on that grid
 
     # transform to spectral, take l=m=0 mode at [1] and normalize for mean
[…]
