### 2024-03-29 60a9c1a
radius always R, note on scaling
