### 2024-03-29 1348bf8
Add circulation calculation
