### 2024-03-29 cce23bd
remove incorrect "set initial condition" comment
