+++ docs/joss/paper.bib
@@ -72,6 +72,20 @@ address = {Exeter, Devon },
 url = {https://scitools.org.uk/cartopy}
 }
 
+@article {CESM,
+      author = {James W. Hurrell and M. M. Holland and P. R. Gent and S. Ghan and Jennifer E. Kay and P. J. Kushner and J.-F. Lamarque and W. G. Large and D. Lawrence and K. Lindsay and W. H. Lipscomb and M. C. Long and N. Mahowald and D. R. Marsh and R. B. Neale and P. Rasch and S. Vavrus and M. Vertenstein and D. Bader and W. D. Collins and J. J. Hack and J. Kiehl and S. Marshall"},
+      title = {The Community Earth System Model: A Framework for Collaborative Research},
+      journal = {Bulletin of the American Meteorological Society},
+      year = {2013},
+      publisher = {American Meteorological Society},
+      address = {Boston MA, USA},
+      volume = {94},
+      number = {9},
+      doi = {10.1175/BAMS-D-12-00121.1},
+      pages= {1339 - 1360},
+      url = {https://journals.ametsoc.org/view/journals/bams/94/9/bams-d-12-00121.1.xml},
+}
+
 @article{Corby1972,
   title = {A General Circulation Model of the Atmosphere Suitable for Long Period Integrations},
   author = {Corby, G. A. and Gilchrist, A. and Newson, R. L.},
@@ -85,6 +99,13 @@ url = {https://scitools.org.uk/cartopy}
   urldate = {2023-03-07},
   abstract = {The design of a general circulation model using the primitive equations in spherical form is described, including a statement of the finite difference forms used to integrate the system and explanations of the motives for unusual aspects of the finite difference scheme. The model incorporates the hydrological cycle, topography, a simple scheme for the radiative exchanges and arrangements for the simulation of deep free convection (sub grid-scale) and for the representation of exchanges of momentum, sensible and latent heat with the underlying surface. An experiment performed with the model forms the subject of a separate paper.},
   langid = {english},
+}
+
+@Manual{Dask,
+  title = {Dask: Library for dynamic task scheduling},
+  author = {{Dask Development Team}},
+  year = {2016},
+  url = {http://dask.pydata.org},
 }
 
 @article{Gorski2005,
@@ -167,6 +188,20 @@ url = {https://scitools.org.uk/cartopy}
   doi = {10.1109/MCSE.2007.55},
   abstract = {Matplotlib is a 2D graphics package used for Python for application development, interactive scripting,and publication-quality image generation across user interfaces and operating systems},
   keywords = {application development,Computer languages,Equations,Graphical user interfaces,Graphics,Image generation,Interpolation,Operating systems,Packaging,Programming profession,Python,scientific programming,scripting languages,User interfaces},
+}
+
+@article{ICON,
+  author = {Giorgetta, M. A. and Brokopf, R. and Crueger, T. and Esch, M. and Fiedler, S. and Helmert, J. and Hohenegger, C. and Kornblueh, L. and Köhler, M. and Manzini, E. and Mauritsen, T. and Nam, C. and Raddatz, T. and Rast, S. and Reinert, D. and Sakradzija, M. and Schmidt, H. and Schneck, R. and Schnur, R. and Silvers, L. and Wan, H. and Zängl, G. and Stevens, B.},
+  title = {ICON-A, the Atmosphere Component of the ICON Earth System Model: I. Model Description},
+  journal = {Journal of Advances in Modeling Earth Systems},
+  volume = {10},
+  number = {7},
+  pages = {1613-1637},
[…]
