### 2024-03-29 SpeedyWeather/SpeedyWeather.jl#480 @milankl
@slayoo can you check whether you are happy with our changes, see pdf [here](https://github.com/SpeedyWeather/SpeedyWeather.jl/blob/mk/josspaper/docs/joss/paper.pdf)? Let us know if there's other points. I believe I've addressed all points, to summarise

1. > the "monolithic" statements call for reference[s]

I'm adding [this](https://ieeexplore.ieee.org/abstract/document/8029803) as a reference here. "Monolithic architecture" or "application" is widely used in general software engineering not so much in scientific modelling so thanks for pointing it out. With the reference I believe this should be easy to understand, but I extended the sentence a bit to make it clearer

2. > ditto for "modern trend sees simulations in Fortran and data analysis in Python" - please add reference[s]

Added as

>> Most weather, ocean and climate models are written in Fortran
(e.g. ICON [@ICON], CESM [@CESM], MITgcm [@MITgcm], NEMO [@NEMO]) [...]
and data analysis in Python (e.g. NumPy [@Numpy], Xarray [@Xarray],
Dask [@Dask], Matplotlib [@Hunter2007]),

3. > what is "large-scale condensation"? (and how would it be different from other-scale condensation?)

In atmospheric modelling we talk about "large-scale condensation" when a grid-cell is saturated as you cannot resolve the droplet formation directly and therefore parameterize _large-scale_ condensation when relative humidity is above 100%. I see that this term might be confusing so I've changed these sentences to use "convection" to avoid the confusion and the need to add an explanation, now reads as

>> To define a new parameterization for convection in a given vertical column
of the atmosphere, one would define `MyConvection` as a new subtype of
`AbstractConvection`. One then only needs to extend the `initialize!`
(executed once during model initialization) and `convection!`
(executed on every time step) functions for this new type. Passing on
`convection = MyConvection()` to the model constructor then implements
this new model component without the need to branch off or overwrite
existing model components.

4. > why is Cartopy cited as "2010-2015"?

That's their official citation recommendation: https://scitools.org.uk/cartopy/docs/latest/citation.html#bibtex-entry

5. > bib: Malardel et al. 2016 - DOI is missing: https://doi.org/10.21957/zwdu9u5i

Added.

6. > bib: Moses & Churavy 2020 - add clickable link (https://proceedings.neurips.cc/paper_files/paper/2020/file/9332c513ef44b682e9347822c2e457ac-Paper.pdf or https://doi.org/10.48550/arXiv.2010.01709)

Added.

7. > bib: Rasp et al. 2018 - DOI is missing: https://doi.org/10.1073/pnas.1810286115

Added.

8. > bib: Stompor 2011 - add clickable URL: https://ascl.net/1110.013

Added.

9. > bib: Willmert 2020 - use permalink: http://web.archive.org/web/*/https://justinwillmert.com/articles/2020/notes-on-calculating-the-spherical-harmonics/

Added.

10. > paper mentions that "simple parallelization across vertical layers is supported" but README informs also about "grid point-wise [multi-threading] for physics", and CUDA is among the dependencies - please clarify and elaborate

Clarified in the manuscript as

>> A simple parallelization
(across vertical layers for the dynamical core, across horizontal grid points
for the parameterizations) is supported by Julia's multithreading.
No distributed-memory parallelization is currently supported,
GPU support is planned.

And also similarly in the README.

11. > if distributed-memory parallelisation (e.g., MPI) is not supported, please mention it explicitly

See above.
