### 2024-03-29 584e26f
Use ringgrids broadcasting throughout docs
