+++ docs/src/analysis.md
@@ -138,16 +138,12 @@ space for the global integral as before. Let us define a `total_energy` function
 ```@example analysis
 using SpeedyWeather
 function total_energy(u, v, η, model)
-    # allocate grid variable
-    h = zero(u)
-    E = zero(u)
-    
     H = model.atmosphere.layer_thickness
     Hb = model.orography.orography
     g = model.planet.gravity
     
-    @. h = η + H - Hb               # layer thickness between the bottom and free surface
-    @. E = h/2*(u^2 + v^2) + g*h^2  # vertically-integrated mechanical energy
+    h = @. η + H - Hb               # layer thickness between the bottom and free surface
+    E = @. h/2*(u^2 + v^2) + g*h^2  # vertically-integrated mechanical energy
 
     # transform to spectral, take l=m=0 mode at [1] and normalize for mean
     return E_mean = real(spectral(E)[1]) / model.spectral_transform.norm_sphere
@@ -218,14 +214,12 @@ f = coriolis(ζ)     # create f on that grid
 
 # layer thickness
 η = simulation.diagnostic_variables.surface.pres_grid
-h = zero(η)
 H = model.atmosphere.layer_thickness
 Hb = model.orography.orography
-@. h = η + H - Hb
+h = @. η + H - Hb
 
 # potential vorticity
-q = zero(ζ)
-@. q = (f + ζ) / h
+q = @. (f + ζ) / h
 nothing # hide
 ```
 
@@ -257,10 +251,6 @@ Following previous examples, let us define a `total_angular_momentum` function a
 using SpeedyWeather
 
 function total_angular_momentum(u, η, model)
-    # allocate grid variable
-    h = zero(u)
-    Λ = zero(u)
-    
     H = model.atmosphere.layer_thickness
     Hb = model.orography.orography
     R = model.spectral_grid.radius
@@ -268,8 +258,8 @@ function total_angular_momentum(u, η, model)
 
     r = R * cos.(model.geometry.lats)       # momentum arm for every grid point
     
-    @. h = η + H - Hb           # layer thickness between the bottom and free surface
-    @. Λ = (u*r + Ω*r^2) * h    # vertically-integrated AAM
+    h = @. η + H - Hb           # layer thickness between the bottom and free surface
+    Λ = @. (u*r + Ω*r^2) * h    # vertically-integrated AAM
 
     # transform to spectral, take l=m=0 mode at [1] and normalize for mean
     return Λ_mean = real(spectral(Λ)[1]) / model.spectral_transform.norm_sphere
@@ -306,10 +296,8 @@ Following previous fashion, we define a function `total_circulation` for this
 
 ```@example analysis
 function total_circulation(ζ, model)
-    f = coriolis(ζ)  # create f on that grid
-    C = zero(ζ)
-    @. C = ζ + f
-
+    f = coriolis(ζ)         # create f on the grid of ζ
+    C = ζ .+ f              # absolute vorticity
     # transform to spectral, take l=m=0 mode at [1] and normalize for mean
     return C_mean = real(spectral(C)[1]) / model.spectral_transform.norm_sphere
 end
@@ -385,9 +373,7 @@ to show how to global integral ``\iint dV`` can be written more efficiently
 # define a global integral, reusing a precomputed SpectralTransform S
 # times surface area of sphere omitted
 function ∬dA(v, h, S::SpectralTransform)
-    vh = zero(h)
-    @. vh = v * h
-    return real(spectral(vh, S)[1]) / S.norm_sphere
+    return real(spectral(v .* h, S)[1]) / S.norm_sphere
 end
 
 # use SpectralTransform from model
[…]
