### 2024-04-02 #6323 @vavrines
@kthyng
Parts of https://github.com/SpeedyWeather/SpeedyWeather.jl/issues/446 are not yet complete. I see @milankl is working on it.
