### 2024-04-02 dad1391
faster similar for same number type
