### 2024-04-02 543e50f
with machine details
