+++ benchmark/manual_benchmarking.jl
@@ -18,7 +18,12 @@ version = SpeedyWeather.pkgversion(SpeedyWeather)
 write(md, "\ncreated for SpeedyWeather.jl v$version on $(Dates.format(Dates.now(), Dates.RFC1123Format)). \n\n")
 write(md, "All simulations have been benchmarked over several seconds (wallclock time) without output. ")
 write(md, "Benchmarking excludes initialization and is started just before the main time loop and finishes right after. ")
-write(md, "All simulations single-threaded on a CPU. \n\n")
+write(md, "All simulations single-threaded on a CPU, ")
+write(md, "<details><summary>Machine details</summary>")
+versioninfo(md)
+write(md,"</details>")
+
+write(md, "\n")
 write(md, "The benchmarking results here are not very robust, timings that change with +-50% are not uncommon. ")
 write(md, "Proper benchmarking for performance optimization uses the minimum or median of many executions, ")
 write(md, "while we run a simulation for several time steps which effectively represents the mean, ")
