### 2024-04-02 55add34
more machine details
