+++ benchmark/manual_benchmarking.jl
@@ -18,7 +18,7 @@ version = SpeedyWeather.pkgversion(SpeedyWeather)
 write(md, "\ncreated for SpeedyWeather.jl v$version on $(Dates.format(Dates.now(), Dates.RFC1123Format)). \n\n")
 write(md, "All simulations have been benchmarked over several seconds (wallclock time) without output. ")
 write(md, "Benchmarking excludes initialization and is started just before the main time loop and finishes right after. ")
-write(md, "All simulations single-threaded on a CPU, ")
+write(md, "All simulations single-threaded on a CPU, more details:")
 write(md, "<details><summary>Machine details</summary>")
 versioninfo(md)
 write(md,"</details>")
