+++ benchmark/benchmarks.md
@@ -1,106 +1,0 @@
-# Benchmarks
-
-created for SpeedyWeather.jl v0.9.0 on Tue, 02 Apr 2024 12:14:41. 
-
-All simulations have been benchmarked over several seconds (wallclock time) without output. Benchmarking excludes initialization and is started just before the main time loop and finishes right after. All simulations single-threaded on a CPU, <details><summary>Machine details</summary>Julia Version 1.10.2
-Commit bd47eca2c8a (2024-03-01 10:14 UTC)
-Build Info:
-  Official https://julialang.org/ release
-Platform Info:
-  OS: macOS (x86_64-apple-darwin22.4.0)
-  CPU: 8 × Intel(R) Core(TM) i5-1030NG7 CPU @ 1.10GHz
-  WORD_SIZE: 64
-  LIBM: libopenlibm
-  LLVM: libLLVM-15.0.7 (ORCJIT, icelake-client)
-Threads: 1 default, 0 interactive, 1 GC (on 8 virtual cores)
-Environment:
-  LD_LIBRARY_PATH = /Users/milan/.julia/conda/3/lib:
-</details>
-The benchmarking results here are not very robust, timings that change with +-50% are not uncommon. Proper benchmarking for performance optimization uses the minimum or median of many executions, while we run a simulation for several time steps which effectively represents the mean, susceptible to outliers that slow down the simulation. However, this is what a user will experience in most situations anyway and the following therefore presents a rough idea of how fast a SpeedyWeather simulation will run, and how much memory it requires.
-
-Explanation
-- NF: Number format, default: Float32
-- T: Spectral resolution, maximum degree of spherical harmonics, default: T31
-- L: Number of vertical layers, default: 8 (for 3D models)
-- Grid: Horizontal grid, default: OctahedralGaussianGrid
-- Rings: Grid-point resolution, number of latitude rings pole to pole
-- Dynamics: With dynamics?, default: true
-- Physics: With physical parameterizations?, default: true (for primitive equation models)
-- Δt: time step [s].
-- SYPD: Speed of simulation, simulated years per wallclock day.
-- Memory: Memory footprint of simulation, variables and constants.
-
-## Models, default setups
-
-| Model | T | L | Physics | Δt | SYPD | Memory|
-| - | - | - | - | - | - | - |
-| BarotropicModel | 31 | 1 | false | 1800 | 22436 | 1.22 MB |
-| ShallowWaterModel | 31 | 1 | false | 1800 | 13086 | 1.24 MB |
-| PrimitiveDryModel | 31 | 8 | true | 1800 | 582 | 3.95 MB |
-| PrimitiveWetModel | 31 | 8 | true | 1800 | 472 | 4.28 MB |
-
-## Grids
-
-| Model | T | L | Grid | Rings | Δt | SYPD | Memory|
-| - | - | - | - | - | - | - | - |
-| PrimitiveWetModel | 63 | 8 | FullGaussianGrid | 96 | 900 | 39 | 22.50 MB |
-| PrimitiveWetModel | 63 | 8 | FullClenshawGrid | 95 | 900 | 39 | 22.29 MB |
[…]
