+++ benchmark/manual_benchmarking.jl
@@ -1,14 +1,14 @@
 using SpeedyWeather, Dates, Printf
 import SpeedyWeather.SpeedyTransforms: prettymemory
 
-# include("benchmark_suite.jl")
-# include("define_benchmarks.jl")
+include("benchmark_suite.jl")
+include("define_benchmarks.jl")
 
-# ## RUN BENCHMARKS
-# for key in keys(benchmarks)
-#     suite = benchmarks[key]
-#     run_benchmark_suite!(suite)
-# end
+## RUN BENCHMARKS
+for key in keys(benchmarks)
+    suite = benchmarks[key]
+    run_benchmark_suite!(suite)
+end
 
 ## WRITE into benchmarks.md
 md = open("README.md", "w")
@@ -18,14 +18,6 @@ version = SpeedyWeather.pkgversion(SpeedyWeather)
 write(md, "\ncreated for SpeedyWeather.jl v$version on $(Dates.format(Dates.now(), Dates.RFC1123Format)). \n\n")
 write(md, "All simulations have been benchmarked over several seconds (wallclock time) without output. ")
 write(md, "Benchmarking excludes initialization and is started just before the main time loop and finishes right after. ")
-write(md, "All simulations single-threaded on a CPU, more details:\n")
-
-write(md, "<details><summary>Machine details</summary>\n")
-write(md, "```julia\njulia> versioninfo()\n")
-versioninfo(md)
-write(md,"```\n</details>")
-
-write(md, "\n")
 write(md, "The benchmarking results here are not very robust, timings that change with +-50% are not uncommon. ")
 write(md, "Proper benchmarking for performance optimization uses the minimum or median of many executions, ")
 write(md, "while we run a simulation for several time steps which effectively represents the mean, ")
@@ -33,7 +25,15 @@ write(md, "susceptible to outliers that slow down the simulation. However, this 
 write(md, "in most situations anyway and the following therefore presents a rough idea of how fast a ")
 write(md, "SpeedyWeather simulation will run, and how much memory it requires.\n\n")
 
-write(md, "Explanation\n")
+write(md, "### Machine details\n\n")
+write(md, "All simulations single-threaded on a CPU:\n")
+
+write(md, "```julia\njulia> versioninfo()\n")
+versioninfo(md)
+write(md,"```\n\n")
+
+write(md, "### Explanation\n\n")
+write(md, "Abbreviations in the tables below are as follows, omitted columns use defaults.\n")
 write(md, "- NF: Number format, default: $(SpeedyWeather.DEFAULT_NF)\n")
 write(md, "- T: Spectral resolution, maximum degree of spherical harmonics, default: T$(SpeedyWeather.DEFAULT_TRUNC)\n")
 write(md, "- L: Number of vertical layers, default: $(SpeedyWeather.DEFAULT_NLEV) (for 3D models)\n")
