+++ benchmark/README.md
@@ -1,12 +1,12 @@
 # Benchmarks
 
-created for SpeedyWeather.jl v0.9.0 on Tue, 02 Apr 2024 13:53:20. 
+created for SpeedyWeather.jl v0.9.0 on Tue, 02 Apr 2024 14:05:51. 
 
 All simulations have been benchmarked over several seconds (wallclock time) without output. Benchmarking excludes initialization and is started just before the main time loop and finishes right after. The benchmarking results here are not very robust, timings that change with +-50% are not uncommon. Proper benchmarking for performance optimization uses the minimum or median of many executions, while we run a simulation for several time steps which effectively represents the mean, susceptible to outliers that slow down the simulation. However, this is what a user will experience in most situations anyway and the following therefore presents a rough idea of how fast a SpeedyWeather simulation will run, and how much memory it requires.
 
 ### Machine details
 
-All simulations single-threaded on a CPU:
+All benchmark simulation were single-threaded on a CPU:
 ```julia
 julia> versioninfo()
 Julia Version 1.10.2
@@ -38,77 +38,84 @@ Abbreviations in the tables below are as follows, omitted columns use defaults.
 - SYPD: Speed of simulation, simulated years per wallclock day.
 - Memory: Memory footprint of simulation, variables and constants.
 
+### Running the benchmarks
+
+The benchmark suite here can be reproduced by executing:
+
+```> julia manual_benchmarking.jl```
+
+inside `the SpeedyWeather.jl/benchmark` folder. It will create this `README.md` which can be pushed to the repository for updates or comparison.
 ## Models, default setups
 
 | Model | T | L | Physics | Δt | SYPD | Memory|
 | - | - | - | - | - | - | - |
-| BarotropicModel | 31 | 1 | false | 1800 | 26309 | 1.22 MB |
-| ShallowWaterModel | 31 | 1 | false | 1800 | 14511 | 1.24 MB |
-| PrimitiveDryModel | 31 | 8 | true | 1800 | 670 | 3.95 MB |
-| PrimitiveWetModel | 31 | 8 | true | 1800 | 510 | 4.28 MB |
+| BarotropicModel | 31 | 1 | false | 1800 | 19349 | 1.22 MB |
+| ShallowWaterModel | 31 | 1 | false | 1800 | 10195 | 1.24 MB |
+| PrimitiveDryModel | 31 | 8 | true | 1800 | 530 | 3.95 MB |
+| PrimitiveWetModel | 31 | 8 | true | 1800 | 399 | 4.28 MB |
 
 ## Grids
 
 | Model | T | L | Grid | Rings | Δt | SYPD | Memory|
 | - | - | - | - | - | - | - | - |
-| PrimitiveWetModel | 63 | 8 | FullGaussianGrid | 96 | 900 | 41 | 22.50 MB |
[…]
