### 2024-04-02 579c490
Particle figure for barotropic vorticity equations
