### 2024-04-02 SpeedyWeather/SpeedyWeather.jl#446 @milankl
@vavrines Many thanks for the review, below are our responses

1. > Three sets of equations are introduced at the beginning of the Summary, where the first two are explained with figures. Maybe it makes sense to do the same for the barotropic vorticity equation?

I actually like this suggestion, so I've added this figure

![image](https://github.com/SpeedyWeather/SpeedyWeather.jl/assets/25530332/0ef0151b-c78b-4b1c-8534-8cccd9afc2a9)

which showcases the barotropic vorticity model but also the particle advection, we've recently added.

2. > In the manuscript, both the physical models, e.g., the atmospheric general circulation model, and the computer programs, e.g., the Dedalus, CLIMLAB, and Oceananigans, are called “models”. The latter may be renamed to avoid ambiguity.

I'm hesitating to rename the computer prorgrams to something else than model, not because I don't agree with the ambiguity here, but because the term "model" to mean a computer program that simulates some geophysical phenonmea or processes by solving equations is so common. I'd therefore rather use "equations" wherever the computer program-independent equations are meant, i.e. "shallow water equations", "barotropic vorticity equations", "primitive equations", if it's about the implementation here the word "model" is added, e.g. "The primitive equation model in SpeedyWeather.jl". Does that sound good?

4. > The performance claims of the software appear to be limited. As the weather forecast is a pretty time-consuming task, the audience would be interested to know about the HPC capacity (distributed computing, GPU, scaling, etc.).

This has been i) better reflected in the readme with #486 ii) addressed in the manuscript following a similar comment from another reviewer, see https://github.com/SpeedyWeather/SpeedyWeather.jl/issues/480#issuecomment-2027535734

> A simple parallelization (across vertical layers for the dynamical core, across horizontal grid points for the parameterizations) is supported by Julia's multithreading. No distributed-memory parallelization is currently supported, GPU support is planned.

and iii) we have added a benchmark suite with #509, see https://github.com/SpeedyWeather/SpeedyWeather.jl/tree/main/benchmark

5. > The parameterization and data-driven parts of the documentation are yet to be completed. Is this intentional?

Yes, I'm still working on this and will let you know about progress.

7. > It is mentioned that “The primitive equation core (wet or dry) is in development” in the README. How complete is this development?

This has been rewritten in the README to reflect that the development of the primitive equation models is indeed completed. Sure, there's always things we will improve over time, but it's in a usable state with many different parameterizations implemented, which allow to study a wide range of atmospheric phenomena.
