### 2024-04-02 d76df12
Higher resolution figure for primitive equation showcase
