+++ docs/src/large_scale_condensation.md
@@ -65,7 +65,8 @@ T_{i+1} - T_i &= -\frac{L_v}{c_p}( q_{i+1} - q_i )
 ```
 
 With Euler forward time stepping this is great, but with our [leapfrog timestepping + RAW filter](@id leapfrog)
-this is very dispersive (see #445) although the implicit formulation is already much better.
+this is very dispersive (see [#445](https://github.com/SpeedyWeather/SpeedyWeather.jl/issues/445))
+although the implicit formulation is already much better.
 We therefore introduce a time step ``\Delta t_c`` which makes the implicit condensation not immediate
 anymore but over several time steps ``\Delta t`` of the leapfrogging.
 
