### 2024-04-03 f156fb2
Add Makie extension
