+++ Project.toml
@@ -1,7 +1,7 @@
 name = "SpeedyWeather"
 uuid = "9e226e20-d153-4fed-8a5b-493def4f21a9"
 authors = ["Milan Klöwer and SpeedyWeather contributors"]
-version = "0.9"
+version = "0.9.0"
 
 [deps]
 AbstractFFTs = "621f4979-c628-5d54-868e-fcf4e3e8185c"
@@ -28,6 +28,12 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
 UnicodePlots = "b8865327-cd53-5732-bb35-84acbb429228"
 
+[weakdeps]
+Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
+
+[extensions]
+SpeedyWeatherMakieExt = "Makie"
+
 [compat]
 AbstractFFTs = "1"
 Adapt = "3, 4"
@@ -44,6 +50,7 @@ GenericFFT = "0.1"
 JLD2 = "0.4"
 KernelAbstractions = "0.9"
 LinearAlgebra = "1.8"
+Makie = "0.20"
 NCDatasets = "0.12, 0.13, 0.14"
 Primes = "0.5"
 Printf = "1.8"
