### 2024-04-03 cdd339d
Large-scale condensation documented (#510)
* Large-scale condensation documented

* Jablonowski initial conditions corrected

* Gremlins removed in docs
