+++ docs/src/convection.md
@@ -28,24 +28,118 @@ conceptually for every vertical column in the atmosphere we
 2. Decide whether convection should take place and whether it is deep (precipitating) or shallow (non-precipitating).
 3. Relax temperature and humidity towards (adjusted) profiles from 1.
 
-### Reference profiles
+## Reference profiles
 
 The dry adiabat is
 
 ```math
-T = T_s (\frac{p}{p_s})^\frac{R}{c_p}
+T = T_0 (\frac{p}{p_0})^\frac{R}{c_p}
 ```
 
 The temperature ``T`` of an air parcel at pressure ``p`` is determined by the
-temperature ``T_s`` it had at pressure ``p_s`` and the gas constant for dry air
-``R`` and the heat capacity ``c_p``. The pseudo adiabat follows the dry adiabat
-until saturation is reached, ``q > q^\star``. Then it follows the pseudoadiabatic
-lapse rate
+temperature ``T_0`` it had at pressure ``p_0`` (this can be surface but it does not have to be)
+and the gas constant for dry air ``R = 287.04 J/K/kg`` and the heat capacity ``c_p = 1004.64 J/K/kg``.
+The pseudo adiabat follows the dry adiabat until saturation is reached
+(the lifting condensation level, often abbreviated to LCL), ``q > q^\star``.
+Then it follows the pseudoadiabatic lapse rate
 
 ```math
 \Gamma = -\frac{dT}{dz} = \frac{g}{c_p}\left(
-    \frac{1 + \frac{qL_v}{(1-q)^2R_dT_v}}{1 + \frac{qL^2}{(1-q)^2 c_p R_vT^2}}\right)
+    \frac{  1 + \frac{q^\star L_v  }{(1-q^\star)^2     R_d T_v}}{
+            1 + \frac{q^\star L_v^2}{(1-q^\star)^2 c_p R_v T^2}}\right)
 ```
+
+with gravity ``g``, heat capacity ``c_p``, the saturation specific humidity of the parcel ``q^\star``
+(which is its specific humidity given that it has already reached saturation),
+latent heat of vaporization ``L_v``, dry gas constant ``R_d``, water vapour
+gas constant ``R_v``, and [Virtual temperature](@ref) ``T_v``.
+Starting with a temperature ``T`` and humidity ``q = q^\star`` at the lifting condensation level 
+temperature aloft changes with $dT = -\frac{d\Phi}{c_p}(...)$ between two layers separated
+``d\Phi`` in geopotential ``\Phi`` apart. On that new layer, ``q^\star`` is recalculated
+as well as the virtual temperature ``T_v = T(1 + \mu q^\star)``. ``\mu`` is derived from
+the ratio of dry to vapour gas constants see [Virtual temperature](@ref).
+Note that the pseudoadiabatic ascent is independent of the environmental temperature
+and humidity and function of temperature and humidity of the parcel only
+(although that one starts with surface temperature and humidity from the environment).
+Solely the level of zero buoyancy is determined by comparing the parcel's virtual
+temperature ``T_v`` to the virtual temperature of the environment ``T_{v,e}`` at that level.
+Level of zero buoyancy is reached when ``T_v = T_{v,e}`` but continues for ``T_v > T_{v,e}``
+which means that the parcel is still buoyant. Note that the virtual temperature includes
+the effect that humidity has on its density.
+
+The (absolute) temperature a lifted parcel has during ascent (following its pseudoadiabat, dry and/
+or moist, until reaching the level of zero buoyancy) is then taken as the reference temperature
+profile ``T_{ref}`` that the Betts-Miller convective parameterization relaxes towards as a first guess
[…]
