### 2024-04-03 ba45334
Convection documentation (#512)
* Convection documentation
