+++ docs/src/index.md
@@ -12,7 +12,7 @@ defined and integrated into SpeedyWeather.
 SpeedyWeather.jl is uses a spherical harmonic transform to simulate
 the general circulation of the atmosphere using a vorticity-divergence formulation,
 a semi-implicit time integration and simple parameterizations to represent various
-climate processes: Radiation, clouds, precipitation, surface fluxes, among others.
+climate processes: Convection, clouds, precipitation, radiation, surface fluxes, among others.
 
 SpeedyWeather.jl defines 
 - [`BarotropicModel`](@ref barotropic_vorticity_model) for the 2D barotropic vorticity equation
@@ -22,7 +22,24 @@ SpeedyWeather.jl defines
 
 and solves these equations in spherical coordinates as described in this documentation.
 
-## Developers
+## Vision
+
+Why another model? You may ask. We believe that most currently available are stiff, difficult to use
+and extend, and therefore slow down research whereas a modern code in a modern language wouldn't have to.
+We decided to use Julia because it combines the best of Fortran and Python: Within a single language
+we can interactively run SpeedyWeather but also extend it, inspect its components, evaluate
+individual terms of the equations, and analyse and visualise output on the fly.
+
+We do not aim to make SpeedyWeather an atmospheric model similar to the production-ready models used
+in weather forecasting, at least not at the cost of our current level of interactivity and ease of
+use or extensibility. If someone wants to implement a cloud parameterization that is very complicated
+and expensive to run then they are more than encouraged to do so, but it will probably live in
+its own repository and we are happy to provide a general interface to do so. But SpeedyWeather's
+defaults should be balanced: Physically accurate yet general; as independently as possible from other
+components and parameter choices; not too complicated to implement and understand; and computationally cheap.
+Finding a good balance is difficult but we try our best. 
+
+## Developers and contributing
 
 The development of  SpeedyWeather.jl is lead by [Milan Klöwer](https://github.com/milankl) and
 [current and past contributors](https://github.com/SpeedyWeather/SpeedyWeather.jl/graphs/contributors) include
@@ -36,6 +53,12 @@ The development of  SpeedyWeather.jl is lead by [Milan Klöwer](https://github.c
 - [Simone Silvestri](https://github.com/simone-silvestri)
 
 Any contributions are always welcome!
+
+Open-source lives from large teams of (even occasional) contributors. If you are interested to
+fix something, implement something, or just use it and provide feedback you are always welcome.
+We are more than happy to guide you, especially when you don't know where to start.
+We can point you to the respective code, highlight how everything is connected and tell you
+about dos and don'ts. Just express your interest to contribute and we'll be happy to have you.
 
 ## Funding
 
