+++ test/lower_triangular_matrix.jl
@@ -430,19 +430,19 @@ end
                 L2 = deepcopy(L1) 
 
                 L2 .*= NF(5)
-                @test L1 .* NF(5) ≈ L2 
+                @test all(L1 .* NF(5) .≈ L2) 
 
                 L1 = adapt(ArrayType, randn(LowerTriangularArray{NF}, 10, 10, idims...))
                 L2 = deepcopy(L1) 
 
                 L2 ./= NF(5)
-                @test L1 ./ NF(5) ≈ L2 
+                @test all(L1 ./ NF(5) .≈ L2) 
 
                 L1 = adapt(ArrayType, randn(LowerTriangularArray{NF}, 10, 10, idims...))
                 L2 = deepcopy(L1)
 
                 L2 .^= NF(2)
-                @test L1 .^ NF(2) ≈ L2
+                @test all(L1 .^ NF(2) .≈ L2)
             end
         end
     end
