### 2024-04-04 c51ebfa
small code cleanup
