+++ src/LowerTriangularMatrices/lower_triangular_matrix.jl
@@ -432,29 +432,6 @@ end
 # GPU methods (should be moved to an extension, in case this becomes a standalone package)
 import GPUArrays._copyto!
 
-"""
-function GPUArrays._copyto!(dest::LowerTriangularArray{T,N,ArrayType}, bc::Broadcasted) where {T,N,ArrayType}
-    axs = axes(dest)
-    axes(bc) == axs || Broadcast.throwdm(axes(bc), axs)
-
-    # write KA kernel
-    dev = KernelAbstractions.get_backend(dest.data)
-    copyto!_kernel!(dev, workgroup_size(dev))(dest.data, bc, dest.m, ndrange=size(dest.data))
-
-    synchronize!(dev)
-    return dest
-end
-
-KernelAbstractions.@kernel function copyto!_kernel!(dest, bc, @Const(m))
-    I = KernelAbstractions.@index(Global, Cartesian)
-    dest[I] = bc[k2ij(I[1], m),I[2:end]...] # TODO: the k2ij is costly, if we intend to accalerate this, this needs to be precomputed
-end 
-
-function workgroup_size(device::KernelAbstractions.Backend)
-    return device isa KernelAbstractions.GPU ? 256 : 16 
-end
-"""
-
 # copied and adjusted from GPUArrays.jl
 @inline function GPUArrays._copyto!(dest::LowerTriangularArray{T,N,ArrayType}, bc::Broadcasted) where {T,N,ArrayType}
     axes(dest) == axes(bc) || Broadcast.throwdm(axes(dest), axes(bc))
@@ -466,10 +443,8 @@ end
             while i < nelem
                 i += 1
                 I = GPUArrays.@cartesianidx(dest.data, i) 
-                @inbounds dest.data[I] = bc[k2ij(I, dest.m)] 
-                # TODO: broken: the indexing doesnt work, probably we need to define a new macro for the lowertriangle to do it better
-                # TODO 1: the k2ij is costly, if we intend to accelarate this, this needs to be precomputed
-            end                                                                     # TODO 2: CartesianIndex can't be indexed with ranges, hence the [I[i] for i=2:N], but this seems suboptimal as well    
+                @inbounds dest.data[I] = bc[k2ij(I, dest.m)] # TODO: the k2ij is costly, if we intend to accelarate this, this needs to be precomputed
+            end                                                                     
             return
         end
 
@@ -484,13 +459,6 @@ end
 
     return dest
 end
-
-import Base: map! 
-
-# TODO: do we need this? does it work? it's in GPUArrays.jl implemented as well
-#function Base.map!(f, dest::LowerTriangularArray{T,N,ArrayType}, xs::AbstractArray...) where {T,N,ArrayType<:GPUArrays.AbstractGPUArray} 
-#    map!(f, dest.data, xs)
-#end 
 
 Broadcast.BroadcastStyle(::Type{LowerTriangularArray{T,N,ArrayType}}) where {T,N,ArrayType <: GPUArrays.AbstractGPUArray} = Broadcast.BroadcastStyle(ArrayType)
 GPUArrays.backend(::Type{LowerTriangularArray{T,N,ArrayType}}) where {T,N,ArrayType <: GPUArrays.AbstractGPUArray} = GPUArrays.backend(ArrayType)
