+++ docs/src/lowertriangularmatrices.md
@@ -4,34 +4,37 @@ LowerTriangularMatrices is a submodule that has been developed for SpeedyWeather
 technically independent (SpeedyWeather.jl however imports it and so does SpeedyTransforms)
 and can also be used without running simulations. It is just not put into its own respective repository.
 
-This module defines `LowerTriangularMatrix`, a lower triangular matrix, which in contrast to
+This module defines `LowerTriangularArray`, a lower triangular matrix format, which in contrast to
 [`LinearAlgebra.LowerTriangular`](https://docs.julialang.org/en/v1/stdlib/LinearAlgebra/#LinearAlgebra.LowerTriangular) does not store the entries above the diagonal. SpeedyWeather.jl
-uses `LowerTriangularMatrix` which is defined as a subtype of `AbstractMatrix` to store
-the spherical harmonic coefficients (see [Spectral packing](@ref)). 
+uses `LowerTriangularArray` which is defined as a subtype of `AbstractArray` to store
+the spherical harmonic coefficients (see [Spectral packing](@ref)). For 2D `LowerTriangularArray` the alias `LowerTriangularMatrix` exists. Higher dimensional `LowerTriangularArray` are 'batches' of 2D `LowerTriangularMatrix`. So, for example a ``(10\times 10\times 10)`` `LowerTriangularArray` holds 10 `LowerTriangularMatrix` of size ``(10\times 10)`` in one array.  
 
-## Creation of `LowerTriangularMatrix`
+## Creation of `LowerTriangularArray` 
 
-A `LowerTriangularMatrix` can be created using `zeros`, `ones`, `rand`, or `randn`
+A `LowerTriangularMatrix` and `LowerTriangularArray` can be created using `zeros`, `ones`, `rand`, or `randn`
 ```@repl LowerTriangularMatrices
 using SpeedyWeather.LowerTriangularMatrices
 
 L = rand(LowerTriangularMatrix{Float32}, 5, 5)
+L2 = rand(LowerTriangularArray{Float32}, 5, 5, 5)
 ```
-or the `undef` initializor `LowerTriangularMatrix{Float32}(undef, 3, 3)`.
+or the `undef` initializors `LowerTriangularMatrix{Float32}(undef, 3, 3)`, `.
 The element type is arbitrary though, you can use any type `T` too.
 
-Alternatively, it can be created through conversion from `Matrix`, which drops the upper triangle
+Alternatively, it can be created through conversion from `Array`, which drops the upper triangle
 entries and sets them to zero
 ```@repl LowerTriangularMatrices
 M = rand(Float16, 3, 3)
-
 L = LowerTriangularMatrix(M)
+
+M2 = rand(Float16, 3, 3, 5)
+L2 = LowerTriangularArray(M)
 ```
 
-## Indexing `LowerTriangularMatrix`
+## Indexing `LowerTriangularArray`
 
-`LowerTriangularMatrix` supports two types of indexing: 1) by denoting two indices, column and row `[l, m]`
-or 2) by denoting a single index `[lm]`. The double index works as expected
+`LowerTriangularArray` supports two types of indexing: 1) by denoting two indices, column and row `[l, m, ..]`
+or 2) by denoting a single index `[lm, ..]`. The double index works as expected
 ```@repl LowerTriangularMatrices
 L
 
@@ -60,7 +63,7 @@ for j in 1:m, i in j:n
     L[ij] = i+j
 end
 ```
-which loops over all lower triangle entries of `L::LowerTriangularMatrix` and the single
+which loops over all lower triangle entries of `L::LowerTriangularArray` and the single
 index `ij` is simply counted up. However, one could also use `[i, j]` as indices in the
 loop body or to perform any calculation (`i+j` here).
[…]
