### 2024-04-04 796b03a
Examples with PythonPlot again +rename
