### 2024-04-04 7132624
output=true for aquaplanet example
