+++ ext/SpeedyWeatherMakieExt.jl
@@ -8,7 +8,7 @@ function Makie.heatmap(
     title::String = "$(RingGrids.get_nlat(grid))-ring $(typeof(grid))",
 )
     full_grid = RingGrids.interpolate(RingGrids.full_grid(typeof(grid)), grid.nlat_half, grid)
-    heatmap(full_grid; title, kwargs...)
+    heatmap(full_grid; title)
 end
 
 function Makie.heatmap(
