### 2024-04-04 0ac2552
use Makie.heatmap in examples.md
