+++ docs/src/how_to_run_speedy.md
@@ -11,7 +11,7 @@ user may want to adjust are chosen and live in their respective model components
 4. [Run that simulation](@ref run).
 
 In the following we will describe these steps in more detail.
-If you want to start with some examples first, see [Model setups](@ref)
+If you want to start with some examples first, see [Examples](@ref)
 which has easy and more advanced examples of how to set up
 SpeedyWeather.jl to run the simulation you want.
 
@@ -175,6 +175,18 @@ However, from a user perspective all that needs to be done here is
 simulation = initialize!(model)
 ```
 and we have initialized the `ShallowWaterModel` we have defined earlier.
+As `initialize!(model)` also initializes the prognostic (and diagnostic)
+variables, it also initializes the clock in
+`simulation.prognostic_variables.clock`. To initialize with a specific
+time, do
+```@example howto
+simulation = intialize!(model, time=DateTime(2020,5,1))
+simulation.prognostic_variables.clock.time
+```
+to set the time to 1st May, 2020 (but you can also do that manually).
+This time is used by components that depend on time, e.g. the solar
+zenith angle calculation.
+
 After this step you can continue to tweak your model setup but note that
 some model components are immutable, or that your changes may not be
 propagated to other model components that rely on it. But you can, for
