### 2024-04-04 101d043
add docstrings and propagate kwargs...
