### 2024-04-04 84420f9
Most examples with CairoMakie now
