+++ docs/src/examples.md
@@ -112,44 +112,61 @@ data is stored in the folder `/run_0001`. In general we can check this also via
 ```@example galewsky_setup
 id = model.output.id
 ```
-So let's plot that data properly (and not just using UnicodePlots). `$id` in the following just means
+So let's plot that data properly (and not just using UnicodePlots.jl). `$id` in the following just means
 that the string is interpolated to `run_0001` if this is the first unnamed run in your folder.
 ```@example galewsky_setup
-using PythonPlot, NCDatasets
-ioff() # hide
+using NCDatasets
 ds = NCDataset("run_$id/output.nc")
 ds["vor"]
 ```
 Vorticity `vor` is stored as a lon x lat x vert x time array, we may want to look at the first time step,
-which is the end of the previous simulation (time=6days) which we didn't store output for.
+which is the end of the previous simulation (time = 6days) which we didn't store output for.
 ```@example galewsky_setup
 t = 1
 vor = Matrix{Float32}(ds["vor"][:, :, 1, t]) # convert from Matrix{Union{Missing, Float32}} to Matrix{Float32}
 lat = ds["lat"][:]
 lon = ds["lon"][:]
 
-fig, ax = subplots(1, 1, figsize=(10, 6))
-ax.pcolormesh(lon, lat, vor')
-ax.set_xlabel("longitude")
-ax.set_ylabel("latitude")
-ax.set_title("Relative vorticity")
-tight_layout() # hide
-savefig("galewsky1.png", dpi=70) # hide
+using CairoMakie
+heatmap(lon, lat, vor)
+save("galewsky0.png", ans) # hide
+nothing # hide
+```
+![Galewsky jet](galewsky0.png)
+
+You see that in comparison the unicode plot heavily coarse-grains the simulation, well it's unicode after all!
+Here, we have unpacked the netCDF file using [NCDatasets.jl](https://github.com/Alexander-Barth/NCDatasets.jl)
+and then plotted via `heatmap(lon, lat, vor)`. While you can do that to give you more control
+on the plotting, SpeedyWeather.jl also defines an extension for Makie.jl, see [Extensions](@ref)
+because if our matrix `vor` here was an `AbstractGrid` (see [RingGrids](@ref)) then all
+its geographic information (which grid point is where) is encoded in the type. So we can also do
+```@example galwesky_setup
+vor_grid = FullGaussianGrid(vor)
+
+using CairoMakie    # this will load the extension so that Makie can plot grids directly
+heatmap(vor_grid, title="Relative vorticity [1/s]")
+save("galewsky1.png", ans) # hide
 nothing # hide
 ```
 ![Galewsky jet pyplot1](galewsky1.png)
 
-You see that in comparison the unicode plot heavily coarse-grains the simulation, well it's unicode after all!
-And now the last time step, that means time = 12days is
+Note that here you need to know which grid the data comes on (an error is thrown if `FullGaussianGrid(vor)`
+is not size compatible). By default the output will be on the FullGaussianGrid, but if you
+play around with other grids, you'd need to change this here. 
+
+We did want to showcase the usage of [NetCDF output](@ref) here, but from now on
+we will use `heatmap` to plot data on our grids directly, without storing output first.
+So for our current simulation, that means at time = 12 days, vorticity on the grid
+is stored in the diagnostic variables and can be visualised with
 
[…]
