### 2024-04-04 8a1fedd
add GeoMakie to docs, particle advection with GeoMakie
