+++ docs/Project.toml
@@ -1,6 +1,7 @@
 [deps]
 CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+GeoMakie = "db073c08-6b98-4ee5-b6a4-5efafb3259c6"
 NCDatasets = "85f8d34a-cbdd-5861-8df4-14fed0d494ab"
 PythonPlot = "274fc56d-3b97-40fa-a1cd-1b4a50311bf9"
 UnicodePlots = "b8865327-cd53-5732-bb35-84acbb429228"
@@ -8,6 +9,7 @@ UnicodePlots = "b8865327-cd53-5732-bb35-84acbb429228"
 [compat]
 CairoMakie = "0.11"
 Documenter = "0.26, 0.27, 1"
+GeoMakie = "0.6"
 NCDatasets = "0.12, 0.13, 0.14"
 PythonPlot = "1"
 UnicodePlots = "3.3"
