### 2024-04-04 c7b73e5
analysis.md with CairoMakie
