+++ docs/src/radiation.md
@@ -1,1 +1,18 @@
 # Radiation
+
+## Longwave radiation
+
+Currently implemented is
+
+```@example radiation
+using SpeedyWeather
+subtypes(SpeedyWeather.AbstractLongwave)
+```
+
+## Shortwave radiation
+
+Currently implemented is
+
+```@example radiation
+subtypes(SpeedyWeather.AbstractShortwave)
+```
