### 2024-04-04 1bba10c
Add Makie extension (#511)
* Add Makie extension

* use Makie.heatmap in examples.md

* add GeoMakie to docs, particle advection with GeoMakie

* analysis.md with CairoMakie
