+++ docs/src/installation.md
@@ -8,7 +8,7 @@ julia> Pkg.add("SpeedyWeather")
 ```
 or, equivalently, (`]` opens the package manager)
 ```julia
-julia> ] add SpeedyWeather
+julia>] add SpeedyWeather
 ```
 which will automatically install the [latest release](https://github.com/SpeedyWeather/SpeedyWeather.jl/releases)
 and all necessary dependencies. If you run into any troubles please raise an
@@ -20,9 +20,22 @@ julia> Pkg.add(url="https://github.com/SpeedyWeather/SpeedyWeather.jl", rev="mai
 ```
 In a similar manner, other branches can be also installed. You can also type, equivalently,
 ```julia
-julia> ] add https://github.com/SpeedyWeather/SpeedyWeather.jl#main
+julia>] add https://github.com/SpeedyWeather/SpeedyWeather.jl#main
 ```
 
 ## Compatibility with Julia versions
 
 SpeedyWeather.jl requires Julia v1.8 or later. The package is tested on Julia v1.8, 1.9, and 1.10.
+
+## Extensions
+
+SpeedyWeather.jl has a weak dependency on
+
+- [Makie.jl](https://github.com/MakieOrg/Makie.jl) to extend `Makie.heatmap`
+
+This is an
+[extension](https://pkgdocs.julialang.org/v1.10/creating-packages/#Conditional-loading-of-code-in-packages-(Extensions)),
+meaning that this functionality is only loaded from SpeedyWeather when `using Makie`
+(or its backends CairoMakie.jl, GLMakie.jl, ...). Hence, if you want to make use of this
+extension (some [Examples](@ref) show this) you need to install Makie.jl manually.
+
