+++ docs/src/radiation.md
@@ -1,6 +1,6 @@
 # Radiation
 
-## Longwave radiation
+## Longwave radiation implementations
 
 Currently implemented is
 
@@ -9,6 +9,55 @@ using SpeedyWeather
 subtypes(SpeedyWeather.AbstractLongwave)
 ```
 
+## Uniform cooling
+
+Following Paulius and Garner[^PG06], the uniform cooling of the atmosphere
+is defined as 
+
+```math
+\frac{\partial T}{\partial t} = \begin{cases} - \tau^{-1}\quad&\text{for}\quad T > T_{min} \\
+                                            \frac{T_{strat} - T}{\tau_{strat}} \quad &\text{else.} \end{cases}
+```
+with ``\tau = 16~h`` resulting in a cooling of -1.5K/day for most of the atmosphere,
+except below temperatures of ``T_{min} = 207.5~K`` in the stratosphere where a
+relaxation towards ``T_{strat} = 200~K`` with a time scale of ``\tau_{strat} = 5~days``
+is present.
+
+## Jeevanjee radiation
+
+Jeevanjee and Zhou [^JZ22] (eq. 2) define a longwave radiative flux ``F`` for atmospheric cooling
+as (following Seeley and Wordsworth [^SW23], eq. 1)
+
+```math
+\frac{dF}{dT} = α*(T_t - T)
+```
+
+The flux ``F`` (in ``W/m^2/K``) is a vertical upward flux between two layers (vertically adjacent)
+of temperature difference ``dT``. The change of this flux across layers depends on the temperature
+``T`` and is a relaxation term towards a prescribed stratospheric temperature ``T_t = 200~K`` with
+a radiative forcing constant ``\alpha = 0.025 W/m^2/K^2``. Two layers of identical temperatures
+``T_1 = T_2`` would have no net flux between them, but a layer below at higher temperature would
+flux into colder layers above as long as its temperature ``T > T_t``. This flux is applied above
+the lowermost layer and above, leaving the surface fluxes unchanged. The uppermost layer is tied
+to ``T_t`` through a relaxation at time scale ``\tau = 6~h``
+
+```math
+\frac{\partial T}{\partial t} = \frac{T_t - T}{\tau}
+```
+
+The flux ``F`` is converted to temperature tendencies at layer ``k`` via
+
+```math
+\frac{\partial T_k}{\partial t} = (F_{k+1/2} - F_{k-1/2})\frac{g}{\Delta p c_p}
+```
+
+The term in parentheses is the absorbed flux in layer ``k`` of the upward
+flux from below at interface ``k+1/2`` (``k`` increases downwards, see
+[Vertical coordinates and resolution](@ref) and [Sigma coordinates](@ref)).
+``\Delta p = p_{k+1/2} - p_{k-1/2}`` is the pressure thickness of layer ``k``,
+gravity ``g`` and heat capacity ``c_p``.
+
+
 ## Shortwave radiation
 
 Currently implemented is
@@ -16,3 +65,11 @@ Currently implemented is
 ```@example radiation
 subtypes(SpeedyWeather.AbstractShortwave)
 ```
+
+## References
+
+[^PG06]: Paulius and Garner, 2006. JAS. DOI:[10.1175/JAS3705.1](https://doi.org/10.1175/JAS3705.1)
+
+[^SW23]: Seeley, J. T. & Wordsworth, R. D. Moist Convection Is Most Vigorous at Intermediate Atmospheric Humidity. Planet. Sci. J. 4, 34 (2023). DOI:[10.3847/PSJ/acb0cb](https://doi.org/10.3847/PSJ/acb0cb)
+
[…]
