### 2024-04-05 c8a38e0
Split into 2D/3D examples
