+++ docs/src/examples_2D.md
@@ -1,8 +1,10 @@
-# Examples
+# [Examples 2D](@id Examples)
 
-The following is a collection of model setups, starting with an easy setup
+The following is a collection of example model setups, starting with an easy setup
 of the [Barotropic vorticity equation](@ref barotropic_vorticity_model) and
-continuing with more complicated setups.
+continuing with the [shallow water model](@ref shallow_water_model).
+
+See also [Examples 3D](@ref) for examples with the primitive equation models.
 
 ## 2D turbulence on a non-rotating sphere
 
@@ -304,7 +306,7 @@ Hb = model.orography.orography
 η = simulation.diagnostic_variables.surface.pres_grid
 h = @. η + H - Hb   # @. to broadcast grid + scalar - grid
 
-heatmap(h, title="Dynamic layer thickness h")
+heatmap(h, title="Dynamic layer thickness h", colormap=:oslo)
 save("gravity_waves.png", ans) # hide
 nothing # hide
 ```
@@ -312,99 +314,6 @@ nothing # hide
 
 Can you spot the Himalayas or the Andes?
 
-## Jablonowski-Williamson baroclinic wave
-
-```@example jablonowski
-using SpeedyWeather
-spectral_grid = SpectralGrid(trunc=31, nlev=8, Grid=FullGaussianGrid, dealiasing=3)
-
-orography = ZonalRidge(spectral_grid)
-initial_conditions = InitialConditions(
-    vordiv = ZonalWind(),
-    temp = JablonowskiTemperature(),
-    pres = ZeroInitially())
-
-model = PrimitiveDryModel(; spectral_grid, orography, initial_conditions, physics=false)
-simulation = initialize!(model)
-model.feedback.verbose = false # hide
-run!(simulation, period=Day(9))
-nothing # hide
-```
-
-The Jablonowski-Williamson baroclinic wave test case[^JW06] using the
-[Primitive equation model](@ref primitive_equation_model) particularly the dry model,
-as we switch off all physics with `physics=false`.
-We want to use 8 vertical levels, and a lower resolution of T31 on a
-[full Gaussian grid](@ref FullGaussianGrid).
-The Jablonowski-Williamson initial conditions are `ZonalWind` for vorticity and divergence
-(curl and divergence of ``u, v``), `JablonowskiTemperature` for temperature and
-`ZeroInitially` for pressure. The orography is just a `ZonalRidge`.
-There is no forcing and the initial conditions are baroclinically unstable which kicks
-off a wave propagating eastward. This wave becomes obvious when visualised with
-
-```@example jablonowski
-using CairoMakie
-
-vor = simulation.diagnostic_variables.layers[end].grid_variables.vor_grid
-heatmap(vor, title="Surface relative vorticity")
-save("jablonowski.png", ans) # hide
-nothing # hide
-```
-![Jablonowski pyplot](jablonowski.png)
-
-## Aquaplanet
-
-```@example aquaplanet
-using SpeedyWeather
-
-# components
-spectral_grid = SpectralGrid(trunc=31, nlev=5)
-ocean = AquaPlanet(spectral_grid, temp_equator=302, temp_poles=273)
-land_sea_mask = AquaPlanetMask(spectral_grid)
-orography = NoOrography(spectral_grid)
-
-# create model, initialize, run
-model = PrimitiveWetModel(; spectral_grid, ocean, land_sea_mask, orography)
-simulation = initialize!(model)
-model.feedback.verbose = false # hide
-run!(simulation, period=Day(50))
-nothing # hide
-```
-
-Here we have defined an aquaplanet simulation by
-- creating an `ocean::AquaPlanet`. This will use constant sea surface temperatures that only vary with latitude.
-- creating a `land_sea_mask::AquaPlanetMask` this will use a land-sea mask with `false`=ocean everywhere.
[…]
