+++ src/physics/surface_fluxes.jl
@@ -1,28 +1,29 @@
-abstract type AbstractSurfaceWind <: AbstractParameterization end
 abstract type AbstractSurfaceThermodynamics <: AbstractParameterization end
-abstract type AbstractSurfaceHeat <: AbstractParameterization end
-abstract type AbstractEvaporation <: AbstractParameterization end
+abstract type AbstractSurfaceWind <: AbstractParameterization end
+abstract type AbstractSurfaceSensibleHeat <: AbstractParameterization end
+abstract type AbstractSurfaceEvaporation <: AbstractParameterization end
 
+# defines the order in which they are called und unpacks to dispatch
 function surface_fluxes!(column::ColumnVariables, model::PrimitiveEquation)
 
     # get temperature, humidity and density at surface
-    surface_thermodynamics!(column, model.surface_thermodynamics, model.atmosphere, model)
+    surface_thermodynamics!(column, model.surface_thermodynamics, model)
 
     # also calculates surface wind speed necessary for other fluxes too
-    surface_wind_stress!(column, model.surface_wind)
+    surface_wind_stress!(column, model.surface_wind, model)
 
-    # now call other heat and humidity fluxes
-    sensible_heat_flux!(column, model.surface_heat_flux, model.atmosphere)
-    evaporation!(column, model)
+    # now call other heat (wet and dry) and humidity fluxes (PrimitiveWet only)
+    sensible_heat_flux!(column, model.surface_heat_flux, model)
+    model isa PrimitiveWet && surface_evaporation!(column, model.surface_evaporation, model)
 end
 
+## SURFACE THERMODYNAMICS
 export SurfaceThermodynamicsConstant
-struct SurfaceThermodynamicsConstant{NF<:AbstractFloat} <: AbstractSurfaceThermodynamics end
-SurfaceThermodynamicsConstant(SG::SpectralGrid) = SurfaceThermodynamicsConstant{SG.NF}()
+struct SurfaceThermodynamicsConstant <: AbstractSurfaceThermodynamics end
+SurfaceThermodynamicsConstant(SG::SpectralGrid) = SurfaceThermodynamicsConstant()
 
 function surface_thermodynamics!(   column::ColumnVariables,
                                     ::SurfaceThermodynamicsConstant,
-                                    atmosphere::AbstractAtmosphere,
                                     model::PrimitiveWet)
 
     # surface value is same as lowest model level
@@ -30,7 +31,7 @@ function surface_thermodynamics!(   column::ColumnVariables,
     column.surface_humid = column.humid[end]    # humidity at surface is the same as 
 
     # surface air density via virtual temperature
-    (; R_dry) = atmosphere
+    (; R_dry) = model.atmosphere
     Tᵥ = column.temp_virt[column.nlev]
     column.surface_air_density = column.pres[end]/(R_dry*Tᵥ)
 end
@@ -44,6 +45,13 @@ function surface_thermodynamics!(   column::ColumnVariables,
     column.surface_temp = column.temp[end]   # todo use constant POTENTIAL temperature
     column.surface_air_density = column.pres[end]/(R_dry*column.surface_temp)
 end
+
+## WIND STRESS
+export NoSurfaceWind
+struct NoSurfaceWind <: AbstractSurfaceWind end
+NoSurfaceWind(::SpectralGrid) = NoSurfaceWind()
+initialize!(::NoSurfaceWind, ::PrimitiveEquation) = nothing
[…]
