### 2024-04-05 c90d5b8
Radiation documentation (#513)
* Radiation docs

* Jeevanjee radiation described
