+++ docs/src/examples_2D.md
@@ -139,9 +139,12 @@ nothing # hide
 You see that in comparison the unicode plot heavily coarse-grains the simulation, well it's unicode after all!
 Here, we have unpacked the netCDF file using [NCDatasets.jl](https://github.com/Alexander-Barth/NCDatasets.jl)
 and then plotted via `heatmap(lon, lat, vor)`. While you can do that to give you more control
-on the plotting, SpeedyWeather.jl also defines an extension for Makie.jl, see [Extensions](@ref)
-because if our matrix `vor` here was an `AbstractGrid` (see [RingGrids](@ref)) then all
-its geographic information (which grid point is where) is encoded in the type. So we can also do
+on the plotting, SpeedyWeather.jl also defines an extension for Makie.jl, see [Extensions](@ref).
+Because if our matrix `vor` here was an `AbstractGrid` (see [RingGrids](@ref)) then all
+its geographic information (which grid point is where) would directly be encoded in the type.
+From the netCDF file you need to use the longitude and latitude dimensions.
+
+So we can also just do
 ```@example galewsky_setup
 vor_grid = FullGaussianGrid(vor)
 
@@ -154,7 +157,8 @@ nothing # hide
 
 Note that here you need to know which grid the data comes on (an error is thrown if `FullGaussianGrid(vor)`
 is not size compatible). By default the output will be on the FullGaussianGrid, but if you
-play around with other grids, you'd need to change this here. 
+play around with other grids, you'd need to change this here,
+see [NetCDF output on other grids](@ref output_grid).
 
 We did want to showcase the usage of [NetCDF output](@ref) here, but from now on
 we will use `heatmap` to plot data on our grids directly, without storing output first.
@@ -162,7 +166,6 @@ So for our current simulation, that means at time = 12 days, vorticity on the gr
 is stored in the diagnostic variables and can be visualised with
 
 ```@example galewsky_setup
-t = ds.dim["time"]
 vor = simulation.diagnostic_variables.layers[1].grid_variables.vor_grid
 heatmap(vor, title="Relative vorticity [1/s]")
 save("galewsky2.png", ans) # hide
@@ -178,9 +181,9 @@ Let's try it out! We create an `EarthOrography` struct like so
 orography = EarthOrography(spectral_grid)
 ```
 
-It will read the orography from file as shown, and there are some smoothing options too, but let's not change them.
-Same as before, create a model, initialize into a simulation, run. This time directly for 12 days so that we can
-compare with the last plot
+It will read the orography from file as shown (only at `initialize!(model)`), and there are some smoothing
+options too, but let's not change them. Same as before, create a model, initialize into a simulation, run.
+This time directly for 12 days so that we can compare with the last plot
 
 ```@example galewsky_setup
 model = ShallowWaterModel(; spectral_grid, orography, initial_conditions)
