### 2024-04-05 568af0f
SpeedyTransforms docs with Makie too
