+++ docs/src/installation.md
@@ -8,7 +8,7 @@ julia> Pkg.add("SpeedyWeather")
 ```
 or, equivalently, (`]` opens the package manager)
 ```julia
-julia>] add SpeedyWeather
+julia> ] add SpeedyWeather
 ```
 which will automatically install the [latest release](https://github.com/SpeedyWeather/SpeedyWeather.jl/releases)
 and all necessary dependencies. If you run into any troubles please raise an
@@ -20,7 +20,7 @@ julia> Pkg.add(url="https://github.com/SpeedyWeather/SpeedyWeather.jl", rev="mai
 ```
 In a similar manner, other branches can be also installed. You can also type, equivalently,
 ```julia
-julia>] add https://github.com/SpeedyWeather/SpeedyWeather.jl#main
+julia> ] add https://github.com/SpeedyWeather/SpeedyWeather.jl#main
 ```
 
 ## Compatibility with Julia versions
