### 2024-04-05 35314c0
Held-Suarez example, NoTypes for surface fluxes (#514)
* Split into 2D/3D examples

* NoTypes for surface fluxes

* Held-Suarez example added

* SpeedyTransforms docs with Makie too
