+++ src/LowerTriangularMatrices/lower_triangular_matrix.jl
@@ -15,7 +15,7 @@ struct LowerTriangularArray{T, N, ArrayType <: AbstractArray{T}} <: AbstractArra
     m::Int              # number of rows
     n::Int              # number of columns
 
-    LowerTriangularArray{T, N, ArrayType}(data, m, n) where {T, N, ArrayType<:AbstractArray} = check_lta_input_array(data, m, n, N) ? new(data, m, n) : error(lta_error_message(data, m, n, T, N, ArrayType))
+    LowerTriangularArray{T, N, ArrayType}(data, m, n) where {T, N, ArrayType<:AbstractArray{T}} = check_lta_input_array(data, m, n, N) ? new(data, m, n) : error(lta_error_message(data, m, n, T, N, ArrayType))
 end
 
 check_lta_input_array(data, m, n, N) = (ndims(data) == N-1) & (length(data) == prod(size(data)[2:end]) * nonzeros(m, n)) 
@@ -25,10 +25,10 @@ function lta_error_message(data, m, n, T, N, ArrayType)
     "$(size(data))-sized $(typeof(data)) cannot be used to create a $size_tuple LowerTriangularArray{$T,$N,$ArrayType}"
 end 
 
-const LowerTriangularMatrix{T, ArrayType} = LowerTriangularArray{T, 2, ArrayType}
+const LowerTriangularMatrix{T} = LowerTriangularArray{T, 2, Vector{T}}
 
 LowerTriangularArray(data::ArrayType, m::Integer, n::Integer) where {T, N, ArrayType<:AbstractArray{T,N}} = LowerTriangularArray{T, N+1, ArrayType}(data, m, n)
-LowerTriangularMatrix(data::ArrayType, m::Integer, n::Integer) where {T, ArrayType<:AbstractVector{T}} = LowerTriangularMatrix{T, ArrayType}(data, m, n)
+LowerTriangularMatrix(data::Vector{T}, m::Integer, n::Integer) where T = LowerTriangularMatrix{T}(data, m, n)
 
 # SIZE ETC
 Base.length(L::LowerTriangularArray) = length(L.data)  # define length as number of non-zero elements
@@ -47,9 +47,8 @@ for zeros_or_ones in (:zeros, :ones)
             return LowerTriangularArray($zeros_or_ones(T, nonzeros(m, n), I...), m, n)
         end
         
-        # use Float64 and Vector as default if type T and ArrayType not provided
-        Base.$zeros_or_ones(::Type{LowerTriangularMatrix}, m::Integer, n::Integer) = $zeros_or_ones(LowerTriangularMatrix{Float64, Vector{Float64}}, m, n)
-        Base.$zeros_or_ones(::Type{LowerTriangularMatrix{T}}, m::Integer, n::Integer) where T = $zeros_or_ones(LowerTriangularMatrix{T, Vector{T}}, m, n)
+        # use Float64 as default if type T is not provided
+        Base.$zeros_or_ones(::Type{LowerTriangularMatrix}, m::Integer, n::Integer) = $zeros_or_ones(LowerTriangularMatrix{Float64}, m, n)
     end
 end
 
@@ -62,8 +61,8 @@ end
 
 LowerTriangularArray{T,N,ArrayType}(::UndefInitializer, size::S) where {T,N,ArrayType<:AbstractArray{T}, S<:Tuple} = LowerTriangularArray{T,N,ArrayType}(undef, size...)
    
-function LowerTriangularMatrix{T,ArrayType}(::UndefInitializer, m::Integer, n::Integer) where {T,ArrayType<:AbstractArray{T}}
-    return LowerTriangularMatrix(ArrayType(undef, nonzeros(m, n)), m, n)
[…]
