### 2024-04-08 836592b
indexing with leading ":" returns LowerTriangularArray
