+++ src/LowerTriangularMatrices/LowerTriangularMatrices.jl
@@ -5,7 +5,6 @@ import Adapt, KernelAbstractions
 import UnicodePlots
 import LinearAlgebra: tril!
 import GPUArrays
-import Base: view
 
 export LowerTriangularMatrix, LowerTriangularArray, eachharmonic
 # export plot
