### 2024-04-08 e374352
remove second spectral_trunction! version
