+++ src/SpeedyTransforms/spectral_truncation.jl
@@ -85,26 +85,6 @@ If `trunc` is larger than the implicit truncation in `alms` obtained from its si
 is automatically called instead, returning `alms_interp`, a coefficient matrix that is larger than `alms`
 with padded zero coefficients."""
 function spectral_truncation(   ::Type{NF},                     # number format NF (can be complex)
-                                alms::LowerTriangularMatrix,    # spectral field to be truncated
-                                ltrunc::Integer,                # truncate to max degree ltrunc
-                                mtrunc::Integer,                # truncate to max order mtrunc
-                                ) where NF
-    
-    lmax, mmax = size(alms) .- 1     # 0-based degree l, order m of the spherical harmonics
-    
-    # interpolate to higher resolution if output larger than input
-    (ltrunc > lmax || mtrunc > mmax) && return spectral_interpolation(NF, alms, ltrunc, mtrunc)
-
-    # preallocate new (smaller) array
-    alms_trunc = zeros(LowerTriangularMatrix{NF}, ltrunc+1, mtrunc+1)  
-
-    # copy data over, copyto! copies the largest matching subset of harmonics
-    copyto!(alms_trunc, alms)
-    return alms_trunc
-end
-
-# TODO: temporary version for N-dim LWA, merge with above one when all tests are done.
-function spectral_truncation(   ::Type{NF},                     # number format NF (can be complex)
     alms::LowerTriangularArray{T,N},    # spectral field to be truncated
     ltrunc::Integer,                # truncate to max degree ltrunc
     mtrunc::Integer,                # truncate to max order mtrunc
