### 2024-04-10 2e436d5
First ideas for AbstractGridArray
