### 2024-04-11 a2818b4
Remove type constraints for model components during construction
