### 2024-04-11 9a5fb9e
rm default for spectral_grid in model construction
