+++ .github/workflows/CI.yml
@@ -14,7 +14,7 @@ jobs:
       matrix:
         version:
           # - '1.8'
-          # - '1.9'
+          - '1.9'
           - '1.10'
         os:
           - ubuntu-latest
