+++ .github/workflows/CI.yml
@@ -13,7 +13,6 @@ jobs:
     strategy:
       matrix:
         version:
-          # - '1.8'
           - '1.9'
           - '1.10'
         os:
