### 2024-04-12 2c1786e
AbstractGridArray constructors
