+++ src/RingGrids/general.jl
@@ -1,195 +1,183 @@
-abstract type AbstractGridArray{T, N, ArrayType <: AbstractArray{T, N}} <: AbstractArray{T,N} end
-abstract type AbstractFullGridArray{T, N, ArrayType <: AbstractArray{T, N}} <: AbstractGridArray{T,N} end
-abstract type AbstractReducedGridArray{T, N, ArrayType <: AbstractArray{T, N}} <: AbstractGridArray{T,N} end
+abstract type AbstractGridArray{T, N, ArrayType <: AbstractArray{T, N}} <: AbstractArray{T, N} end
+const AbstractGrid{T} = AbstractGridArray{T, 1, Vector{T}}
 
-# all AbstractGrids have their grid points stored in a vector field `data`
-# propagate length, size, getindex, setindex! for that
+# nonparametric_type(G::Type{<:AbstractGridArray}) = @warn "Please define nonparametric_type(::$G)"
+# horizontal_grid_type(G::Type{<:AbstractGridArray}) = @warn "Please define horizontal_grid_type(::$G)"
+
+# LENGTH SIZE
 Base.length(G::AbstractGridArray) = length(G.data)
 Base.size(G::AbstractGridArray) = size(G.data)
 Base.sizeof(G::AbstractGridArray) = sizeof(G.data)
 
-@inline Base.getindex(G::AbstractGridArray, ijk::Integer...) = getindex(G.data,ijk...)
-@inline Base.getindex(G::AbstractGridArray, r::AbstractRange) = get_index(G.data, r)
-
-
-
-@inline function Base.getindex(L::AbstractGridArray, col::Colon, I...)
-    return AbstractGridArray(getindex(L.data, col, I...), L.m, L.n)
-end
-
-
-
-@inline Base.getindex(L::AbstractGridArray, r::AbstractRange) = getindex(L.data,r)
-@inline Base.getindex(L::AbstractGridArray, r::AbstractRange, I...) = getindex(L.data, r, I...)
-@inline Base.getindex(L::AbstractGridArray, i::Integer) = getindex(L.data,i)
-@inline Base.getindex(L::AbstractGridArray, I::CartesianIndex) = getindex(L, Tuple(I)...)
-
-
-@inline Base.setindex!(G::AbstractGrid, x::AbstractVector, r::AbstractRange) = setindex!(G.data, x, r)
-@inline Base.setindex!(G::AbstractGrid, x, k::Integer) = setindex!(G.data, x, k)
-
-
-function grids_match(A::AbstractGridArray, B::AbstractGridArray)
-    length(A) == length(B) && return _grids_match(typeof(A), typeof(B))
-    return false
-end
-
-function _grids_match(A::Type{<:AbstractGridArray}, B::Type{<:AbstractGridArray})
-    return nonparametric_type(A) == nonparametric_type(B)
-end
-
-"""
-    abstract type AbstractGrid{T} <: AbstractVector{T} end
-
-The abstract supertype for all spatial grids on the sphere supported by SpeedyWeather.jl.
-Every new grid has to be of the form
-
-    abstract type AbstractGridClass{T} <: AbstractGrid{T} end
-    struct MyNewGrid{T} <: AbstractGridClass{T}
-        data::Vector{T}     # all grid points unravelled into a vector
-        nlat_half::Int      # resolution: latitude rings on one hemisphere (Equator incl)
-    end
-    
-`MyNewGrid` should belong to a grid class like `AbstractFullGrid`, `AbstractOctahedralGrid` or
-`AbstractHEALPixGrid` (that already exist but you may introduce a new class of grids) that share
-certain features such as the number of longitude points per latitude ring and indexing, but may
[…]

[…]
