### 2024-04-14 14453c7
similar, eachgrid for AbstractGridArray
