+++ src/RingGrids/grids/octahedral_clenshaw.jl
@@ -1,124 +1,4 @@
 """
-    abstract type AbstractOctahedralGrid{T} <: AbstractGrid{T} end
-
-An `AbstractOctahedralGrid` is a horizontal grid with 16+4i longitude
-points on the latitude ring i starting with i=1 around the pole.
-Different latitudes can be used, Gaussian latitudes, equi-angle latitdes, or others."""
-abstract type AbstractOctahedralGrid{T} <: AbstractGrid{T} end
-
-get_nlon_max(::Type{<:AbstractOctahedralGrid}, nlat_half::Integer) = nlon_octahedral(nlat_half)
-
-function get_nlon_per_ring(Grid::Type{<:AbstractOctahedralGrid}, nlat_half::Integer, j::Integer)
-    nlat = get_nlat(Grid, nlat_half)
-    @assert 0 < j <= nlat "Ring $j is outside O$nlat_half grid."
-    j = j > nlat_half ? nlat - j + 1 : j      # flip north south due to symmetry
-    return nlon_octahedral(j)
-end
-
-function get_colatlons(Grid::Type{<:AbstractOctahedralGrid}, nlat_half::Integer)
-    
-    colat = get_colat(Grid, nlat_half)
-    nlat = get_nlat(Grid, nlat_half)
-    
-    npoints = get_npoints(Grid, nlat_half)
-    colats = zeros(npoints)                 # preallocate arrays
-    lons = zeros(npoints)
-
-    ij = 1                                  # continuous index
-    for j in 1:nlat                         # populate arrays ring by ring
-        nlon = get_nlon_per_ring(Grid, nlat_half, j)
-        lon = collect(0:2π/nlon:2π-π/nlon)
-
-        colats[ij:ij+nlon-1] .= colat[j]
-        lons[ij:ij+nlon-1] .= lon
-
-        ij += nlon
-    end
-
-    return colats, lons
-end
-
-function each_index_in_ring(Grid::Type{<:AbstractOctahedralGrid},
-                            j::Integer,                     # ring index north to south
-                            nlat_half::Integer)             # resolution param
-
-    nlat = get_nlat(Grid, nlat_half)
-    @boundscheck 0 < j <= nlat || throw(BoundsError)        # ring index valid?
-    if j <= nlat_half                                       # northern hemisphere incl Equator
-        index_1st = 2j*(j+7) - 15                           # first in-ring index i
-        index_end = 2j*(j+9)                                # last in-ring index i
-    else                                                    # southern hemisphere excl Equator
-        j = nlat - j + 1                                    # mirror ring index around Equator
-        n = get_npoints(Grid, nlat_half) + 1                 # number of grid points + 1
-        index_1st = n - 2j*(j+9)                            # count backwards
-        index_end = n - (2j*(j+7) - 15)
-    end
-    return index_1st:index_end                              # range of i's in ring
-end
-
-function each_index_in_ring!(   rings::Vector{<:UnitRange{<:Integer}},
-                                Grid::Type{<:AbstractOctahedralGrid},
-                                nlat_half::Integer) # resolution param
-
-    nlat = length(rings)
-    @boundscheck nlat == get_nlat(Grid, nlat_half) || throw(BoundsError)
-
-    index_end = 0
-    @inbounds for j in 1:nlat_half                  # North incl Eq only
-        index_1st = index_end + 1                   # 1st index is +1 from prev ring's last index
-        index_end += 16 + 4j                        # add number of grid points per ring
[…]
