### 2024-04-16 52c9d27
OctahedralGaussian and ClenshawArrays
