+++ src/RingGrids/grids/octahedral_clenshaw.jl
@@ -1,40 +1,64 @@
-"""
-    G = OctahedralClenshawGrid{T}
+struct OctahedralClenshawArray{T, N, ArrayType <: AbstractArray{T, N}} <: AbstractReducedGridArray{T, N, ArrayType}
+    data::ArrayType                 # data array, ring by ring, north to south
+    nlat_half::Int                  # number of latitudes on one hemisphere
+    rings::Vector{UnitRange{Int}}   # TODO make same array type as data?
 
-An Octahedral Clenshaw grid that uses `nlat` equi-spaced latitudes. Like FullClenshawGrid, the central
-latitude ring is on the Equator. Like OctahedralGaussianGrid, the number of longitude points per
-latitude ring decreases towards the poles. Starting with 20 equi-spaced longitude points (starting at 0˚E)
-on the rings around the poles, each latitude ring towards the equator has consecuitively 4 more points,
-one for each face of the octahedron. E.g. 20, 24, 28, 32, ...nlon-4, nlon, nlon, nlon-4, ..., 32, 28, 24, 20.
-The maximum number of longitue points is `nlon`. The values of all grid points are stored in a vector
-field `v` that unravels the data 0 to 360˚, then ring by ring, which are sorted north to south."""
-struct OctahedralClenshawGrid{T} <: AbstractOctahedralGrid{T}
-    data::Vector{T}    # data vector, ring by ring, north to south
-    nlat_half::Int  # number of latitudes on one hemisphere (incl Equator)
-
-    # check that `nlat_half` match the vector `v` length
-    OctahedralClenshawGrid{T}(data::AbstractVector, nlat_half::Integer) where T = length(data) == npoints_octahedral(nlat_half, true) ?
-    new(data, nlat_half) : error("$(length(data))-element Vector{$(eltype(data))}"*
-    "cannot be used to create a O$(nlat_half) OctahedralClenshawGrid{$T}.")
+    OctahedralClenshawArray(data::A, nlat_half, rings) where {A <: AbstractArray{T, N}} where {T, N} =
+        check_inputs(data, nlat_half, rings, OctahedralClenshawArray) ?
+        new{T, N, A}(data, nlat_half, rings) :
+        error_message(data, nlat_half, rings, OctahedralClenshawArray, T, N, A)
 end
 
-nonparametric_type(::Type{<:OctahedralClenshawGrid}) = OctahedralClenshawGrid
+# TYPES
+const OctahedralClenshawGrid{T} = OctahedralClenshawArray{T, 1, Vector{T}}
+nonparametric_type(::Type{<:OctahedralClenshawArray}) = OctahedralClenshawArray
+horizontal_grid_type(::Type{<:OctahedralClenshawArray}) = OctahedralClenshawGrid
+full_grid(::Type{<:OctahedralClenshawArray}) = FullClenshawArray
 
-# infer nlat_half from data vector length, infer parametric type from eltype of data
-OctahedralClenshawGrid{T}(data::AbstractVector) where T = OctahedralClenshawGrid{T}(data,
-                                                        nlat_half_octahedral(length(data), true))
-OctahedralClenshawGrid(data::AbstractVector, n::Integer...) = OctahedralClenshawGrid{eltype(data)}(data, n...)
+# SIZE
+nlat_odd(::Type{<:OctahedralClenshawArray}) = true
+npoints_pole(::Type{<:OctahedralClenshawArray}) = 0
+npoints_added_per_ring(::Type{<:OctahedralClenshawArray}) = 4
 
-nlat_odd(::Type{<:OctahedralClenshawGrid}) = true
[…]
