+++ src/RingGrids/grids/octahedral_gaussian.jl
@@ -1,94 +1,74 @@
-"""
-    G = OctahedralGaussianGrid{T}
+struct OctahedralGaussianArray{T, N, ArrayType <: AbstractArray{T, N}} <: AbstractReducedGridArray{T, N, ArrayType}
+    data::ArrayType                 # data array, ring by ring, north to south
+    nlat_half::Int                  # number of latitudes on one hemisphere
+    rings::Vector{UnitRange{Int}}   # TODO make same array type as data?
 
-An Octahedral Gaussian grid that uses `nlat` Gaussian latitudes, but a decreasing number of longitude
-points per latitude ring towards the poles. Starting with 20 equi-spaced longitude points (starting at 0˚E)
-on the rings around the poles, each latitude ring towards the equator has consecuitively 4 more points,
-one for each face of the octahedron. E.g. 20, 24, 28, 32, ...nlon-4, nlon, nlon, nlon-4, ..., 32, 28, 24, 20.
-The maximum number of longitue points is `nlon`. The values of all grid points are stored in a vector
-field `v` that unravels the data 0 to 360˚, then ring by ring, which are sorted north to south."""
-struct OctahedralGaussianGrid{T} <: AbstractOctahedralGrid{T}
-    data::Vector{T}    # data vector, ring by ring, north to south
-    nlat_half::Int  # number of latitudes on one hemisphere
-
-    # check that `nlat_half` match the vector `v` length
-    OctahedralGaussianGrid{T}(data::AbstractVector, nlat_half::Integer) where T = length(data) == npoints_octahedral(nlat_half, false) ?
-    new(data, nlat_half) : error("$(length(data))-element Vector{$(eltype(data))}"*
-    "cannot be used to create a O$(nlat_half) OctahedralGaussianGrid{$T}.")
+    OctahedralGaussianArray(data::A, nlat_half, rings) where {A <: AbstractArray{T, N}} where {T, N} =
+        check_inputs(data, nlat_half, rings, OctahedralGaussianArray) ?
+        new{T, N, A}(data, nlat_half, rings) :
+        error_message(data, nlat_half, rings, OctahedralGaussianArray, T, N, A)
 end
 
-nonparametric_type(::Type{<:OctahedralGaussianGrid}) = OctahedralGaussianGrid
+# TYPES
+const OctahedralGaussianGrid{T} = OctahedralGaussianArray{T, 1, Vector{T}}
+nonparametric_type(::Type{<:OctahedralGaussianArray}) = OctahedralGaussianArray
+horizontal_grid_type(::Type{<:OctahedralGaussianArray}) = OctahedralGaussianGrid
+full_grid(::Type{<:OctahedralGaussianArray}) = FullGaussianArray
 
-# number of points and longitudes per ring on the octahedral grid
-npoints_octahedral(nlat_half::Integer, nlat_oddp::Bool) =
-    nlat_oddp ? max(0, 4nlat_half^2 + 32nlat_half - 16) : 4nlat_half^2 + 36nlat_half # max(0, ...) needed to avoid negative array size when nlat_half==0
-nlat_half_octahedral(npoints::Integer, nlat_oddp::Bool) =
-    nlat_oddp ? round(Int, -4+sqrt(20 + npoints/4)) : round(Int, -9/2+sqrt((9/2)^2 + npoints/4))  # inverse
-nlon_octahedral(j::Integer) = 16+4j
+# SIZE
+nlat_odd(::Type{<:OctahedralGaussianArray}) = false
+npoints_pole(::Type{<:OctahedralGaussianArray}) = 16
+npoints_added_per_ring(::Type{<:OctahedralGaussianArray}) = 4
 
-# infer nside from data vector length, infer parametric type from eltype of data
[…]
