+++ src/RingGrids/grids/healpix.jl
@@ -1,54 +1,63 @@
-"""
-    abstract type AbstractHEALPixGrid{T} <: AbstractGrid{T} end
+struct HEALPixArray{T, N, ArrayType <: AbstractArray{T, N}} <: AbstractReducedGridArray{T, N, ArrayType}
+    data::ArrayType                 # data array, ring by ring, north to south
+    nlat_half::Int                  # number of latitudes on one hemisphere
+    rings::Vector{UnitRange{Int}}   # TODO make same array type as data?
 
-An `AbstractHEALPixGrid` is a horizontal grid similar to the standard HEALPixGrid,
-but different latitudes can be used, the default HEALPix latitudes or others."""
-abstract type AbstractHEALPixGrid{T} <: AbstractGrid{T} end
-
-npoints_healpix(nlat_half::Integer) = 3*nlat_half^2
-nside_healpix(nlat_half::Integer) = nlat_half÷2
-nlat_half_healpix(npoints::Integer) = round(Int, sqrt(npoints/3))  # inverse of npoints_healpix
-nlon_healpix(nlat_half::Integer, j::Integer) = min(4j, 2nlat_half, 8nlat_half-4j)
-nlon_max_healpix(nlat_half::Integer) = 2nlat_half
-
-nlat_odd(::Type{<:AbstractHEALPixGrid}) = true
-get_nlon_max(::Type{<:AbstractHEALPixGrid}, nlat_half::Integer) = nlon_max_healpix(nlat_half)
-
-function get_nlon_per_ring(G::Type{<:AbstractHEALPixGrid}, nlat_half::Integer, j::Integer)
-    nlat = get_nlat(G, nlat_half)
-    @assert 0 < j <= nlat "Ring $j is outside H$nlat_half grid."
-    return nlon_healpix(nlat_half, j)
+    HEALPixArray(data::A, nlat_half, rings) where {A <: AbstractArray{T, N}} where {T, N} =
+        check_inputs(data, nlat_half, rings, HEALPixArray) ?
+        new{T, N, A}(data, nlat_half, rings) :
+        error_message(data, nlat_half, rings, HEALPixArray, T, N, A)
 end
 
-get_npoints(::Type{<:AbstractHEALPixGrid}, nlat_half::Integer) = npoints_healpix(nlat_half)
+## TYPES
+const HEALPixGrid{T} = HEALPixArray{T, 1, Vector{T}}
+nonparametric_type(::Type{<:HEALPixArray}) = HEALPixArray
+horizontal_grid_type(::Type{<:HEALPixArray}) = HEALPixGrid
+full_array_type(::Type{<:HEALPixArray}) = FullHEALPixArray
 
-function get_colatlons(Grid::Type{<:AbstractHEALPixGrid}, nlat_half::Integer)
+## SIZE
+nlat_odd(::Type{<:HEALPixArray}) = true
+get_npoints2D(::Type{<:HEALPixArray}, nlat_half::Integer) = 3*nlat_half^2
+get_nlat_half(::Type{<:HEALPixArray}, npoints2D::Integer) = round(Int, sqrt(npoints2D/3))
+function get_nlon_per_ring(Grid::Type{<:HEALPixArray}, nlat_half::Integer, j::Integer)
     nlat = get_nlat(Grid, nlat_half)
-    npoints = get_npoints(Grid, nlat_half)
-    nside = nside_healpix(nlat_half)
-    colat = get_colat(Grid, nlat_half)
-
-    colats = zeros(npoints)
-    lons = zeros(npoints)
-
-    ij = 1
-    for j in 1:nlat
-        nlon = get_nlon_per_ring(Grid, nlat_half, j)
[…]
