### 2024-04-17 652a6d0
Loop indexing tested
