### 2024-04-17 917b1a0
add Adapt, KernelAbstractions, GPUArrays deps, JLArrays extra
