+++ Project.toml
@@ -15,6 +15,7 @@ DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 FLoops = "cc61a311-1640-44b5-9fba-1b764f453329"
 FastGaussQuadrature = "442a2c76-b920-505d-bb47-c5924d526838"
+GPUArrays = "0c68f7d7-f131-5f86-a1c3-88cf8149b2d7"
 GenericFFT = "a8297547-1b15-4a5a-a998-a2ac5f1cef28"
 JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 KernelAbstractions = "63c18a36-062a-441e-b654-da1e3ab1ce7c"
@@ -46,6 +47,7 @@ DocStringExtensions = "0.9"
 FFTW = "1"
 FLoops = "0.2"
 FastGaussQuadrature = "0.4, 0.5, 1"
+GPUArrays = "10"
 GenericFFT = "0.1"
 JLD2 = "0.4"
 KernelAbstractions = "0.9"
@@ -62,7 +64,8 @@ UnicodePlots = "3.3"
 julia = "1.9"
 
 [extras]
+JLArrays = "27aeb0d3-9eb9-45fb-866b-73c2ecf80fcb"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test"]
+test = ["Test", "JLArrays"]
