### 2024-04-17 3ac3b6e
GPU broadcasting first draft
