+++ src/LowerTriangularMatrices/lower_triangular_matrix.jl
@@ -1,34 +1,38 @@
-
-
 """
 $(TYPEDSIGNATURES)
 A lower triangular array implementation that only stores the non-zero entries explicitly.
 `L<:AbstractMatrix` although in general we have `L[i] != Matrix(L)[i]`, the former skips
 zero entries, tha latter includes them. 
 
-Supports n-dimensional lower triangular arrays, so that for all trailing dimensions `L[:,:,..]`
+Supports n-dimensional lower triangular arrays, so that for all trailing dimensions `L[:, :, ..]`
 is a matrix in lower triangular form, e.g. a (5x5x3)-LowerTriangularArray would hold 3 lower
-triangular matrices.
-"""
+triangular matrices."""
 struct LowerTriangularArray{T, N, ArrayType <: AbstractArray{T}} <: AbstractArray{T,N}
     data::ArrayType     # non-zero elements unravelled into an array with one dimension less N-1
     m::Int              # number of rows
     n::Int              # number of columns
 
-    LowerTriangularArray{T, N, ArrayType}(data, m, n) where {T, N, ArrayType<:AbstractArray{T}} = check_lta_input_array(data, m, n, N) ? new(data, m, n) : error(lta_error_message(data, m, n, T, N, ArrayType))
+    LowerTriangularArray{T, N, ArrayType}(data, m, n) where {T, N, ArrayType<:AbstractArray{T}} =
+        check_lta_input_array(data, m, n, N) ?
+        new(data, m, n) :
+        error(lta_error_message(data, m, n, T, N, ArrayType))
 end
 
-check_lta_input_array(data, m, n, N) = (ndims(data) == N-1) & (length(data) == prod(size(data)[2:end]) * nonzeros(m, n)) 
+check_lta_input_array(data, m, n, N) =
+    (ndims(data) == N-1) & (length(data) == prod(size(data)[2:end]) * nonzeros(m, n)) 
 
 function lta_error_message(data, m, n, T, N, ArrayType) 
     size_tuple = (m, n, size(data[2:end])...)
     "$(size(data))-sized $(typeof(data)) cannot be used to create a $size_tuple LowerTriangularArray{$T,$N,$ArrayType}"
 end 
 
+"""2-dimensional `LowerTriangularArray` of type `T`` with its non-zero entries unravelled into a `Vector{T}`"""
 const LowerTriangularMatrix{T} = LowerTriangularArray{T, 2, Vector{T}}
 
-LowerTriangularArray(data::ArrayType, m::Integer, n::Integer) where {T, N, ArrayType<:AbstractArray{T,N}} = LowerTriangularArray{T, N+1, ArrayType}(data, m, n)
-LowerTriangularMatrix(data::Vector{T}, m::Integer, n::Integer) where T = LowerTriangularMatrix{T}(data, m, n)
+LowerTriangularArray(data::ArrayType, m::Integer, n::Integer) where {T, N, ArrayType<:AbstractArray{T,N}} =
+    LowerTriangularArray{T, N+1, ArrayType}(data, m, n)
+LowerTriangularMatrix(data::Vector{T}, m::Integer, n::Integer) where T =
+    LowerTriangularMatrix{T}(data, m, n)
 
 # SIZE ETC
 Base.length(L::LowerTriangularArray) = length(L.data)  # define length as number of non-zero elements
@@ -36,26 +40,29 @@ Base.size(L::LowerTriangularArray) = (L.m, L.n, size(L.data)[2:end]...)  # defin
[…]
