### 2024-04-17 1a71da6
getindex boundscheck, inbounds corrected, linting
