+++ src/LowerTriangularMatrices/lower_triangular_matrix.jl
@@ -2,7 +2,7 @@
 $(TYPEDSIGNATURES)
 A lower triangular array implementation that only stores the non-zero entries explicitly.
 `L<:AbstractMatrix` although in general we have `L[i] != Matrix(L)[i]`, the former skips
-zero entries, tha latter includes them. 
+zero entries, tha latter includes them.
 
 Supports n-dimensional lower triangular arrays, so that for all trailing dimensions `L[:, :, ..]`
 is a matrix in lower triangular form, e.g. a (5x5x3)-LowerTriangularArray would hold 3 lower
@@ -35,18 +35,31 @@ LowerTriangularMatrix(data::Vector{T}, m::Integer, n::Integer) where T =
     LowerTriangularMatrix{T}(data, m, n)
 
 # SIZE ETC
-Base.length(L::LowerTriangularArray) = length(L.data)  # define length as number of non-zero elements
-Base.size(L::LowerTriangularArray) = (L.m, L.n, size(L.data)[2:end]...)  # define size as matrix size (doesnt give an error, even for LowerTriangularMatrix due to tuple indexing logic)
-Base.sizeof(L::LowerTriangularArray) = sizeof(L.data)  # sizeof the underlying data vector
+"""$(TYPEDSIGNATURES)
+Length of a `LowerTriangularArray` defined as number of non-zero elements."""
+Base.length(L::LowerTriangularArray) = length(L.data)
+
+"""$(TYPEDSIGNATURES)
+Size of a `LowerTriangularArray` defined as matrix size including zero upper triangle."""
+Base.size(L::LowerTriangularArray) = (L.m, L.n, size(L.data)[2:end]...)
+
+# sizeof the underlying data vector
+Base.sizeof(L::LowerTriangularArray) = sizeof(L.data)
 
 # CREATE INSTANCES (ZEROS, ONES, UNDEF)
 for f in (:zeros, :ones, :rand, :randn)
     @eval begin
-        function Base.$f(::Type{LowerTriangularArray{T, N, ArrayType}}, m::Integer, n::Integer, k::Integer...) where {T, N, ArrayType}
+        # use ArrayType from LowerTriangularArray parameter
+        function Base.$f(
+            ::Type{LowerTriangularArray{T, N, ArrayType}}, 
+            m::Integer,
+            n::Integer,
+            k::Integer...
+        ) where {T, N, ArrayType}
             return LowerTriangularArray(ArrayType($f(T, nonzeros(m, n), k...)), m, n)
         end
         
-        # default CPU 
+        # default CPU, use Array
         function Base.$f(::Type{LowerTriangularArray{T}}, m::Integer, n::Integer, k::Integer...) where T
             return LowerTriangularArray($f(T, nonzeros(m, n), k...), m, n)
         end
@@ -66,7 +79,12 @@ function LowerTriangularArray{T, N, ArrayType}(
     return LowerTriangularArray(ArrayType(undef, nonzeros(I[1], I[2]), I[3:end]...), I[1], I[2])
 end
 
-LowerTriangularArray{T,N,ArrayType}(::UndefInitializer, size::S) where {T,N,ArrayType<:AbstractArray{T}, S<:Tuple} = LowerTriangularArray{T,N,ArrayType}(undef, size...)
+function LowerTriangularArray{T, N, ArrayType}(
+    ::UndefInitializer,
+    size::S,
+) where {T,N,ArrayType<:AbstractArray{T}, S<:Tuple}
+    return LowerTriangularArray{T, N, ArrayType}(undef, size...)
[…]
