+++ src/SpeedyWeather.jl
@@ -37,7 +37,10 @@ export initialize!, finish!
 include("utility_functions.jl")
 
 # LowerTriangularMatrices for spherical harmonics
-export LowerTriangularMatrices, LowerTriangularMatrix, LowerTriangularArray
+export  LowerTriangularMatrices, 
+        LowerTriangularMatrix,
+        LowerTriangularArray
+        
 include("LowerTriangularMatrices/LowerTriangularMatrices.jl")
 using .LowerTriangularMatrices
 
