### 2024-04-17 963988b
using Adapt, JLArrays in ringgrids test
