### 2024-04-17 cfae034
full_grid_type vs full_array_type plus tests
