### 2024-04-17 38d4062
use octahealpix weights for octahealpix not healpix
