+++ src/RingGrids/grids/full_octahealpix.jl
@@ -25,7 +25,7 @@ get_colat(::Type{<:FullOctaHEALPixArray}, nlat_half::Integer) = get_colat(OctaHE
 get_lon(::Type{<:FullOctaHEALPixArray}, nlat_half::Integer) = get_lon(FullGaussianArray, nlat_half)
 
 # QUADRATURE
-get_quadrature_weights(::Type{<:FullOctaHEALPixArray}, nlat_half::Integer) = healpix_weights(nlat_half)
+get_quadrature_weights(::Type{<:FullOctaHEALPixArray}, nlat_half::Integer) = octahealpix_weights(nlat_half)
 
 """
     G = FullOctaHEALPixGrid{T}
