### 2024-04-18 f819cc1
docstrings for general and full grids
