### 2024-04-18 0d2d8c8
Exports more gradient operators from SpeedyTransforms
