+++ src/SpeedyWeather.jl
@@ -58,7 +58,8 @@ using .RingGrids
 
 # SpeedyTransforms
 export SpeedyTransforms, SpectralTransform
-export spectral, gridded, spectral_truncation
+export spectral, gridded
+export spectral_truncation, spectral_truncation!
 export curl, divergence, curl!, divergence!
 export ∇, ∇², ∇⁻², ∇!, ∇²!, ∇⁻²!
 
