### 2024-04-18 7b5ad57
Reduced grid documentation
