### 2024-04-19 cfb79fc
More @inbounds checks
