### 2024-04-19 bfb2822
disable out of bounds access for github actions CI (always checks)
