+++ ext/SpeedyWeatherJLArraysExt.jl
@@ -1,0 +1,8 @@
+module SpeedyWeatherMakieExt
+
+using SpeedyWeather, JLArrays
+
+# for RingGrids, every Array needs this method to strip away the parameters
+SpeedyWeather.RingGrids.nonparametric_type(::Type{JLArray}) = JLArray
+
+end # module
