+++ Project.toml
@@ -31,9 +31,11 @@ UnicodePlots = "b8865327-cd53-5732-bb35-84acbb429228"
 
 [weakdeps]
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
+JLArrays = "27aeb0d3-9eb9-45fb-866b-73c2ecf80fcb"
 
 [extensions]
 SpeedyWeatherMakieExt = "Makie"
+SpeedyWeatherJLArraysExt = "JLArrays
 
 [compat]
 AbstractFFTs = "1"
@@ -49,6 +51,7 @@ FLoops = "0.2"
 FastGaussQuadrature = "0.4, 0.5, 1"
 GPUArrays = "10"
 GenericFFT = "0.1"
+JLArrays = "0.1.4"
 JLD2 = "0.4"
 KernelAbstractions = "0.9"
 LinearAlgebra = "1.9"
