### 2024-04-19 abb781f
radius keyword for all allocating gradient operators
