+++ src/SpeedyTransforms/spectral_gradients.jl
@@ -1,3 +1,5 @@
+const DEFAULT_RADIUS = 1
+
 """
 $(TYPEDSIGNATURES)
 Curl of a vector `u, v` written into `curl`, `curl = ∇×(u, v)`.
@@ -15,7 +17,7 @@ function curl!( curl::LowerTriangularMatrix,
 
     # = -(∂λ - ∂θ) or (∂λ - ∂θ), adding or overwriting the output curl
     kernel(o, a, b, c) = flipsign ? (add ? o-(a+b-c) : -(a+b-c)) :
-                                 (add ? o+(a+b-c) :   a+b-c )    
+                                    (add ? o+(a+b-c) :   a+b-c )    
     _divergence!(kernel, curl, v, u, S)             # flip u, v -> v, u
 end
 
@@ -36,7 +38,7 @@ function divergence!(   div::LowerTriangularMatrix,
 
     # = -(∂λ + ∂θ) or (∂λ + ∂θ), adding or overwriting the output div
     kernel(o, a, b, c) = flipsign ? (add ? o-(a-b+c) : -(a-b+c)) :
-                                 (add ? o+(a-b+c) :   a-b+c )                
+                                    (add ? o+(a-b+c) :   a-b+c )                
     _divergence!(kernel, div, u, v, S)
 end
 
@@ -104,19 +106,28 @@ This function requires both `u, v` to be transforms of fields that are scaled wi
 Both `div` and `div_grid` are scaled with the radius.
 """
 function divergence(u::LowerTriangularMatrix,
-                    v::LowerTriangularMatrix)
+                    v::LowerTriangularMatrix;
+                    radius = DEFAULT_RADIUS)
 
     @assert size(u) == size(v) "Size $(size(u)) and $(size(v)) incompatible."
 
     S = SpectralTransform(u)
     div = similar(u)
     divergence!(div, u, v, S, add=false, flipsign=false)
+
+    if radius != 1
+        div .*= inv(radius)
+    end
+
     return div
 end
 
-function _div_or_curl(  kernel!,
-                        u::Grid,
-                        v::Grid) where {Grid<:AbstractGrid}
+function _div_or_curl(  
+    kernel!,
+    u::Grid,
+    v::Grid;
+    radius = DEFAULT_RADIUS
+) where {Grid<:AbstractGrid}
 
     @assert size(u) == size(v) "Size $(size(u)) and $(size(v)) incompatible."
 
@@ -132,6 +143,11 @@ function _div_or_curl(  kernel!,
 
     div_or_vor = similar(us)
     kernel!(div_or_vor, us, vs, S, add=false, flipsign=false)
+
+    if radius != 1
+        div_or_vor .*= inv(radius)
+    end
+
     return div_or_vor
 end
 
@@ -141,7 +157,7 @@ Divergence (∇⋅) of two vector components `u, v` on a grid.
 Applies 1/coslat scaling, transforms to spectral space and returns
 the spectral divergence, which is scaled with the radius
 of the sphere. Divide by radius for unscaling."""
-divergence(u::Grid, v::Grid) where {Grid<:AbstractGrid} = _div_or_curl(divergence!, u, v)
+divergence(u::Grid, v::Grid; kwargs...) where {Grid<:AbstractGrid} = _div_or_curl(divergence!, u, v; kwargs...)
 
 """
 $(TYPEDSIGNATURES)
@@ -149,7 +165,7 @@ Curl (∇×) of two vector components `u, v` on a grid.
 Applies 1/coslat scaling, transforms to spectral space and returns
[…]
