### 2024-04-19 5b087e9
CUDA.CuArray for nonparametric_type
