### 2024-04-19 b741672
Add radius keyword to docstrings
