+++ src/SpeedyTransforms/spectral_gradients.jl
@@ -3,7 +3,8 @@ const DEFAULT_RADIUS = 1
 """
 $(TYPEDSIGNATURES)
 Curl of a vector `u, v` written into `curl`, `curl = ∇×(u, v)`.
-`u, v` are expected to have a 1/coslat-scaling included, then `curl` is not scaled.
+`u, v` are expected to have a 1/coslat-scaling included, otherwise `curl` is scaled.
+Acts on the unit sphere, i.e. it omits 1/radius scaling as all inplace gradient operators.
 `flipsign` option calculates -∇×(u, v) instead. `add` option calculates `curl += ∇×(u, v)` instead.
 `flipsign` and `add` can be combined. This functions only creates the kernel and calls the generic
 divergence function _divergence! subsequently with flipped u, v -> v, u for the curl."""
@@ -24,7 +25,8 @@ end
 """
 $(TYPEDSIGNATURES)
 Divergence of a vector `u, v` written into `div`, `div = ∇⋅(u, v)`. 
-`u, v` are expected to have a 1/coslat-scaling included, then `div` is not scaled.
+`u, v` are expected to have a 1/coslat-scaling included, otherwise `div` is scaled.
+Acts on the unit sphere, i.e. it omits 1/radius scaling as all inplace gradient operators.
 `flipsign` option calculates -∇⋅(u, v) instead. `add` option calculates `div += ∇⋅(u, v)` instead.
 `flipsign` and `add` can be combined. This functions only creates the kernel and calls
 the generic divergence function _divergence! subsequently."""
@@ -46,7 +48,9 @@ end
 $(TYPEDSIGNATURES)
 Generic divergence function of vector `u`, `v` that writes into the output into `div`.
 Generic as it uses the kernel `kernel` such that curl, div, add or flipsign
-options are provided through `kernel`, but otherwise a single function is used."""
+options are provided through `kernel`, but otherwise a single function is used.
+Acts on the unit sphere, i.e. it omits 1/radius scaling as all inplace gradient operators.
+"""
 function _divergence!(  kernel,
                         div::LowerTriangularMatrix{Complex{NF}},
                         u::LowerTriangularMatrix{Complex{NF}},
@@ -94,16 +98,16 @@ $(TYPEDSIGNATURES)
 Divergence (∇⋅) of two vector components `u, v` which need to have size (n+1)xn,
 the last row will be set to zero in the returned `LowerTriangularMatrix`.
 This function requires both `u, v` to be transforms of fields that are scaled with
-`1/cos(lat)`. An example usage is therefore
+`1/cos(lat)`. Acts on the unit sphere, i.e. it omits 1/radius scaling unless
+`radius` keyword argument is provided.
+An example usage is therefore
 
     RingGrids.scale_coslat⁻¹!(u_grid)
     RingGrids.scale_coslat⁻¹!(v_grid)
     u = spectral(u_grid, one_more_degree=true)
     v = spectral(v_grid, one_more_degree=true)
-    div = divergence(u, v)
+    div = divergence(u, v, radius = 6.371e6)
     div_grid = gridded(div)
-
-Both `div` and `div_grid` are scaled with the radius.
 """
 function divergence(u::LowerTriangularMatrix,
                     v::LowerTriangularMatrix;
@@ -122,6 +126,7 @@ function divergence(u::LowerTriangularMatrix,
     return div
 end
 
+# called by divergence or curl
 function _div_or_curl(  
     kernel!,
     u::Grid,
@@ -155,16 +160,17 @@ end
 $(TYPEDSIGNATURES)
 Divergence (∇⋅) of two vector components `u, v` on a grid.
[…]
