### 2024-04-20 c2fcdfe
Gradient example in docs
