+++ docs/src/gradients.md
@@ -1,0 +1,214 @@
+## Gradient operators
+
+[SpeedyTransforms](@ref) also includes many gradient operators to take derivatives in
+spherical harmonics. These are in particular ``\nabla, \nabla \cdot, \nabla \times,
+\nabla^2, \nabla^{-2}``. We call them `divergence`, `curl`, `∇`, `∇²`, `∇⁻²`
+(as well as their in-place versions with `!`) within the limits of unicode characters
+and Julia syntax. These functions are defined for inputs being spectral coefficients
+(i.e. `LowerTriangularMatrix`) or gridded fields (i.e. `<:AbstractGrid`) and
+also allow as an additional argument a spectral transform object
+(see [SpectralTransform](@ref SpectralTransform)) which avoids recalculating it
+under the hood.
+
+!!! info "SpeedyTransforms assumes a unit sphere"
+    The gradient operators in SpeedyTransforms generally assume a sphere of radius ``R=1``.
+    For the transforms themselves that does not make a difference, but the gradient operators
+    `divergence`, `curl`, `∇`, `∇²`, `∇⁻²` omit the radius scaling unless you provide the optional
+    keyword `radius` (or you can do `./= radius` manually). 
+    Also note that meridional derivates in spectral space expect a ``\cos^{-1}(\theta)`` scaling.
+    Details are always outlined in the respective docstrings, `?∇` for example.
+
+
+The actually implemented operators are,
+in contrast to the mathematical [Derivatives in spherical coordinates](@ref)
+due to reasons of scaling as follows. Let the implemented operators be
+``\hat{\nabla}`` etc.
+
+```math
+\hat{\nabla} A = \left(\frac{\partial A}{\partial \lambda}, \cos(\theta)\frac{\partial A}{\partial \theta} \right) =
+R\cos(\theta)\nabla A
+```
+So the zonal derivative omits the radius and the ``\cos^{-1}(\theta)`` scaling.
+The meridional derivative adds a ``\cos(\theta)`` due to a recursion relation
+being defined that way, which, however, is actually convenient because the whole
+operator is therefore scaled by ``R\cos(\theta)``. The curl and divergence operators
+expect the input velocity fields to be scaled by ``\cos^{-1}(\theta)``, i.e.
+
+```math
+\begin{aligned}
+\hat{\nabla} \cdot (\cos^{-1}(\theta)\mathbf{u}) &= \frac{\partial u}{\partial \lambda} +
+\cos\theta\frac{\partial v}{\partial \theta} = R\nabla \cdot \mathbf{u}, \\
+\hat{\nabla} \times (\cos^{-1}(\theta)\mathbf{u}) &= \frac{\partial v}{\partial \lambda} -
+\cos\theta\frac{\partial u}{\partial \theta} = R\nabla \times \mathbf{u}.
+\end{aligned}
+```
+
+And the Laplace operators omit a ``R^2`` (radius ``R``) scaling, i.e.
+
+```math
+\hat{\nabla}^{-2}A = \frac{1}{R^2}\nabla^{-2}A , \quad \hat{\nabla}^{2}A = R^2\nabla^{2}A
+```
+## Gradient `∇`
+
+We illustrate the usage of the gradient function `∇`. Let us create some fake
+data `G` on the grid first 
+
+```@example gradient
+using SpeedyWeather, CairoMakie
+
+# create some data with wave numbers 0,1,2,3,4
+trunc = 64                  # 1-based maximum degree of spherical harmonics
+L = randn(LowerTriangularMatrix{ComplexF32}, trunc, trunc)
+spectral_truncation!(L, 5)  # remove higher wave numbers
+G = gridded(L)
[…]
