+++ docs/src/speedytransforms.md
@@ -17,13 +17,6 @@ Combined with the spectral transform, you could for example start with a velocit
 field in grid-point space, transform to spectral, compute its divergence and transform
 back to obtain the divergence in grid-point space. Examples are outlined in [Gradient operators](@ref).
 
-!!! info "SpeedyTransforms assumes a unit sphere"
-    The operators in SpeedyTransforms generally assume a sphere of radius ``R=1``.
-    For the transforms themselves that does not make a difference, but the gradient operators
-    `div`, `curl`, `∇`, `∇²`, `∇⁻²` omit the radius scaling. You will have to do this manually.
-    Also note that the meridional derivate expects a ``\cos^{-1}(\theta)`` scaling.
-    More in [Gradient operators](@ref).
-
 ## Example transform
 
 Lets start with a simple transform. We could be `using SpeedyWeather` but to be more verbose
@@ -100,7 +93,7 @@ to interpolate directly between grids. If you want to control directly the resol
 grid `gridded` is supposed to transform onto you have to provide a `SpectralTransform` instance.
 More on that now.
 
-## The `SpectralTransform` struct
+## [The `SpectralTransform` struct](@id SpectralTransform)
 
 Both `spectral` and `gridded` create an instance of `SpectralTransform` under the hood. This object
 contains all precomputed information that is required for the transform, either way: 
@@ -273,162 +266,6 @@ to kilobytes
 ```@example speedytransforms
 SpectralTransform(spectral_grid, recompute_legendre=true)
 ```
-
-## Gradient operators
-
-`SpeedyTransforms` also includes many gradient operators to take derivatives in
-spherical harmonics. These are in particular ``\nabla, \nabla \cdot, \nabla \times,
-\nabla^2, \nabla^{-2}``. However, the actually implemented operators are,
-in contrast to the mathematical [Derivatives in spherical coordinates](@ref)
-due to reasons of scaling as follows. Let the implemented operators be
-``\hat{\nabla}`` etc.
-
-```math
-\hat{\nabla} A = \left(\frac{\partial A}{\partial \lambda}, \cos(\theta)\frac{\partial A}{\partial \theta} \right) =
-R\cos(\theta)\nabla A
-```
-So the zonal derivative omits the radius and the ``\cos^{-1}(\theta)`` scaling.
-The meridional derivative adds a ``\cos(\theta)`` due to a recursion relation
-being defined that way, which, however, is actually convenient because the whole
-operator is therefore scaled by ``R\cos(\theta)``. The curl and divergence operators
-expect the input velocity fields to be scaled by ``\cos^{-1}(\theta)``, i.e.
-
-```math
-\begin{aligned}
-\hat{\nabla} \cdot (\cos^{-1}(\theta)\mathbf{u}) &= \frac{\partial u}{\partial \lambda} +
-\cos\theta\frac{\partial v}{\partial \theta} = R\nabla \cdot \mathbf{u}, \\
-\hat{\nabla} \times (\cos^{-1}(\theta)\mathbf{u}) &= \frac{\partial v}{\partial \lambda} -
-\cos\theta\frac{\partial u}{\partial \theta} = R\nabla \times \mathbf{u}.
-\end{aligned}
-```
-
-And the Laplace operators omit a ``R^2`` (radius ``R``) scaling, i.e.
-
-```math
-\hat{\nabla}^{-2}A = \frac{1}{R^2}\nabla^{-2}A , \quad \hat{\nabla}^{2}A = R^2\nabla^{2}A
-```
-
-## Example: Geostrophy
-
-Now, we want to use the following example to illustrate
-their use: We have ``u, v`` and want to calculate ``\eta`` in the shallow water
[…]
