### 2024-04-20 52b624d
Gradient docs add some heatmap titles
