### 2024-04-20 SpeedyWeather/SpeedyWeather.jl#443 @kthyng
@milankl How is this work coming along?
