### 2024-04-22 61dc760
more setindex tests for AbstractGridArray
