### 2024-04-22 1ca3a90
SpectralTransform uses nonparametric type of AbstractGridArray
