+++ src/SpeedyTransforms/spectral_transform.jl
@@ -6,8 +6,8 @@ for the spectral transform."""
 struct SpectralTransform{NF<:AbstractFloat}
 
     # GRID
-    Grid::Type{<:AbstractGrid}  # grid type used
-    nlat_half::Int              # resolution parameter of grid (# of latitudes on one hemisphere, Eq incl)
+    Grid::Type{<:AbstractGridArray} # grid type used
+    nlat_half::Int                  # resolution parameter of grid (# of latitudes on one hemisphere, Eq incl)
 
     # SPECTRAL RESOLUTION
     lmax::Int               # Maximum degree l=[0, lmax] of spherical harmonics
@@ -71,26 +71,28 @@ Generator function for a SpectralTransform struct. With `NF` the number format,
 `Grid` the grid type `<:AbstractGrid` and spectral truncation `lmax, mmax` this function sets up
 necessary constants for the spetral transform. Also plans the Fourier transforms, retrieves the colatitudes,
 and preallocates the Legendre polynomials (if recompute_legendre == false) and quadrature weights."""
-function SpectralTransform( ::Type{NF},                     # Number format NF
-                            Grid::Type{<:AbstractGrid},     # type of spatial grid used
-                            lmax::Int,                      # Spectral truncation: degrees
-                            mmax::Int;                      # Spectral truncation: orders
+function SpectralTransform( ::Type{NF},                         # Number format NF
+                            Grid::Type{<:AbstractGridArray},    # type of spatial grid used
+                            lmax::Int,                          # Spectral truncation: degrees
+                            mmax::Int;                          # Spectral truncation: orders
                             recompute_legendre::Bool = true,        # re or precompute legendre polynomials?
                             legendre_shortcut::Symbol = :linear,    # shorten Legendre loop over order m
                             dealiasing::Real=DEFAULT_DEALIASING
                             ) where NF
 
+    Grid = RingGrids.nonparametric_type(Grid)   # always use nonparametric super type
+
     # RESOLUTION PARAMETERS
-    nlat_half = get_nlat_half(mmax, dealiasing)  # resolution parameter nlat_half,
+    nlat_half = get_nlat_half(mmax, dealiasing) # resolution parameter nlat_half,
                                                 # number of latitude rings on one hemisphere incl equator
-    nlat = get_nlat(Grid, nlat_half)             # 2nlat_half but one less if grids have odd # of lat rings
-    nlon_max = get_nlon_max(Grid, nlat_half)     # number of longitudes around the equator
+    nlat = get_nlat(Grid, nlat_half)            # 2nlat_half but one less if grids have odd # of lat rings
+    nlon_max = get_nlon_max(Grid, nlat_half)    # number of longitudes around the equator
                                                 # number of longitudes per latitude ring (one hemisphere only)
     nlons = [RingGrids.get_nlon_per_ring(Grid, nlat_half, j) for j in 1:nlat_half]
     nfreq_max = nlon_max÷2 + 1                      # maximum number of fourier frequencies (real FFTs)
 
     # LATITUDE VECTORS (based on Gaussian, equi-angle or HEALPix latitudes)
-    colat = RingGrids.get_colat(Grid, nlat_half)     # colatitude in radians                             
+    colat = RingGrids.get_colat(Grid, nlat_half)    # colatitude in radians                             
     cos_colat = cos.(colat)                         # cos(colat)
[…]
