### 2024-04-22 b9bd961
corrected quadrature weights for full healpix grids
