### 2024-04-23 77a7592
any(::AbstractGridArray) added
