### 2024-04-23 20b0938
LowerTriangular arithmetic via broadcasting
