+++ src/LowerTriangularMatrices/lower_triangular_matrix.jl
@@ -23,7 +23,8 @@ check_lta_input_array(data, m, n, N) =
 
 function lta_error_message(data, m, n, T, N, ArrayType) 
     size_tuple = (m, n, size(data[2:end])...)
-    "$(size(data))-sized $(typeof(data)) cannot be used to create a $size_tuple LowerTriangularArray{$T,$N,$ArrayType}"
+    return "$(size(data))-sized $(typeof(data)) cannot be used to create "*
+                "a $size_tuple LowerTriangularArray{$T,$N,$ArrayType}"
 end 
 
 """2-dimensional `LowerTriangularArray` of type `T`` with its non-zero entries unravelled into a `Vector{T}`"""
@@ -104,9 +105,9 @@ $(TYPEDSIGNATURES)
 Converts the index pair `i, j` of an `m`x`n` LowerTriangularMatrix `L` to a single
 index `k` that indexes the same element in the corresponding vector that stores
 only the lower triangle (the non-zero entries) of `L`."""
-@inline ij2k(i::Integer, j::Integer, m::Integer) = i+(j-1)*m-j*(j-1)÷2
+@inline ij2k(i::Integer, j::Integer, m::Integer) = i + (j-1)*m - j*(j-1)÷2
 @inline triangle_number(n::Integer) = n*(n+1)÷2
-nonzeros(m::Integer, n::Integer) = m*n-triangle_number(n-1)
+nonzeros(m::Integer, n::Integer) = m*n - triangle_number(n-1)
 
 """
 $(TYPEDSIGNATURES)
@@ -413,74 +414,60 @@ Base.similar(L::LowerTriangularArray{T,N,ArrayType}, ::Type{T}) where {T, N, Arr
 Base.similar(L::LowerTriangularArray{T}) where T = similar(L, T)
  
 # ARITHMETIC
-# only mul/div with scalar and addition/subtraction, others are converted to Matrix
-function Base.:(*)(L::LowerTriangularArray{T}, s::Number) where T
-    sT = convert(T, s)
-    LowerTriangularArray(L.data .* sT, L.m, L.n)
-end
-
-Base.:(*)(s::Number, L::LowerTriangularArray) = L*s         # commutative
-Base.:(/)(L::LowerTriangularArray, s::Number) = L*inv(s)
-
-function Base.:(+)(
-    L1::LowerTriangularArray{T, N, ArrayType},
-    L2::LowerTriangularArray{S, N, ArrayType},
-) where {T, S, N, ArrayType}
-    @boundscheck (L1.m == L2.m) && (L1.n == L2.n) || throw(BoundsError)
-    LowerTriangularArray(L1.data + L2.data, L1.m, L1.n)
-end
-
-function Base.:(-)(
-    L1::LowerTriangularArray{T,N,ArrayType},
-    L2::LowerTriangularArray{S,N,ArrayType}
-) where {T, S, N, ArrayType}
-    @boundscheck (L1.m == L2.m) && (L1.n == L2.n) || throw(BoundsError)
-    LowerTriangularArray(L1.data - L2.data, L1.m, L1.n)
-end
-
-Base.:(-)(L::LowerTriangularArray) = LowerTriangularArray(-L.data, L.m, L.n)
 Base.prod(L::LowerTriangularArray{NF}) where NF = zero(NF)
 
 """
 $(TYPEDSIGNATURES)
 Fills the elements of `L` with `x`. Faster than fill!(::AbstractArray, x)
 as only the non-zero elements in `L` are assigned with x."""
-function Base.fill!(L::LowerTriangularArray{T}, x) where T
-    xT = convert(T, x)
-    fill!(L.data, xT)
-
-    L
[…]
