### 2024-04-23 d3ec392
remove .== tests as broadcasting still fails for that
