### 2024-04-23 35d0ad5
isstructurepreserving, fzeropresering still incorrectly used, always return false
