### 2024-04-24 138810a
broadcasting with type promotion
