### 2024-04-24 f7c68a4
maybe define all LowerTriangularArray operations as structure and zero preserving?
