### 2024-04-24 090fd10
isapprox also propagated to .data
