### 2024-04-24 b5e7d6a
don't require type equality for == to compare across eltypes too
