+++ src/LowerTriangularMatrices/lower_triangular_matrix.jl
@@ -425,8 +425,8 @@ function Base.fill!(L::LowerTriangularArray, x)
     return L
 end
 
-Base.:(==)(L1::LowerTriangularArray, L2::LowerTriangularArray) =
-    typeof(L1) == typeof(L2) && L1.data == L2.data
+Base.:(==)(L1::LowerTriangularArray, L2::LowerTriangularArray) = 
+    L1.m == L2.m && L1.n == L2.n && L1.data == L2.data
 Base.isapprox(L1::LowerTriangularArray, L2::LowerTriangularArray; kwargs...) =
     isapprox(L1.data, L2.data; kwargs...)
 Base.all(L::LowerTriangularArray) = all(L.data)
