### 2024-04-24 59c8723
integrate u->eta via transforms
