+++ src/dynamics/initial_conditions.jl
@@ -132,23 +132,16 @@ function initialize!(   progn::PrognosticVariables,
     λ = perturb_lon*2π/360          # perturbation longitude [radians]
 
     (; rotation, gravity) = model.planet
-    (; radius) = model.spectral_grid
+    (; Grid, NF, radius, nlat_half) = model.spectral_grid
+    (; coslat⁻¹) = model.geometry
 
-    # always create on F64 grid then convert to spectral and interpolate there
-    Grid = FullGaussianGrid
-    nlat_half = 64
-    u_grid = zeros(Grid, nlat_half)
-    η_grid = zeros(Grid, nlat_half)
-    colats = RingGrids.get_colat(Grid, nlat_half)
+    u_grid = zeros(Grid{NF}, nlat_half)
+    η_perturb_grid = zeros(Grid{NF}, nlat_half)
+    lat = RingGrids.get_lat(Grid, nlat_half)
     _, lons = RingGrids.get_colatlons(Grid, nlat_half)
-    weights = FastGaussQuadrature.gausslegendre(2nlat_half)[2]
-    η_sum = 0
-    latstr = π/2
 
-    for (j, ring) in enumerate(eachring(u_grid, η_grid))
-        θ = π/2 - colats[j]             # latitude in radians
-        coslat⁻¹j = 1/cos(θ)
-        f = 2rotation*sin(θ)
+    for (j, ring) in enumerate(eachring(u_grid, η_perturb_grid))
+        θ = lat[j]             # latitude in radians
         
         # velocity per latitude
         if θ₀ < θ < θ₁
@@ -157,44 +150,56 @@ function initialize!(   progn::PrognosticVariables,
             u_θ = 0
         end
 
-        # integration for layer thickness h / interface height η
-        w = weights[j]
-        dlat = θ - latstr
-        latstr = θ
-        # TODO
-        # we may need a better algorithm here
-        η_sum += -dlat*(radius*u_θ/gravity * (f + tan(θ)/radius*u_θ))
-        #η_sum += 2w*(radius*u_θ/gravity * (f + tan(θ)/radius*u_θ))
-
         # lon-constant part of perturbation
         ηθ = perturb_height*cos(θ)*exp(-((θ₂-θ)/β)^2)
 
         # store in all longitudes
         for ij in ring
-            u_grid[ij] = u_θ/radius*coslat⁻¹j   # include scaling for curl!
+            u_grid[ij] = u_θ/radius*coslat⁻¹[j]   # include scaling for curl!
             
             # calculate perturbation (possibly shifted in lon compared to Galewsky 2004)
             ϕ = lons[ij] - λ
-            η_grid[ij] = η_sum + exp(-(ϕ/α)^2)*ηθ
+            η_perturb_grid[ij] = exp(-(ϕ/α)^2)*ηθ
         end
     end
 
-    u = spectral(u_grid)
-    η = spectral(η_grid)
-
-    # interpolate in spectral space to desired resolution
-    (; lmax, mmax) = model.spectral_transform
-    (; NF) = model.spectral_grid
-    u = spectral_truncation(complex(NF), u, lmax, mmax)
+    # the following obtain initial conditions for η from u, v=0 via
+    # 0 = -∇⋅((ζ+f)*(-v, u)) - ∇²((u^2 + v^2)/2), i.e.
+    # invert the Laplacian for
+    # ∇²(gη) = -∇⋅(0, (ζ+f)*u) - ∇²(u^2/2)
+    u = spectral(u_grid, model.spectral_transform)
     
     # get vorticity initial conditions from curl of u, v
     v = zero(u)     # meridional velocity zero for these initial conditions
[…]
