+++ src/models/primitive_dry.jl
@@ -46,7 +46,7 @@ Base.@kwdef mutable struct PrimitiveDryModel{
 } <: PrimitiveDry
 
     spectral_grid::SpectralGrid
-    device_setup::DS = DeviceSetup(CPUDevice())
+    device_setup::DS = DeviceSetup(spectral_grid.device)
     
     # DYNAMICS
     dynamics::Bool = true

[…]
