+++ docs/src/orography.md
@@ -1,0 +1,188 @@
+# Orography
+
+Orography (in height above the surface) forms the surface boundary of the
+lowermost layer in SpeedyWeather. 
+
+In the [shallow-water equations](@ref shallow_water_model) the orography
+``H_b`` enters the equations when computing the layer thickness ``h = \eta + H_0 - H_b``
+for the volume fluxes ``\mathbf{u}h`` in the continuity equation.
+Here, the orography is used in meters above the surface which shortens
+``h`` over mountains. The orography here is needed in grid-point space.
+
+In the [primitive equations](@ref primitive_equation_model) the orography enters
+the equations when computing the [Geopotential](@ref). So actually required here
+is the surface geopotential ``\Phi_s = gz_s`` where ``z_s``
+is the orography height in meters as used in the shallow-water equations too
+``z_s = H_b``.
+However, the primitive equations require the orography in spectral
+space as the geopotential calculation is a linear operation in the
+horizontal and can therefore be applied in either grid-point or spectral space.
+The latter is more convenient as SpeedyWeather solves the equations
+to avoid additional transforms.
+
+In the current formulation of the
+[barotropic vorticity equations](@ref barotropic_vorticity_model) there is no
+orography. In fact, the field `model.orography` is not defined for
+`model::BarotropicModel`.
+
+## Orographies implemented
+
+Currently implemented are
+```@example orography
+using InteractiveUtils # hide
+using SpeedyWeather
+subtypes(SpeedyWeather.AbstractOrography)
+```
+which are 
+
+- ``\Phi_s = z_s = H_b = 0`` for `NoOrography`
+- For `ZonalRidge` the zonal ridge from the Jablonowski and Williamson initial conditions, see [Jablonowski-Williamson baroclinic wave](@ref)
+- For `EarthOrography` a high-resolution orography is loaded and interpolated to the resolution as defined by `spectral_grid`.
+
+all orographies need to be created with `spectral_grid::SpectralGrid` as the first argument,
+so that the respective fields for `geopot_surf`, i.e. ``\Phi_s`` and `orography`, i.e. ``H_b``
+can be allocated in the right size and number format.
+
+## Earth's orography
+
+Earth's orography can be created with (here we use a resolution of T85, about 165km globally)
+
+```@example orography
+using SpeedyWeather
+spectral_grid = SpectralGrid(trunc=85)
+orography = EarthOrography(spectral_grid)
+```
+
+but note that only allocates the orography, it does not actually load
+and interpolate the orography which happens at the `initialize!`
+step. Visualised with
+
+```@example orography
+model = PrimitiveDryModel(;spectral_grid, orography)
+initialize!(orography, model)   # happens also in simulation = initialize!(model)
+
+using CairoMakie
+heatmap(orography.orography, title="Earth's orography at T85 resolution, no smoothing")
+save("earth_orography.png", ans) # hide
+nothing # hide
+```
+![EarthOrography](earth_orography.png)
+
+typing `?EarthOrography` shows the various options that are provided.
+An orogaphy at T85 resolution that is as smooth as it would be at T31
+for example can be created with
+
+```@example orography
+orography = EarthOrography(spectral_grid, smoothing=true, smoothing_truncation=31)
+initialize!(orography, model)
+
+heatmap(orography.orography, title="Earth's orography at T85 resolution, smoothed to T31")
+save("earth_orography_smooth.png", ans) # hide
+nothing # hide
+```
+![EarthOrography_smooth](earth_orography_smooth.png)
+
+## Load orography from file
+
+The easiest to load another orography from a netCDF file is to reuse the
+`EarthOrography`, e.g.
+
+```julia
+mars_orography = EarthOrography(spectal_grid, 
+                                path="path/to/my/orography",
[…]
