+++ src/dynamics/particle_advection.jl
@@ -2,7 +2,7 @@ abstract type AbstractParticleAdvection <: AbstractModelComponent end
 
 export NoParticleAdvection
 struct NoParticleAdvection <: AbstractParticleAdvection end
-NoParticleAdvection(SG::SpectralGrid) = NoParticleAdvection()
+NoParticleAdvection(::SpectralGrid) = NoParticleAdvection()
 initialize!(::NoParticleAdvection, ::ModelSetup) = nothing
 particle_advection!(progn, diagn, lf, ::NoParticleAdvection) = nothing
 
@@ -33,7 +33,7 @@ function initialize!(
 
     (; every_n_timesteps) = particle_advection
     # Δt [˚*s/m] is scaled by radius to convert more easily from velocity [m/s]
-    # to [˚/s] for particle locations in degree
+    # to [˚/s] for particle locations in degree
     particle_advection.Δt[] = every_n_timesteps * model.time_stepping.Δt
     particle_advection.Δt[] *= (360/2π)
 end
@@ -53,27 +53,72 @@ function initialize!(
     end
 end
 
-# function barrier
-function particle_advection!(progn, diagn, lf, adv::ParticleAdvection2D)
-    particle_advection!(progn.particles, diagn, progn.clock, lf, adv)
-end
-
-function particle_advection!(
+"""$(TYPEDSIGNATURES)
+Initialize particle advection time integration: Store u,v interpolated initial conditions
+in `diagn.particles.u` and `.v`  to be used when particle advection actually executed for first time."""
+function initialize!(
     particles::Vector{Particle{NF}},
-    diagn::AbstractVariables,
-    clock::Clock,
-    lf::Int,    # leapfrog index used to distinguish between 1st and other steps
+    progn::PrognosticVariables,
+    diagn::DiagnosticVariables,
     particle_advection::ParticleAdvection2D,
 ) where NF
 
     # escape immediately for no particles
     length(particles) == 0 && return nothing
 
-    # escape immediately if advection not on this timestep
-    clock.timestep_counter % particle_advection.every_n_timesteps == 0 || return nothing
+    k = particle_advection.layer
+    (; u_grid, v_grid) = diagn.layers[k].grid_variables
+    (; interpolator) = diagn.particles
+    
+    # interpolate initial velocity on initial locations
+    lats = diagn.particles.u    # reuse u,v arrays as only used for u, v
+    lons = diagn.particles.v    # after update_locator!
+    
+    for i in eachindex(particles)
+        # modulo all particles here
+        # i.e. one can start with a particle at -120˚E which moduloed to 240˚E here
+        particles[i] = mod(particles[i])
+        lats[i] = particles[i].lat
+        lons[i] = particles[i].lon
+    end
 
-    # don't do particle advection on the 2nd time step (which is half a leapfrog step that isn't counted)
-    clock.timestep_counter == 0 && lf == 2 && return nothing
+    RingGrids.update_locator!(interpolator, lats, lons)
+    u0 = diagn.particles.u      # now reused arrays are actually u, v
+    v0 = diagn.particles.v
+    interpolate!(u0, u_grid, interpolator)
+    interpolate!(v0, v_grid, interpolator)
+
+    print("PA: Initial velocities interpolated on initial locations")
+    @info (progn.clock.timestep_counter, progn.clock.time)
+end
+
+
+# function barrier
+function particle_advection!(progn, diagn, adv::ParticleAdvection2D)
+    particle_advection!(progn.particles, diagn, progn.clock, adv)
+end
+
+function particle_advection!(
+    particles::Vector{Particle{NF}},
+    diagn::AbstractVariables,
+    clock::Clock,
+    particle_advection::ParticleAdvection2D,
+) where NF
+
+    # escape immediately for no particles
+    length(particles) == 0 && return nothing
+
[…]
