### 2024-05-03 SpeedyWeather/SpeedyWeather.jl#443 @milankl
@kthyng Thanks for the reminder, it's just a lot of documentation to write... I'm still working on @natgeo-wong's points (any feedback welcome to our recent progress on this issue!) who rightly requested to have more documentation, in particular we've added now

- Vertical diffusion #532
- Surface fluxes #530
- Orography #526
- Land-sea mask and aquaplanet simulations #496
- Held-Suarez simulations #514
- Radiation #513
- Convection #512
- Large-scale condensation #510
- Using gradient operators from SpeedyWeather #523
- Online model analysis with operators in SpeedyWeather and callbacks #505, #497
- Model and Simulation structure #492

I guess what's still missing (please @natgeo-wong check whether you agree) is

- Some documentation how to use/extend the ocean model
- An overview of how to define new parameterizations
