### 2024-05-06 SpeedyWeather/SpeedyWeather.jl#443 @milankl
Responses to individual points @natgeo-wong

> 1. In light of the above, the page Model Setups in the documentation should be re-
named as Examples. The page Model Setups should more follow the style of the Oceananigans.jl documentation, in that it specifies and expands upon explaining the
different AbstractTypes that are available for the different model types.

We now have 2D and 3D examples

- https://speedyweather.github.io/SpeedyWeather.jl/dev/examples_2D/
- https://speedyweather.github.io/SpeedyWeather.jl/dev/examples_3D/

introducing some of the types/components users can change and play around with. Others are discussed in other parts of the documentation, e.g. https://speedyweather.github.io/SpeedyWeather.jl/dev/land_sea_mask/#Time-dependent-land-sea-mask documents a time-dependent land-sea mask, or https://speedyweather.github.io/SpeedyWeather.jl/dev/analysis/ explains how to check for conservation of mass, energy etc on the fly using gradient operators and the spectral transform for spatial integrations: https://speedyweather.github.io/SpeedyWeather.jl/dev/gradients/

> 2. There is a lot of explanation on the theoretical physics that underlies this model in the documentation (i.e., a page dedicated to the physics behind say, the Primitive Equation model). I recommend grouping these all under a single category in the documentation named Model Physics.

We did this. Thanks for the suggestion.

> 3. While it is always useful to keep a reference list of API of Functions and Types at the end, some of these Functions and Types would be better placed at where the documentation either explains the Model Physics or somewhere within the Model Setup.

Thanks for the suggestion but in the end we decided against this because we wanted to separate the theoretical/physical/mathematical aspects from the implementation details, API, function signatures. We now usually have some example outlining how a given function/type is supposed to be used or extended and in that sense one can always request the docstring in the REPL like `?model` or search the documentation if the usage of a function is not clear. I personally prefer to have most parts of the documentation example-driven instead of hitting a user with all optional keywords to a function which the docstring, however, definitely should provide. But please let us know if there's anything in the documentation where any of this is not clear.

> 4. There needs to be a section somewhere in the Documentation or README on how to officially contribute to SpeedyWeather.jl, even if it’s just a note to make an issue/pull-request

We have added this to the readme and the documentation. There's many ways to contribute but we encourage people being interested to open an issue or reach out to us individually.

> 5. what specific AbstractType modules are planned to be made available in Speedy-
Weather.jl for the different physical parameterizations

See https://speedyweather.github.io/SpeedyWeather.jl/dev/parameterizations/ and also more in answers below.

> 6. what are the standard fields required for each of these AbstractType modules

Only few have required fields, like orography, see https://speedyweather.github.io/SpeedyWeather.jl/dev/orography/, others like a custom coriolis struct would need to have a field `f::Vector` but we don't really encourage people to change them although it's certainly possible to do (e.g. you can mathematically think of a planet where northern and southern hemisphere rotate in opposite directions, meaning you have low pressure system rotating anti-clockwise on both hemispheres). Whenever there is a required field we specify this in the respective part of the documentation.

> 7. which of these AbstractType modules are relevant for each of the different types of
models (e.g. ShallowWaterModel, PrimitiveDryModel) available in SpeedyWeather.jl

For the parameterizations we document the abstract types now in https://speedyweather.github.io/SpeedyWeather.jl/dev/parameterizations/

Technically you can always do
```julia
julia> subtypes(SpeedyWeather.AbstractModelComponent)
16-element Vector{Any}:
 SpeedyWeather.AbstractAdiabaticConversion
 SpeedyWeather.AbstractAlbedo
 SpeedyWeather.AbstractAtmosphere
 SpeedyWeather.AbstractCoriolis
 SpeedyWeather.AbstractDrag
 SpeedyWeather.AbstractForcing
 SpeedyWeather.AbstractGeopotential
 SpeedyWeather.AbstractHorizontalDiffusion
 SpeedyWeather.AbstractImplicit
 SpeedyWeather.AbstractInitialConditions
 SpeedyWeather.AbstractOrography
 SpeedyWeather.AbstractParameterization
 SpeedyWeather.AbstractParticleAdvection
 SpeedyWeather.AbstractPlanet
 SpeedyWeather.AbstractSchedule
 SpeedyWeather.AbstractTimeStepper

julia> subtypes(SpeedyWeather.AbstractParameterization)
12-element Vector{Any}:
 SpeedyWeather.AbstractBoundaryLayer
 SpeedyWeather.AbstractClausiusClapeyron
 SpeedyWeather.AbstractCondensation
 SpeedyWeather.AbstractConvection
 SpeedyWeather.AbstractRadiation
 SpeedyWeather.AbstractSurfaceEvaporation
 SpeedyWeather.AbstractSurfaceHeatFlux
 SpeedyWeather.AbstractSurfaceThermodynamics
 SpeedyWeather.AbstractSurfaceWind
 SpeedyWeather.AbstractTemperatureRelaxation
 SpeedyWeather.AbstractVegetation
 SpeedyWeather.AbstractVerticalDiffusion
```

But we are hesitating to put this list into the documentation as it sounds like we encourage users to just go ahead and define new concrete types for any of them. While some are straightforward to subtype, others are more complicated and so we'd rather explain those that we encourage users to experiment with and document them well but slowly growing the documentation over time.

> 8. My preference would be for the creation of detailed documentation for at least two or
three different modules before this paper is accepted for publication. The documentation
for the AbstractPlanet and AbstractOrography superTypes, for example, could be pretty
helpful.

Documented now are:
- Forcing and drag, https://speedyweather.github.io/SpeedyWeather.jl/dev/forcing_drag/
- Parameterizations https://speedyweather.github.io/SpeedyWeather.jl/dev/parameterizations/
- Orography: https://speedyweather.github.io/SpeedyWeather.jl/dev/orography/
- Land-sea mask: https://speedyweather.github.io/SpeedyWeather.jl/dev/land_sea_mask/
- Ocean: https://speedyweather.github.io/SpeedyWeather.jl/dev/ocean/
- Callbacks: https://speedyweather.github.io/SpeedyWeather.jl/dev/callbacks/

On `AbstractPlanet`: We don't document or expect someone to define another planet like `Jupiter <: AbstractPlanet`, while this is mathematically straight forward, we haven't tested any of this (see #462) and so don't want to document something that we don't know works. When we can guarantee that a planet with another radius, rotation, gravity does not cause numerical stabilities (it likely will) then we'll document this. Otherwise we're afraid that too many users will run into instabilities. In other words, while easy to implement, other planets are currently not supported from our side and entirely experimental.

> 9. And also, a question I had is, what is the criteria that determines whether something
needs to become modularized? For example, I believe that SpeedyWeather does have the
AbstractOrography superType, but within this superType after some exploring I have seen
that you can point towards different files for orography. And if you change the file for orog-
raphy file, technically, you can run drastically non-Earthlike orographies in your experiment.
So why do you need an AbstractOrography superType?

From a software engineering perspective it's easier and clearer to implement a very similar structure for (almost) all components. That's why we have abstract `VerticalCoordinates`, `AbstractTimeStepper`, `AbstractImplicit` although implementing a new concrete type is anything but straight forward. Some components are easy to abstract and make modular, others like the timestepping are so deeply ingrained into the model structure that it's more work to abstract them (and hence provide an API for custom) than we currently believe is worth it. So these "criteria" you are requesting are somewhat a grey zone: Some components feel very natural to abstract and modularise others require much more effort or are components that we believe a user would not want to experiment with. But so depending on future demands this line can shift: At the moment we don't document or expect someone to define another planet like `Jupiter <: AbstractPlanet`, while this is mathematically straight forward, we haven't tested any of this (see #462) and so don't want to document something that we don't know works.

On Orography: We have `EarthOrography`, `ZonalRidge` and `NoOrography`. You are right than one can just use another file within `EarthOrography` and get, e.g. Mars' orography. However, maybe someone wants to define some orography as a function of lat, lon or maybe also orography with some randomness (rough ground etc). Orography is now documented here https://speedyweather.github.io/SpeedyWeather.jl/dev/orography/

> 10. In the Primitive Equations, there is a note that radiation is subsumed into a forcing
term Q/cp. Is there any plan to add a radiative scheme, and which scheme? And right
now, for clarification purposes, the model is entirely driven by the surface temperatures?

We have now several radiation schemes, see #502 #444 even though they are still very simple. More complicated ones on the todo list are #461, the Frierson 2006 2-band radiation (grey) and Fortran SPEEDY's 4-band radiation scheme (humidity and clouds impact radiation, +CO2 forcing). A simple `TransparentShortwave` radiation can force the model, but if disabled the forcing comes from sea surface temperatures indeed. The current ones are also documented: https://speedyweather.github.io/SpeedyWeather.jl/dev/radiation/. While there is much more todo on this, we agree, we believe it's beyond the scope of a JOSS paper and more something towards a model evaluation paper. So we're in favour of not implementing more complicated radiation schemes at this stage.
