### 2024-05-13 SpeedyWeather/SpeedyWeather.jl#443 @natgeo-wong
Very nice work and detailed documentation! I have one passing comment, is that certain parameterizations and physics may only be relevant to `PrimitiveModel` Types, and not `ShallowWater` or `Barotropic` types. Maybe some demarcation of this would help? Other than that, all things look good to me.
