### 2024-05-13 #6323 @natgeo-wong
All comments have been addressed. I have not yet closed https://github.com/SpeedyWeather/SpeedyWeather.jl/issues/443, but I believe that the editor can move forward with this paper.
