### 2024-05-13 SpeedyWeather/SpeedyWeather.jl#443 @milankl
All parameterizations act on vertical columns and are therefore *only* available for `PrimitiveEquation` as the others are 2D models where any additional terms have to be formulated as forcing/drag. Hmmm maybe that's not clear?
