### 2024-05-14 SpeedyWeather/SpeedyWeather.jl#443 @natgeo-wong
Maybe it should be mentioned somewhere?
