### 2024-05-14 SpeedyWeather/SpeedyWeather.jl#443 @milankl
@natgeo-wong just added a prominent info block in the Parameterization docs for this! ☝🏼 Hope this is what you imagined and we can close this issue?
