### 2024-05-14 SpeedyWeather/SpeedyWeather.jl#443 @milankl
Sorry was automatically closed with the pr merged. You should close it
