### 2024-05-15 SpeedyWeather/SpeedyWeather.jl#443 @milankl
Good suggestion, I'll address those and references the pull requests here!
