+++ docs/src/extensions.md
@@ -1,0 +1,167 @@
+# Extending SpeedyWeather
+
+Generally, SpeedyWeather is built in a very modular, extensible way.
+While that sounds fantastic in general, it does not save you from understanding
+its [modular logic](@ref logic) before you can extend SpeedyWeather.jl easily yourself.
+We highly recommend you to read the following sections if you would like to
+extend SpeedyWeather in some way, but it also gives you a good understanding
+of how we build SpeedyWeather in the first place. Because in the end there
+is no difference between internally or externally defined model components.
+Having said that, there is a question of the [Scope of variables](@ref)
+meaning that some functions or types require its module to be explicitly named,
+like `SpeedyWeather.some_function` instead of just `some_function`. But that's
+really it.
+
+Before and especially after reading this section you are welcome to
+[raise an issue](https://github.com/SpeedyWeather/SpeedyWeather.jl/issues)
+about whatever you would like to do with SpeedyWeather. We are happy to help.
+
+## [SpeedyWeather's modular logic](@id logic)
+
+Almost every component in SpeedyWeather is implemented in three steps:
+
+1. Define a new type,
+2. define its initialization,
+3. extend a function that defines what it does.
+
+To be a bit more explicit (Julia always encourages you to think more abstractly,
+which can be difficult to get started...) we will use the example of defining
+a new forcing for the `Barotropic` or `ShallowWater` models. But the concept
+is the same whether you want to define a forcing, a drag, or a new parameterization
+for the primitive equations, etc. For the latter, see [Parameterizations](@ref)
+In general, you can define a new component also just in a notebook or in the Julia REPL,
+you do not have to branch off from the repository and write directly into it.
+However, note the [Scope of variables](@ref) if you define a component externally.
+
+To define a new forcing type, at the most basic level you would do
+```@example extending
+using SpeedyWeather
+
+struct MyForcing{NF} <: SpeedyWeather.AbstractForcing
+    # define some parameters and work arrays here
+    a::NF
+    v::Vector{NF}
+end
+```
+In Julia this introduces a new (so-called compound) type that is a subtype of `AbstractForcing`,
+we have a bunch of these abstract super types defined (see [Abstract model components](@ref))
+and you want to piggy-back on them because of multiple-dispatch. This new type could also be 
+a `mutable struct`, could have keywords defined with `Base.@kwdef` and can
+also be parametric with respect to the number format `NF` or grid, but let's skip
+those details for now. Conceptually you include into the type any parameters
+(example the float `a` here) that you may need and especially those that you want to change
+(ideally not work arrays, see discussion in [Use `ColumnVariables` work arrays](@ref)).
+This type will get a user-facing interface so that one can quickly create a new
+forcing but with altered parameters. Generally you should also include any
+kind of precomputed arrays (here a vector `v`). For example, you want to
+apply your forcing only in certain parts of the globe? Then you probably want
+to define a mask here that somehow includes the information of your region.
+For a more concrete example see [Custom forcing and drag](@ref).
+
+To define the new type's initialization, at the most basic level you need
+to extend the `initialize!` function for this new type
+```@example extending
+function initialize!(forcing::MyForcing, model::ModelSetup)
+    # fill in/change any fields of your new forcing here
+    forcing.v[1] = 1
+    # you can use information from other model components too
+    forcing.v[2] = model.planet.gravity
+end
+```
+This function is called _once_ during model initialisation (which is in fact
+just the initialisation of all its components, like the forcing here)
+and it allows you to precompute values or arrays also based on
+parameters of other model components. Like in this example, we want to
+use the gravity that is defined in `model.planet`. If you need a value
[…]
