### 2024-05-29 #6323 @kthyng
And let me know which version you'd like associated with the JOSS publication — are you ready for v0.10?
