### 2024-05-30 #6323 @milankl
Yes, we're ready for v0.10 but we have an open pull request (https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/387) for the JOSS paper that I'd like to merge first before tagging v0.10 as the JOSS release? So do you have any comments/corrections for the manuscript? If not I'll merge that to have the manuscript proof in v0.10 and then I can release that and post the Zenodo DOI here?
