### 2024-05-31 #6323 @kthyng
Generally we don't care if the version includes all changes to the paper, but perhaps you do?
* line 85 needs a space before the reference. Have you thought about exporting as zarr? Is it too late to switch?
* Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example, "julia" is not capitalized in at least one reference.

Can you go through the rest of the bulleted list above too?
