### 2024-06-05 #6323 @navidcy
> The order in the Zenodo archive is different and there are some contributors that are not on the paper (mostly maintenance commits, but several of them are acknowledged) and some paper authors that are not contributors (as they have contributed intellectually, not through code contributions)

@milankl I believe you can edit the author list in Zenodo. If JOSS requires that (not sure) then perhaps only for this Zenodo version we can have as Zenodo authors the same authors/order as in the JOSS paper and then continue with the rest?
