### 2024-06-06 #6323 @kthyng
First off, changing the Zenodo author list is not required. But also, this change is only requested for this version that is associated with the JOSS paper so that it is a cohesive archive together between as the JOSS paper and code archive together, with Zenodo, not meant to erase history to have this consistent for this particular paper release. But still, you can do as you wish.
