### 2024-06-06 #6323 @kthyng
@milankl That invalid DOI is on the MITgcm paper. Just remove the prefix.
